@@ -21,134 +21,108 @@
         "electron": "electron .",
         "locale:extract": "lingui extract",
         "locale:compile": "lingui compile",
-        "locale": "lingui extract && lingui compile",
-        "eslint": "eslint src/**/*.[t]s{,x}",
-        "format": "prettier --write src/**/*.[jt]s{,x} && eslint --fix src/**/*.[jt]s{,x}",
-        "lint": "tsc --noEmit && npm run eslint"
+        "locale": "lingui extract && lingui compile"
     },
     "dependencies": {
-        "@chia/core": "^1.0.0",
-        "@chia/icons": "^1.0.0",
-        "@chia/api": "^1.0.0",
-        "@chia/api-react": "^1.0.0",
-        "@chia/wallets": "^1.0.0",
-        "@electron/remote": "^2.0.1",
-        "@lingui/core": "^3.13.0",
-        "@lingui/macro": "^3.13.0",
-        "@lingui/react": "^3.13.0",
-        "@material-ui/core": "^4.12.3",
-        "@material-ui/icons": "^4.11.2",
+        "@chia/core": "1.0.0",
+        "@chia/icons": "1.0.0",
+        "@chia/api": "1.0.0",
+        "@chia/api-react": "1.0.0",
+        "@chia/wallets": "1.0.0",
+        "@electron/remote": "2.0.1",
+        "@lingui/core": "3.13.0",
+        "@lingui/macro": "3.13.0",
+        "@lingui/react": "3.13.0",
+        "@material-ui/core": "4.12.3",
+        "@material-ui/icons": "4.11.2",
         "@material-ui/lab": "4.0.0-alpha.60",
-        "@material-ui/styles": "^4.11.4",
-        "@nivo/core": "^0.74.0",
-        "@nivo/line": "^0.74.0",
-        "@rehooks/local-storage": "2.4.0",
-        "bech32": "^2.0.0",
-        "big.js": "^6.1.1",
-        "byte-size": "^8.1.0",
-        "core-js": "^3.19.1",
-        "crypto-browserify": "^3.12.0",
-        "dotenv": "^10.0.0",
-        "es6-promise": "^4.2.8",
-        "history": "^5.1.0",
-        "is-electron": "^2.2.1",
-        "isomorphic-fetch": "^3.0.0",
-        "js-yaml": "^4.1.0",
-        "lodash": "^4.17.21",
-        "make-plural": "^7.0.0",
-        "match-sorter": "^6.3.1",
-        "moment": "^2.29.1",
-        "normalize-url": "^7.0.2",
-        "npm-audit-resolver": "^2.3.1",
-        "react": "^17.0.2",
-        "react-dom": "^17.0.2",
-        "react-dropzone": "^11.4.2",
-        "react-hook-form": "^7.20.4",
-        "react-redux": "^7.2.6",
-        "react-router": "^6.0.2",
-        "react-router-dom": "^6.0.2",
-        "react-scroll-to-bottom": "^4.2.0",
-        "react-teleporter": "^2.2.1",
-        "react-use": "^17.3.1",
-        "redux": "^4.1.2",
-        "regenerator-runtime": "^0.13.9",
-        "seedrandom": "^3.0.5",
-        "stream-browserify": "^3.0.0",
-        "styled-components": "^5.3.3",
-        "unique-names-generator": "^4.6.0",
-        "use-dark-mode": "^2.3.1",
-        "validator": "^13.7.0",
-        "ws": "^8.3.0"
+        "@material-ui/styles": "4.11.4",
+        "@nivo/core": "0.79.0",
+        "@nivo/line": "0.79.1",
+        "@rehooks/local-storage": "2.4.4",
+        "bech32": "2.0.0",
+        "big.js": "6.1.1",
+        "byte-size": "8.1.0",
+        "core-js": "3.20.3",
+        "crypto-browserify": "3.12.0",
+        "dotenv": "14.2.0",
+        "es6-promise": "4.2.8",
+        "history": "5.2.0",
+        "is-electron": "2.2.1",
+        "isomorphic-fetch": "3.0.0",
+        "js-yaml": "4.1.0",
+        "lodash": "4.17.21",
+        "make-plural": "7.0.0",
+        "match-sorter": "6.3.1",
+        "moment": "2.29.1",
+        "normalize-url": "7.0.3",
+        "react": "17.0.2",
+        "react-dom": "17.0.2",
+        "react-dropzone": "11.5.1",
+        "react-hook-form": "7.24.1",
+        "react-redux": "7.2.6",
+        "react-router": "6.2.1",
+        "react-router-dom": "6.2.1",
+        "react-scroll-to-bottom": "4.2.0",
+        "react-teleporter": "2.2.1",
+        "react-use": "17.3.2",
+        "redux": "4.1.2",
+        "regenerator-runtime": "0.13.9",
+        "seedrandom": "3.0.5",
+        "stream-browserify": "3.0.0",
+        "styled-components": "5.3.3",
+        "unique-names-generator": "4.6.0",
+        "use-dark-mode": "2.3.1",
+        "validator": "13.7.0",
+        "ws": "8.4.2"
     },
     "devDependencies": {
-        "@babel/core": "^7.16.0",
-        "@babel/plugin-proposal-class-properties": "^7.16.0",
-        "@babel/plugin-proposal-export-default-from": "^7.16.0",
-        "@babel/plugin-transform-flow-strip-types": "^7.16.0",
-        "@babel/preset-env": "^7.16.4",
-        "@babel/preset-react": "7.16.0",
-        "@babel/preset-typescript": "7.16.0",
-        "@babel/register": "^7.16.0",
-        "@lingui/cli": "^3.13.0",
-        "@loadable/babel-plugin": "^5.13.2",
-        "@loadable/webpack-plugin": "^5.15.1",
-        "@pmmmwh/react-refresh-webpack-plugin": "^0.5.3",
-        "@svgr/webpack": "^5.5.0",
-        "@types/big.js": "^6.1.2",
-        "@types/jest": "^27.0.3",
-        "@types/lodash": "^4.14.177",
-        "@types/node": "^16.11.10",
-        "@types/react": "^17.0.37",
-        "@types/react-dom": "^17.0.11",
-        "@types/react-redux": "^7.1.20",
-        "@types/react-router-dom": "^5.3.2",
-        "@types/seedrandom": "^3.0.1",
-        "@types/styled-components": "^5.1.15",
-        "@types/validator": "^13.7.0",
-        "@typescript-eslint/eslint-plugin": "^5.4.0",
-        "@typescript-eslint/parser": "^5.4.0",
-        "babel-core": "^7.0.0-bridge.0",
-        "babel-loader": "^8.2.3",
-        "babel-plugin-lodash": "^3.3.4",
-        "babel-plugin-macros": "^3.1.0",
-        "babel-plugin-styled-components": "^2.0.2",
-        "babel-plugin-transform-imports": "^2.0.0",
-        "concurrently": "^6.4.0",
-        "copy-webpack-plugin": "^10.0.0",
-        "cross-env": "^7.0.3",
-        "css-loader": "^6.5.1",
-        "electron": "^14.2.1",
-        "electron-winstaller": "^5.0.0",
-        "eslint": "^7.32.0",
-        "eslint-config-airbnb": "^18.2.1",
-        "eslint-config-airbnb-typescript": "^14.0.0",
-        "eslint-config-prettier": "^8.3.0",
-        "eslint-plugin-eslint-comments": "^3.2.0",
-        "eslint-plugin-import": "^2.25.3",
-        "eslint-plugin-jest": "^25.3.0",
-        "eslint-plugin-jsx-a11y": "^6.5.1",
-        "eslint-plugin-prettier": "^4.0.0",
-        "eslint-plugin-promise": "^5.1.1",
-        "eslint-plugin-react": "^7.27.1",
-        "eslint-plugin-react-hooks": "^4.3.0",
-        "eslint-plugin-unicorn": "^36.0.0",
-        "eslint-webpack-plugin": "^3.1.1",
-        "file-loader": "^6.2.0",
-        "foreman": "^3.0.1",
-        "html-loader": "^3.0.1",
-        "html-webpack-plugin": "^5.5.0",
-        "lodash-webpack-plugin": "^0.11.6",
-        "prettier-eslint": "^13.0.0",
-        "prettier-eslint-cli": "^5.0.1",
-        "react-refresh": "^0.11.0",
-        "style-loader": "^3.3.1",
-        "terser-webpack-plugin": "^5.2.5",
-        "typescript": "^4.5.2",
-        "url-loader": "^4.1.1",
-        "webpack": "^5.64.4",
-        "webpack-cli": "^4.9.1",
-        "webpack-dev-server": "^4.6.0",
-        "xmldom": "github:xmldom/xmldom#0.7.5"
+        "@babel/core": "7.16.7",
+        "@babel/preset-env": "7.16.8",
+        "@babel/preset-react": "7.16.7",
+        "@babel/preset-typescript": "7.16.7",
+        "@babel/register": "7.16.9",
+        "@lingui/cli": "3.13.0",
+        "@loadable/babel-plugin": "5.13.2",
+        "@loadable/webpack-plugin": "5.15.2",
+        "@pmmmwh/react-refresh-webpack-plugin": "0.5.4",
+        "@svgr/webpack": "6.2.0",
+        "@types/big.js": "6.1.2",
+        "@types/jest": "27.4.0",
+        "@types/lodash": "4.14.178",
+        "@types/node": "17.0.10",
+        "@types/react": "17.0.38",
+        "@types/react-dom": "17.0.11",
+        "@types/react-redux": "7.1.22",
+        "@types/react-router-dom": "5.3.2",
+        "@types/seedrandom": "3.0.1",
+        "@types/styled-components": "5.1.20",
+        "@types/validator": "13.7.1",
+        "babel-core": "7.0.0-bridge.0",
+        "babel-loader": "8.2.3",
+        "babel-plugin-lodash": "3.3.4",
+        "babel-plugin-macros": "3.1.0",
+        "babel-plugin-styled-components": "2.0.2",
+        "babel-plugin-transform-imports": "2.0.0",
+        "concurrently": "7.0.0",
+        "copy-webpack-plugin": "10.2.0",
+        "cross-env": "7.0.3",
+        "css-loader": "6.5.1",
+        "electron": "14.2.3",
+        "electron-winstaller": "5.0.0",
+        "file-loader": "6.2.0",
+        "html-loader": "3.1.0",
+        "html-webpack-plugin": "5.5.0",
+        "lodash-webpack-plugin": "0.11.6",
+        "react-refresh": "0.11.0",
+        "style-loader": "3.3.1",
+        "terser-webpack-plugin": "5.3.0",
+        "typescript": "4.5.4",
+        "url-loader": "4.1.1",
+        "webpack": "5.66.0",
+        "webpack-cli": "4.9.1",
+        "webpack-dev-server": "4.7.3",
+        "xmldom": "github:xmldom/xmldom#0.8.0"
     },
     "browserslist": {
         "production": [
