@@ -3,7 +3,7 @@ import optionsForPlotter from '../utils/optionsForPlotter';
 import defaultsForPlotter from '../utils/defaultsForPlotter';
 
 export default {
-  displayName: "Chia Proof of Space",
+  displayName: 'Chia Proof of Space',
   options: optionsForPlotter(PlotterName.CHIAPOS),
   defaults: defaultsForPlotter(PlotterName.CHIAPOS),
   installInfo: { installed: true },
