@@ -14,12 +14,20 @@ const StyledCardTitle = styled(Box)`
   padding: ${({ theme }) => `${theme.spacing(2)}px ${theme.spacing(2)}px`};
 `;
 
-const StyledCardMaterial = styled(({ cursor, opacity, clickable, ...rest }) => (
+const StyledCardMaterial = styled(({ cursor, opacity, clickable, fullHeight, ...rest }) => (
   <CardMaterial {...rest}/>
 ))`
   cursor: ${({ clickable }) => clickable ? 'pointer' : 'default'};
   opacity: ${({ disabled }) => disabled ? '0.5': '1'};
+  height: ${({ fullHeight }) => fullHeight ? '100%': 'auto'};
+`;
 
+const StyledCardContent = styled(({ fullHeight, ...rest }) => (
+  <CardContent {...rest}/>
+))`
+  display: flex;
+  flex-direction: column;
+  height: ${({ fullHeight }) => fullHeight ? '100%': 'auto'};
 `;
 
 type Props = {
@@ -32,10 +40,11 @@ type Props = {
   action?: ReactNode;
   onSelect?: () => void;
   disabled?: boolean;
+  fullHeight?: boolean;
 };
 
 export default function Card(props: Props) {
-  const { children, title, tooltip, actions, gap, interactive, action, onSelect, disabled } = props;
+  const { children, title, tooltip, actions, gap, interactive, action, onSelect, disabled, fullHeight } = props;
 
   const headerTitle = tooltip ? (
     <Flex alignItems="center" gap={1}>
@@ -53,7 +62,7 @@ export default function Card(props: Props) {
   }
 
   return (
-    <StyledCardMaterial onClick={handleClick} clickable={!!onSelect} disabled={disabled}>
+    <StyledCardMaterial onClick={handleClick} clickable={!!onSelect} disabled={disabled} fullHeight={fullHeight}>
       {title && (
         <StyledCardTitle>
           <Flex gap={2} alignItems="center" flexWrap="wrap">
@@ -64,9 +73,9 @@ export default function Card(props: Props) {
           </Flex>
         </StyledCardTitle>
       )}
-      <CardContent>
-        <Flex flexDirection="column" gap={3}>
-          <Flex flexDirection="column" gap={gap}>
+      <StyledCardContent fullHeight={fullHeight}>
+        <Flex flexDirection="column" gap={3} flexGrow={1}>
+          <Flex flexDirection="column" gap={gap} flexGrow={1}>
             {children}
           </Flex>
           {actions && (
@@ -75,7 +84,7 @@ export default function Card(props: Props) {
             </Grid>
           )}
         </Flex>
-      </CardContent>
+      </StyledCardContent>
     </StyledCardMaterial>
   );
 }
