@@ -1,11 +1,9 @@
 import React, { ReactNode, Suspense } from 'react';
 import styled from 'styled-components';
 import { useNavigate, Outlet } from 'react-router-dom';
-import { t } from '@lingui/macro';
+import { t, Trans } from '@lingui/macro';
 import { AppBar, Toolbar, Drawer, Divider, Container, IconButton } from '@material-ui/core';
 import {
-  DarkModeToggle,
-  LocaleToggle,
   Flex,
   Logo,
   ToolbarSpacing,
@@ -14,6 +12,9 @@ import {
 import { DashboardTitleTarget } from '../DashboardTitle';
 import { useLogout } from '@chia/api-react';
 import { ExitToApp as ExitToAppIcon } from '@material-ui/icons';
+import Settings from '../Settings';
+import Tooltip from '../Tooltip';
+// import LayoutFooter from '../LayoutMain/LayoutFooter';
 
 const StyledRoot = styled(Flex)`
   height: 100%;
@@ -29,9 +30,9 @@ const StyledAppBar = styled(AppBar)`
   background-color: ${({ theme }) =>
     theme.palette.type === 'dark' ? '#424242' : 'white'};
   box-shadow: 0px 0px 8px rgba(0, 0, 0, 0.2);
-  width: ${({ theme }) => `calc(100% - ${theme.drawer.width})`};
-  margin-left: ${({ theme }) => theme.drawer.width};
-  z-index: ${({ theme }) => theme.zIndex.drawer + 1};
+  width: ${({ theme, drawer }) => drawer ? `calc(100% - ${theme.drawer.width})` : '100%'};
+  margin-left: ${({ theme, drawer }) => drawer ? theme.drawer.width : 0};
+  z-index: ${({ theme }) => theme.zIndex.drawer + 1};};
 `;
 
 const StyledDrawer = styled(Drawer)`
@@ -65,10 +66,11 @@ export type LayoutDashboardProps = {
   children?: ReactNode;
   sidebar?: ReactNode;
   outlet?: boolean;
+  settings?: ReactNode;
 };
 
 export default function LayoutDashboard(props: LayoutDashboardProps) {
-  const { children, sidebar, outlet } = props;
+  const { children, sidebar, settings, outlet } = props;
 
   const navigate = useNavigate();
   const logout = useLogout();
@@ -84,14 +86,20 @@ export default function LayoutDashboard(props: LayoutDashboardProps) {
       <Suspense fallback={<Loading center />}>
         {sidebar ? (
           <>
-            <StyledAppBar position="fixed" color="transparent" elevation={0}>
+            <StyledAppBar position="fixed" color="transparent" elevation={0} drawer>
               <StyledToolbar>
                 <Container maxWidth="lg">
                   <Flex alignItems="center">
                     <DashboardTitleTarget />
                     <Flex flexGrow={1} />
-                    <LocaleToggle />
-                    <DarkModeToggle />
+                    <Tooltip title={<Trans>Logout</Trans>}>
+                      <IconButton color="inherit" onClick={handleLogout} title={t`Log Out`}>
+                        <ExitToAppIcon />
+                      </IconButton>
+                    </Tooltip>
+                    <Settings>
+                      {settings}
+                    </Settings>
                   </Flex>
                 </Container>
               </StyledToolbar>
@@ -105,21 +113,24 @@ export default function LayoutDashboard(props: LayoutDashboardProps) {
             </StyledDrawer>
           </>
         ): (
-          <Toolbar>
-            <Container maxWidth="lg">
-              <Flex>
-                <StyledBrandWrapper>
-                  <Logo height={1} />
-                </StyledBrandWrapper>
-                <Flex flexGrow={1} alignItems="flex-end" gap={1} />
-                <LocaleToggle />
-                <DarkModeToggle />
-                <IconButton color="inherit" onClick={handleLogout} title={t`Log Out`}>
-                  <ExitToAppIcon />
-                </IconButton>
-              </Flex>
-            </Container>
-          </Toolbar>
+          <StyledAppBar position="fixed" color="transparent" elevation={0}>
+            <StyledToolbar>
+              <Container maxWidth="lg">
+                <Flex alignItems="center">
+                  <Logo width="100px" />
+                  <Flex flexGrow={1} />
+                  <Tooltip title={<Trans>Logout</Trans>}>
+                    <IconButton color="inherit" onClick={handleLogout} title={t`Log Out`}>
+                      <ExitToAppIcon />
+                    </IconButton>
+                  </Tooltip>
+                  <Settings>
+                    {settings}
+                  </Settings>
+                </Flex>
+              </Container>
+            </StyledToolbar>
+          </StyledAppBar>
         )}
 
         <StyledBody flexDirection="column" flexGrow={1}>
@@ -129,6 +140,7 @@ export default function LayoutDashboard(props: LayoutDashboardProps) {
               <Suspense fallback={<Loading center />}>
                 {outlet ? <Outlet /> : children}
               </Suspense>
+              {/* <LayoutFooter /> */}
             </Flex>
           </StyledContainer>
         </StyledBody>
