@@ -4,10 +4,10 @@ import { useLocalStorage, writeStorage } from '@rehooks/local-storage';
 const LOCAL_STORAGE_KEY = 'skipMigration';
 
 export default function useSkipMigration(): [boolean, (skip: boolean) => void] {
-  let [skip] = useLocalStorage<boolean>(LOCAL_STORAGE_KEY, false);
+  const [skip] = useLocalStorage<boolean>(LOCAL_STORAGE_KEY, false);
 
   const handleSetSkipMigration = useCallback((newSkip: boolean) => {
-    writeStorage('locale', newSkip);
+    writeStorage(LOCAL_STORAGE_KEY, newSkip);
   }, []);
 
   return [skip, handleSetSkipMigration];
