@@ -14,14 +14,14 @@ module.exports = function babel(api) {
         loose: LOOSE,
       }],
       '@babel/preset-typescript',
-      '@babel/preset-react',
+      ['@babel/preset-react', {
+        runtime: 'automatic',
+      }],
     ],
     plugins: [
       'macros',
       '@loadable/babel-plugin',
       ['babel-plugin-styled-components'],
-      ['@babel/plugin-proposal-class-properties', { loose: LOOSE }],
-      '@babel/plugin-proposal-export-default-from',
     ],
   };
 };
