@@ -12,7 +12,7 @@ import {
   Button,
   InputLabel,
 } from '@material-ui/core';
-import { AlertDialog, Card, Flex } from '@chia/core';
+import { AlertDialog, Card, Flex, useShowSaveDialog } from '@chia/core';
 import isElectron from 'is-electron';
 import { newBuy, newSell, addTrade, resetTrades } from '../../modules/trade';
 import { colouredcoin_to_mojo } from '../../util/chia';
@@ -43,6 +43,10 @@ export default function CreateOffer() {
   const wallets = useSelector((state) => state.wallet_state.wallets);
   const dispatch = useDispatch();
   const navigate = useNavigate();
+<<<<<<< HEAD
+=======
+  const showSaveDialog = useShowSaveDialog();
+>>>>>>> 207cb1a67d4bce2ecd46a9125678de02d66d71b1
   let amount_input = null;
   let buy_or_sell = null;
   let wallet_id = null;
@@ -105,7 +109,11 @@ export default function CreateOffer() {
     }
     if (isElectron()) {
       const dialogOptions = {};
+<<<<<<< HEAD
       const result = await window.ipcRenderer?.send('showSaveDialog', dialogOptions);
+=======
+      const result = await showSaveDialog(dialogOptions);
+>>>>>>> 207cb1a67d4bce2ecd46a9125678de02d66d71b1
       const { filePath } = result;
       const offer = {};
       for (const trade of trades) {
@@ -115,7 +123,7 @@ export default function CreateOffer() {
           offer[trade.wallet_id] = -trade.amount;
         }
       }
-      dispatch(create_trade_action(offer, filePath, history));
+      dispatch(create_trade_action(offer, filePath, navigate));
     } else {
       dispatch(
         openDialog(
