@@ -5,4 +5,11 @@ import ReactDOM from 'react-dom';
 import './config/env';
 import AppRouter from './components/app/AppRouter';
 
-ReactDOM.render(<AppRouter />, document.querySelector('#root'));
+// we need to use additional root for hot reloading
+function Root() {
+  return (
+    <App />
+  );
+}
+
+ReactDOM.render(<Root />, document.querySelector('#root'));
