@@ -1,5 +1,3 @@
-const LOOSE = false;
-
 module.exports = function babel(api) {
   api.cache(true);
 
@@ -11,17 +9,16 @@ module.exports = function babel(api) {
         },
         useBuiltIns: 'entry',
         corejs: 3,
-        loose: LOOSE,
       }],
       '@babel/preset-typescript',
-      '@babel/preset-react',
+      ["@babel/preset-react", {
+        "runtime": "automatic"
+      }]
     ],
     plugins: [
       'macros',
       '@loadable/babel-plugin',
       ['babel-plugin-styled-components'],
-      ['@babel/plugin-proposal-class-properties', { loose: LOOSE }],
-      '@babel/plugin-proposal-export-default-from',
     ],
   };
 };
