@@ -1,12 +1,11 @@
-import { useMemo, useState } from 'react';
-import { orderBy } from 'lodash';
+import { useState } from 'react';
 import { useGetTransactionsQuery, useGetTransactionsCountQuery } from '@chia/api-react';
 import Transaction from '../types/Transaction';
 
 export default function useWalletTransactions(
   walletId: number, 
-  defaultRowsPerPage: number = 10, 
-  defaultPage: number = 0, 
+  defaultRowsPerPage = 10, 
+  defaultPage = 0, 
   sortKey?: 'CONFIRMED_AT_HEIGHT' | 'RELEVANCE',
   reverse?: boolean,
 ): {
