@@ -1,1 +1,1 @@
-/*eslint-disable*/module.exports={messages:{"(Optional)":"（可选）","*Want to explore Chia’s blocks further? Check out <0>Chia Explorer</0> built by an open source developer.":"* 想要进一步了解奇亚？来看看一位开源开发者创建的 <0>奇亚区块浏览器</0>。","+":"+","+ Add New Plot NFT":"+ 添加新的农田NFT","+ Add Token":"+ Add Token","+ Add Wallet":"+ 添加钱包","-":"-","0 automatically chooses bucket count":"0 automatically chooses bucket count","128 buckets is recommended":"排序桶数量推荐设置为128","A harvester is a service running on a machine where plot(s) are actually stored. A farmer and harvester talk to a full node to see the state of the chain. View your network of connected harvesters below Learn more":"收割机是一个在存有农田文件的机器上运行的服务。农民和收割机通过和全节点交互来得到链的状态。你可以在下面看到你网络所连接的收割机。了解更多","About Chia Blockchain":"关于奇亚Chia区块链项目","About Chia Wallet":"About Chia Wallet","Accept":"接受","Accepted at time:":"接受时间：","Action":"操作","Actions":"操作","Add":"添加","Add Backup ID":"添加备份 ID","Add Custom Token":"Add Custom Token","Add Plot Directory":"添加农田文件夹","Add Plot to Queue":"将农田开垦任务加入队列","Add Token":"Add Token","Add Wallet":"Add Wallet","Add a Plot":"开垦农田","Add a Plot NFT":"开垦农田NTF","Add a plot":"开垦农田","Add plot directory":"添加农田文件夹","Add {currencyCode} from the Faucet":["从水龙头获取 ",["currencyCode"]],"Adding {0} token":["Adding ",["0"]," token"],"Address":"地址","Address / Puzzle hash":"地址 / 谜语哈希","After your keys are successfully migrated to the new keyring, you may choose to have your keys removed from the old keyring.":"After your keys are successfully migrated to the new keyring, you may choose to have your keys removed from the old keyring.","All":"All","Amount":"数量","Amount ({cc_unit})":["数量 (",["cc_unit"],")"],"Amount For Initial Coin":"初始币数量","Amount must be an even amount.":"Amount must be an even amount.","Are you sure you want to delete the plot? The plot cannot be recovered.":"你确定要删除这个农田吗？删除后将不可恢复。","Are you sure you want to delete unconfirmed transactions?":"确定要删除未确认的交易吗？","Are you sure you want to disconnect?":"是否要断开连接？","Are you sure you want to quit?":"Are you sure you want to quit?","Are you sure you want to quit? GUI Plotting and farming will stop.":"确定要退出吗? 农田开垦和耕种将会停止。","Are you sure you want to use k={plotSize}":["Are you sure you want to use k=",["plotSize"]],"Are you sure you want to use k={plotSize}?":["你确定要使用 k=",["plotSize"],"？"],"Asset Id":"Asset Id","Atomic Swap Wallet":"Atomic Swap Wallet","Attempt {attempt}":["Attempt ",["attempt"]],"Authorized Payee Wallet":"Authorized Payee Wallet","Autogenerated name from pool contract address":"从农业合作社合约地址自动生成的名称","Back":"返回","Backup file is used to restore smart wallets.":"备份文件将用于恢复智能钱包。","Balance":"余额","Base Farmer Reward Amount":"基础耕种奖励数量","Below are the current block challenges. You may or may not have a proof of space for these challenges. These blocks do not currently contain a proof of time.":"下面是当前区块的挑战。 你的农田或许包含了对这些挑战的空间证明。这些区块目前还未包含时间证明。","Best estimate over last 24 hours":"最近24小时的最佳估值","Best estimate over last 5 hours":"Best estimate over last 5 hours","Block":"区块","Block Test":"区块测试","Block VDF Iterations":"区块可验证延迟函数迭代","Block at height {0} in the Chia blockchain":["奇亚区块链上高度为 ",["0"]," 的区块"],"Block hash":"Block hash","Block with hash {headerHash}":["区块哈希",["headerHash"]],"Block with hash {headerHash} does not exist.":["哈希为 ",["headerHash"]," 的区块不存在。"],"Blocks":"区块","Both values are empty. This is ok if you are, for example, sending your rewards to an offline wallet.":"Both values are empty. This is ok if you are, for example, sending your rewards to an offline wallet.","Browse":"浏览","But you are currently farming <0/>":"但你目前正在耕种<0/>","Buy":"购买","CAT Wallet":"CAT Wallet","CC Wallet":"CC Wallet","CCListItems.title":"CCListItems.title","Can be backed up to mnemonic seed":"可备份为助记词","Cancel":"取消","Cancel and Spend":"取消并支付","Cannot send chia to coloured address. Please enter a chia address.":"Cannot send chia to coloured address. Please enter a chia address.","Caution, deleting these plots will delete them forever. Check that the storage devices are properly connected.":"注意，删除这些农田的操作是永久的。请检查存储设备连接正常。","Challenge":"挑战","Challenge Hash":"挑战哈希","Change":"变更","Change Pool":"更换农业合作社","Chat on KeyBase":"在Keybase上参与讨论","Chia":"奇亚","Chia Asset Token":"Chia Asset Token","Chia Blockchain Wiki":"奇亚区块链项目维基","Chia Wallet":"奇亚钱包","Choose Number of Plots":"选择要开垦的数量","Choose Plot Size":"选择农田文件大小","Claim Rewards":"申领奖励","Close":"关闭","Closing down node and server":"正在关闭节点与服务","Coin Name":"代币名称","Coins:":"币：","Colour":"颜色","Colour Info":"色彩信息","Colour String":"色彩字符串","Colour:":"色彩：","Coloured Coin":"染色币","Coloured Coin Options":"染色币选项","Confirm":"确认","Confirm Disconnect":"确认断开连接","Confirm Passphrase":"Confirm Passphrase","Confirmation":"确认","Confirmed":"Confirmed","Confirmed at Height":"Confirmed at Height","Confirmed at block:":"区块上已确认：","Confirmed at height {0}":["确认于区块高度 ",["0"]],"Connect":"连接","Connect to other peers":"连接到其它节点","Connect to pool":"连接到农业合作社","Connected":"已连接","Connecting do daemon":"Connecting do daemon","Connecting to daemon":"Connecting to daemon","Connecting to wallet":"正在连接到钱包","Connection Status":"连接状态","Connection Status:":"连接状态：","Connection type":"连接类型","Connections":"连接数","Contribute on GitHub":"在 GitHub 上贡献","Copied":"已复制","Copy":"复制","Copy to Clipboard":"复制到剪贴板","Create":"创建","Create An Attestation Packet":"Create an Attestation Packet","Create Chia Asset Token Wallet":"Create Chia Asset Token Wallet","Create Chia Asset Token Wallet from Existing TAIL":"Create Chia Asset Token Wallet from Existing TAIL","Create DID wallet":"Create DID wallet","Create Distributed Identity Wallet":"创建分布式身份钱包","Create New Chia Asset Token Wallet":"Create New Chia Asset Token Wallet","Create New Wallet":"新建钱包","Create Offer":"创建报价","Create Plot":"Create Plot","Create Rate Limited Admin Wallet":"创建费用限定型管理钱包","Create Rate Limited User Wallet":"创建费用限定型用户钱包","Create Token":"Create Token","Create Trade":"Create Trade","Create Trade Offer":"发起交易出价","Create Transaction":"发起交易","Create a Backup":"创建备份","Create a Plot NFT":"创建农田NFT","Create a new private key":"创建新私钥","Create admin wallet":"创建管理者钱包","Create custom CAT Wallet":"Create custom CAT Wallet","Create from Existing":"Create from Existing","Create new coloured coin":"创建新的染色币","Create user wallet":"创建用户钱包","Create wallet for colour":"创建染色钱包","Create wallet for existing colour":"基于已有的染色币来创建一个钱包","CreateWallet.addWallet":"CreateWallet.addWallet","Created At:":"创建于：","Created by us:":"创建人：","Creating Plot NFT and Joining the Pool":"正在创建农田NFT并加入农业合作社","Creating Plot NFT for Self Pooling":"正在创建自耕种的农田NFT","Currency code is not defined":"货币代码未定义","Current Difficulty":"当前难度","Current Points Balance":"当前积点余额","Current trade status":"当前交易状态","Custody Wallet":"Custody Wallet","Custom":"Custom","DANGER: permanently delete private key":"注意：永久性删除私钥","DID Wallet":"DID Wallet","DIDListItems.title":"DIDListItems.title","Date":"日期","Delay":"延迟","Delete":"删除","Delete Key":"Delete Key","Delete Plot":"删除农田","Delete Unconfirmed Transactions":"删除未确认的交易","Delete all keys":"删除所有密钥","Delete key":"Delete key","Delete key {fingerprint}":["删除密钥 ",["fingerprint"]],"Deleting all keys will permanently remove the keys from your computer, make sure you have backups. Are you sure you want to continue?":"删除所有密钥将会把你的密钥从你的计算机中永久移除，请确保你已经将其备份。确定继续？","Deleting key":"Deleting key","Deleting the key will permanently remove the key from your computer, make sure you have backups. Are you sure you want to continue?":"删除此密钥将会把密钥从你的计算机中永久移除，请确保你已经将其备份。确定继续？","Developer":"开发者","Developer Tools":"开发者工具","Difficulty":"难度","Disable bifield plotting when your temporary directory is on fast storage like SSD or you need lower per process RAM requirements. Plotting with bitfield enabled has about 20% less overall writes.":"Disable bifield plotting when your temporary directory is on fast storage like SSD or you need lower per process RAM requirements. Plotting with bitfield enabled has about 20% less overall writes.","Disable bitfield plotting":"开垦时禁用位域","Discard":"放弃","Disconnect":"断开连接","Distributed Identity":"分布式身份认证","Distributed Identity Options":"Distributed Identity Options","Do you have existing plots on this machine? <0>Add Plot Directory</0>":"你的机器上有已经完成的农田文件吗？ <0>添加农田文件路径</0>","Does your machine support parallel plotting?":"您的机器是否支持并行开垦？","Drag and drop attestation packet(s)":"Drag and drop attestation packet(s)","Drag and drop offer file":"拖放报价文件","Drag and drop your backup file":"拖放你的备份文件","Drag and drop your recovery backup file":"Drag and drop your recovery backup file","Edit":"编辑","Edit Payout Instructions":"编辑支付指令","Enter a strong passphrase and click Migrate Keys to secure your keys":"Enter a strong passphrase and click Migrate Keys to secure your keys","Enter the 24 word mnemonic that you have saved in order to restore your Chia wallet.":"请输入你保存的24个助记词来恢复奇亚钱包","Error":"错误","Error the entered address appears to be for a different colour.":"Error the entered address appears to be for a different colour.","Error: Cannot send chia to coloured address. Please enter a chia address.":"错误：无法发送奇亚到染色地址。请输入奇亚钱包地址。","Estimated Network Space":"全网已占用空间（估算）","Estimated Time to Win":"预计区块发现时间","Estimated network space":"Estimated network space","Estimated sum of all the plotted disk space of all farmers in the network":"估计网络中所有农民已开垦农田占用磁盘空间量","Exclude final directory":"不耕种最终输出文件夹","Expected Time to Win":"Expected Time to Win","Failed to open (invalid plots)":"打开失败（农田无效）","Farm":"农场","Farmer Public Key":"农民公钥","Farmer Puzzle Hash":"农民解谜哈希","Farmer Reward Address":"农民奖励地址","Farmer Reward Address is not properly formatted.":"农民奖励地址格式不正确。","Farmer Reward Address is not properly formatted. Using 'paste' is recommended.":"Farmer Reward Address is not properly formatted. Using 'paste' is recommended.","Farmer Reward Address must not be empty.":"农民奖励地址不能为空。","Farmer is not connected":"农民节点无法连接","Farmer is not running":"农民节点没有运行","Farmer public key:":"农民公钥：","Farmers earn block rewards and transaction fees by committing spare space to the network to help secure transactions. This is where your farm will be once you add a plot. <0>Learn more</0>":"农民们通过贡献自己的空余存储空间，可以帮助提高整个网络的安全性，进而获得区块奖励和交易费用的奖励。当你的第一块农田开垦完成后，这里就是你的农场。<0>了解更多</0>","Farming":"正在耕种","Farming Status":"耕种状态","Fee":"费用","Fee (TXCH)":"Fee (TXCH)","Fee ({0})":["Fee (",["0"],")"],"Fee ({currencyCode})":["交易费 (",["currencyCode"],")"],"Fees Amount":"费用","Fees Puzzle Hash":"Fees Puzzle Hash","File":"文件","Filename":"文件名","Final folder location":"最终输出文件夹位置","Finished":"已完成","Follow on Twitter":"关注我们的Twitter","Frequently Asked Questions":"常见问题解答","Full Node":"全节点","Full Node Overview":"全节点概览","Full Node Status":"Full Node Status","Full Screen":"全屏","Generate New Colour":"生成新的色彩","Get Pool Login Link":"Get Pool Login Link","Grid view":"网格视图","HD or Hierarchical Deterministic keys are a type of public key/private key scheme where one private key can have a nearly infinite number of different public keys (and therefor wallet receive addresses) that will all ultimately come back to and be spendable by a single private key.":"分层确定性钱包中的私钥可以对应有无限多个公钥（也就是你的钱包接受地址），这所有的公钥都受这个单一的私钥控制。","Harvester ID":"Harvester ID","Harvester Plots":"收割机农田","Header Hash":"头部哈希","Header hash":"头部哈希","Height":"高度","Help":"帮助","Help translate":"改进翻译","Hide Advanced Options":"隐藏高级选项","Hint":"Hint","History":"历史","Host Name":"主机名","IP address":"IP地址","IP address / host":"IP地址 / 主机","If none selected, then it will default to the temporary directory.":"如果未指定，将会使用上述临时目录。","Import Wallet from Mnemonics":"使用助记词导入钱包","Import from Mnemonics (24 words)":"自助记词导入（24个词）","In Progress":"处理中","Incoming":"收入","Incorrect value":"数值错误","Index":"索引","Indicated if this offer was created by us":"表明此出价是否由我们创建","Indicated what time this offer was accepted":"表明次出价被接受的时间","Info Packet":"信息包","Initial Amount":"初始金额","Initialize a Rate Limited User Wallet:":"初始化费用限定型用户钱包：","Interval":"间隔","Invalid state":"无效状态","It is a one-time login link that can be used to log in to a pool's website. It contains a signature using the farmer's key from the plot NFT. Not all pools support this feature.":"这是一个用于登录农业合作社网站的一次性链接。此链接包含农田NFT对应的农民密钥的签名。并非所有耕种池都支持此功能。","Join Pool":"加入农业合作社","Join a Pool":"加入农业合作社","Join a pool and get consistent XCH farming rewards. The average returns are the same, but it is much less volatile.":"加入农业合作社来获得持续的XCH耕种收益。平均回报是相同的，但是收益会更加稳定。","Join a pool and get consistent XCH farming rewards. The average returns are the same, but it is much less volatile. Assign plots to a plot NFT. You can easily switch pools without having to re-plot.":"加入农业合作社来获得持续的XCH耕种收益。平均回报是相同的，但是收益会更加稳定。为已开垦的农田分配农田NFT。你可以轻松切换耕种池，而不比重新开垦现有农田。","Join a pool and get more consistent XCH farming rewards. Create a plot NFT and assign your new plots to a group.":"加入农业合作社来获得持续的XCH耕种收益。创建一个农田NFT，然后将新农田分配到组。","K-Size":"K值大小","Keyring migration failed: {0}":["Keyring migration failed: ",["0"]],"Keys":"密钥","KiB Up/Down":"KiB 上传/下载","Last Attempted Proof":"最近尝试过的证明","Last Height Farmed":"最新收割的区块高度","Latest Block Challenges":"最新的区块挑战","Launcher Id":"启动器ID","Learn More":"了解更多","Leaving Pool":"退出农业合作社","List of connections is empty":"List of connections is empty","List view":"列表视图","Loading Plot NFTs":"正在加载农田NFT","Loading configuration":"Loading configuration","Loading keyring status":"Loading keyring status","Loading list of the keys":"Loading list of the keys","Loading list of wallets":"正在加载钱包列表","Loading...":"正在载入……","Local Harvester Plots":"Local Harvester Plots","Log Out":"Log Out","Logging in":"正在登录","MainWalletList.colouredCoin":"MainWalletList.colouredCoin","MainWalletList.createAdminWallet":"MainWalletList.createAdminWallet","MainWalletList.createDIDWallet":"MainWalletList.createDIDWallet","MainWalletList.createNewColouredCoin":"MainWalletList.createNewColouredCoin","MainWalletList.createUserWallet":"MainWalletList.createUserWallet","MainWalletList.createWalletForExistingColour":"MainWalletList.createWalletForExistingColour","MainWalletList.distributedIdentity":"MainWalletList.distributedIdentity","MainWalletList.rateLimited":"MainWalletList.rateLimited","MainWalletList.recoverDIDWallet":"MainWalletList.recoverDIDWallet","MainWalletList.title":"MainWalletList.title","Manage Farming Rewards":"管理耕种奖励","Manage Recovery DIDs":"管理恢复去中心化标识","Manage Your Farming Rewards Target Addresses":"管理耕种奖励发放地址","Memo":"Memo","Memos":"Memos","MiB Up/Down":"MiB 上传/下载","Migrate Keys":"Migrate Keys","Migration required":"Migration required","Minimum Difficulty":"最小难度","Minutes":"分钟","More memory slightly increases speed":"增加内存可以略微提高速度","Multi Sig Wallet":"Multi Sig Wallet","My DID Wallet":"我的去中心化标识钱包","My Pubkey":"我的公钥","Name":"Name","Network Name":"网络名称","Network has not yet been launched, once it launches full node will start automatically.":"Network has not yet been launched, once it launches full node will start automatically.","Network not launched yet, waiting for genesis challenge!":"Network not launched yet, waiting for genesis challenge!","New Address":"新地址","New Wallet":"新钱包","Next":"下一步","Nickname":"昵称","No":"否","No 24 word seed, since this key is imported.":"没有24个助记词种子，因为此密钥已导入。","No blocks farmed yet":"尚未有区块有收获","No previous transactions":"无过往交易","No private keys for one or both addresses. Safe only if you are sending rewards to another wallet.":"本机没有所输入地址对应的私钥。请确认是否要将奖励发送至非本机钱包。","Node ID":"节点ID","Node Id":"节点ID","None":"无","None of your plots have passed the plot filter yet.":"你的农田还没有通过过滤器检查.","Not Available":"不可用","Not Synced":"未同步","Not accepted yet":"尚未被接受","Not confirmed yet":"尚未被确认","Not connected":"未连接","Not found Plots":"未找到农田","Note that this does not change your pooling payout addresses. This only affects old format plots, and the 0.25XCH reward for pooling plots.":"请注意，这将不会改变你的合作社耕作支付地址。这只会影响旧农田文件，以及0.25XCH的合作社耕作奖励。","Number of Plots":"农田数","Number of buckets":"桶数量","Number of threads":"线程数量","OK":"确定","Offer":"出价","Offers Created":"已创建的出价","On average there is one minute between each transaction block. Unless there is congestion you can expect your transaction to be included in less than a minute.":"每个交易区块平均打包时间为一分钟。通常您的交易将在一分钟内被确认，除非出现拥堵。","Only one backup file is allowed.":"只允许一个备份文件。","Only one backup file is allowed. To add a new backup file, please delete any previously uploaded files and try again.":"Only one backup file is allowed. To add a new backup file, please delete any previously uploaded files and try again.","Outgoing":"支出","Passphrase":"Passphrase","Passphrase Hint":"Passphrase Hint","Passphrase Hint (Optional)":"Passphrase Hint (Optional)","Payout Address":"支付地址","Peak Height":"最高高度","Peak Sub-block Height":"Peak Sub-block Height","Peak Time":"最高时间","Pending":"待处理","Pending Balance":"待处理的余额","Pending Change":"待处理的找回","Pending Total Balance":"待处理总余额","Please Confirm":"请确认","Please add a trade pair":"请添加一个交易对","Please enter 0 fee. Positive fees not supported yet for RL.":"请输入费用数字0。正费用目前在RL上还不支持。","Please enter a coin":"Please enter a coin","Please enter a filename":"请输入文件名","Please enter a passphrase":"Please enter a passphrase","Please enter a pubkey":"请输入公钥","Please enter a puzzlehash":"请输入谜题哈希","Please enter a valid CAT name":"Please enter a valid CAT name","Please enter a valid TAIL":"Please enter a valid TAIL","Please enter a valid asset id":"Please enter a valid asset id","Please enter a valid initial coin amount":"请输入一个有效的初始币数量","Please enter a valid integer of 0 or greater for the number of Backup IDs needed for recovery.":"请输入一个大于或者等于 0 的整数，用于恢复所需的备份ID。","Please enter a valid numeric amount":"请输入一个正确的数目","Please enter a valid numeric amount.":"请输入有效的数字金额。","Please enter a valid numeric fee":"请输入一个正确的费用","Please enter a valid numeric interval length":"请输入一个有效的长度数字","Please enter a valid numeric spendable amount":"请输入一个有效的支出数目","Please enter a valid pubkey":"请输入一个正确的公钥","Please enter a valid token name":"Please enter a valid token name","Please enter valid wallet name":"Please enter valid wallet name","Please enter your passphrase":"Please enter your passphrase","Please enter your passphrase to proceed":"Please enter your passphrase to proceed","Please finish syncing before making a transaction":"同步完成后才能发起交易。","Please select amount":"请选择金额","Please select backup file first":"请先选择备份文件","Please select buy or sell":"请选择买或卖","Please select coin colour":"请选择币的色彩","Please select coin type":"Please select coin type","Please specify final directory":"请指定最终输出文件夹","Please specify temporary directory":"请指定临时目录","Please wait for synchronization":"请等待同步","Please wait for wallet synchronization":"请等待钱包同步","Plot":"农田","Plot Count":"农田数","Plot Id":"农田ID","Plot Key":"农田密钥","Plot NFT":"农田NFT","Plot NFT with p2_singleton_puzzle_hash {plotNFTId} does not exists":["p2_singleton_puzzle_hash为 ",["plotNFTId"]," 的农田NTF不存在。"],"Plot Name":"Plot Name","Plot Public Key":"农田公钥","Plot Seed":"Plot Seed","Plot Size":"农田大小","Plot in Parallel":"并发开垦","Plot is duplicate of {0}":["农田与 ",["0"]," 重复"],"PlotNFT is transitioning to (target state)":"PlotNFT is transitioning to (target state)","PlotNFT is transitioning to (target state). This can take a while. Please do not close the application until this is finalized.":"农田NFT正在过渡到（目标状态）。这可能需要一段时间。请在完成之前不要关闭应用程序。","Plots":"农田","Plots Passed Filter":"通过初筛的农田","Plots are allocated space on your hard drive used to farm and earn Chia. <0>Learn more</0>":"农田就是在你的硬盘上分配的一块空间，用来收获奇亚币。<0>了解更多</0>","Plotting":"正在开垦","Plotting in parallel can save time. Otherwise, add plot(s) to the queue.":"并发开垦可以有效节省开垦时间但对系统资源要求很高。系统硬件不够的话建议选择按队列顺序开垦。","Plotting with bitfield enabled has about 30% less overall writes and is now almost always faster. You may see reduced memory requirements with bitfield plotting disabled. If your CPU design is from before 2010 you may have to disable bitfield plotting.":"在开垦农田时启用位域会减少约30%的磁盘写入，而且大多数情况下会更快。但如果您在开垦时禁用位域，占用的内存则可能会减少。如果您的CPU是2010年以前的型号，您可能必须禁用位域才能开垦。","Plotting with bitfield enabled has about 30% less overall writes and is now almost always faster. You may see reduced memory requirements with bitfield plotting disabled. If your CPU design is from before 2010 you may have to disbale bitfield plotting.":"Plotting with bitfield enabled has about 30% less overall writes and is now almost always faster. You may see reduced memory requirements with bitfield plotting disabled. If your CPU design is from before 2010 you may have to disbale bitfield plotting.","Points Found Since Start":"从开始以来获取的积点","Points Found in Last 24 Hours":"最近24小时获取的积点","Points Successful in Last 24 Hours":"最近24小时成功获取的积点","Pool":"农业合作社","Pool Contract Address":"农业合作社合约地址","Pool Key":"池密钥","Pool Login Link":"农业合作社登录链接","Pool Payout Instructions":"农业合作社支付说明","Pool Public Key":"池公钥","Pool Puzzle Hash":"池谜语哈希","Pool Reward Address":"池奖励地址","Pool Reward Address is not properly formatted.":"池奖励地址格式不正确。","Pool Reward Address is not properly formatted. Using 'paste' is recommended.":"Pool Reward Address is not properly formatted. Using 'paste' is recommended.","Pool Reward Address must not be empty.":"池奖励地址不能为空。","Pool Reward Amount":"池奖励金额","Pool does not provide relative_lock_height.":"农业合作社未提供 relative_lock_height。","Pool does not provide target_puzzle_hash.":"农业合作社未提供 target_puzzle_hash。","Pool public key:":"奖励池公钥：","Pool:":"农业合作社：","Pooling":"合作社耕作","Pooling Wallet":"Pooling Wallet","Port":"端口","Preparing Plot NFT":"正在准备农田NFT","Preparing standard wallet":"正在准备标准钱包","Previous":"上一个","Previous Block Hash":"Previous Block Hash","Previous Header Hash":"上个区块的头部哈希","Previous Sub Block Hash":"Previous Sub Block Hash","Private key with public fingerprint {fingerprint}":["公共指纹为 ",["fingerprint"]," 的私钥"],"Private key {0}":["私钥 ",["0"]],"Private key {fingerprint}":["Private key ",["fingerprint"]],"Private key:":"私钥:","Proof of Space Size":"空间证明的大小","Proofs Found":"已找到证明","Protocol Version":"协议版本","Pubkey":"公钥","Public key:":"公钥：","Puzzlehash":"谜题哈希","Queue Name":"队列名","Queue name":"队列名","Queued":"排队中","Quit":"Quit","RAM max usage":"内存最大使用量","RL Wallet":"RL Wallet","RLListItems.title":"RLListItems.title","Rate Limited":"费用限定","Rate Limited Info":"费用限制信息","Rate Limited Options":"费用限定型选项","Rate Limited User Wallet Setup":"费用限定型用户钱包设置","Receive Address":"接收地址","Recipient address is not a coloured wallet address. Please enter a coloured wallet address":"Recipient address is not a coloured wallet address. Please enter a coloured wallet address","Recover":"恢复","Recover DID Wallet":"恢复去中心化标识钱包","Recover DID wallet":"Recover DID wallet","Recover Distributed Identity Wallet":"恢复分布式身份钱包","Recover Wallet":"恢复钱包","Recoverable Wallet":"Recoverable Wallet","Recovery Wallet":"Recovery Wallet","Refresh Plots":"刷新农田列表","Relative Lock Height":"相对锁高度","Release Notes":"版本发布说明","Remove keys from old keyring upon successful migration":"Remove keys from old keyring upon successful migration","Removing":"删除中","Rename":"重命名","Rename Wallet":"Rename Wallet","Report an Issue...":"报告问题","Restore Metadata for Coloured Coins and other Smart Wallets from Backup":"从备份中恢复染色币与智能钱包的元信息","Restore Smart Wallets From Backup":"Restore Smart Wallets From Backup","Retire":"Retire","SB Height":"SB Height","Safe To Skip":"可安全跳过此步骤","Save":"保存","Save passphrase":"Save passphrase","Search":"Search","Search block by header hash":"按头部哈希搜索区块","Second temporary folder location":"第二临时文件夹位置","See private key":"查看私钥：","Seed:":"种子：","Select 2nd Temporary Directory":"选择第二临时文件夹","Select Final Directory":"选择最终输出文件夹","Select Key":"选择密钥","Select Offer":"选择出价","Select Temporary Directory":"选择临时文件夹","Select Wallet":"Select Wallet","Select Wallet Type":"选择钱包类别","Select the final destination for the folder where you would like the plot to be stored. We recommend you use a large slow hard drive (like external HDD).":"选择您希望保存农田文件的最终输出文件夹。我们建议您使用大容量、慢速的硬盘（比如外置机械硬盘）。","Select the temporary destination for the folder where you would like the plot to be stored. We recommend you use a fast SSD.":"Select the temporary destination for the folder where you would like the plot to be stored. We recommend you use a fast SSD.","Select the temporary destination for the folder where you would like the plot to be stored. We recommend you use a fast drive.":"Select the temporary destination for the folder where you would like the plot to be stored. We recommend you use a fast drive.","Select your Plot NFT":"选择你的农田NFT","Select your Plot NFT from the dropdown or create a new one.":"从下拉菜单中选择现有农田NFT，或者新建。","Selected":"已选择","Selected recovery file:":"选择恢复文件：","Self Pooling":"自耕种","Self pool. When you win a block you will earn XCH rewards.":"自耕种。一旦发现新区块，你将得到全部XCH区块奖励。","Sell":"卖出","Send":"发送","Send this info packet to your Rate Limited Wallet user who must use it to complete setup of their wallet:":"把这段信息发给你的费用限定型钱包用户, 以便他/她可以完成自己的钱包设置:","Send your pubkey to your Rate Limited Wallet admin:":"将你的公钥发送到你的费用限定型管理者钱包:","Show Advanced Options":"显示高级选项","Show Asset Id":"Show Asset Id","Show Token and Asset Issuance Limitations":"Show Token and Asset Issuance Limitations","Show app":"Show app","Side":"边","Sign In":"登录","Skip":"Skip","Skip Keyring Migration":"Skip Keyring Migration","Skips adding [final dir] to harvester for farming":"Skips adding [final dir] to harvester for farming","Skips adding a final directory to harvester for farming":"勾选此项，将不会自动把最终目录添加到收割机进行耕种","Smooth out your XCH farming rewards by joining a pool.":"加入农业合作社来稳定XCH耕作收益。","Something went wrong":"出错了","Speech":"演说","Spendable Amount":"可用金额","Spendable Amount Per Interval":"每次迭代的支出数目","Spendable Balance":"可用余额","Spending Interval (number of blocks): {interval}":["支出迭代(区块数)",["interval"]],"Spending Interval Length (number of blocks)":"支出迭代长度(区块数)","Spending Limit (chia per interval): {0}":["消费限额（单位时间可消费的奇亚币）：",["0"]],"Standard Wallet":"Standard Wallet","Starting Wallet":"Starting Wallet","Starting service {0}":["Starting service ",["0"]],"Starting {0}":["Starting ",["0"]],"State":"状态","Status":"状态","Status:":"状态：","StatusCard.connections":"StatusCard.connections","StatusCard.height":"StatusCard.height","StatusCard.status":"StatusCard.status","StatusCard.title":"StatusCard.title","Stripe Size":"Stripe Size","Sub Block Height":"Sub Block Height","Sub Block at height {0} in the Chia blockchain":["Sub Block at height ",["0"]," in the Chia blockchain"],"Submit":"提交","Success":"Success","Synced":"已同步","Syncing":"正在同步","Syncing <0/>/<1/>":"正在同步 <0/> / <1/>","Syncing {progress}/{tip}":["Syncing ",["progress"],"/",["tip"]],"TXCH Block Rewards":"TXCH Block Rewards","TXCH Farming Rewards":"TXCH Farming Rewards","TXCH Fees Collected":"TXCH Fees Collected","TXCH User Fees":"TXCH User Fees","TXCH User Transaction Fees":"TXCH User Transaction Fees","Target Puzzle Hash":"目标迷题哈希","Temporary files are created during the plotting process which exceeds the size of the final plot files. Make sure you have enough space. <0>Learn more</0>":"Temporary files are created during the plotting process which exceeds the size of the final plot files. Make sure you have enough space. <0>Learn more</0>","Temporary folder location":"临时文件夹位置","The application will stop working at block height 193536.":"这个版本会在区块高度达到193536时停止工作。","The full node that your farmer is connected to is below. <0>Learn more</0>":"你的农场连接的全节点如下.<0>了解更多</0>","The minimum required size for mainnet is k=32":"主网要求的最小农田大小为k=32","The node is not synced":"节点没有同步","The node is syncing, which means it is downloading blocks from other nodes, to reach the latest block in the chain":"节点正在同步中, 也就是说它正在从其他节点下载区块, 以便达到最新的区块高度","The number of Backup IDs needed for recovery cannot exceed the number of Backup IDs added.":"The number of Backup IDs needed for recovery cannot exceed the number of Backup IDs added.","The pool URL \"{normalizedUrl}\" is not working. Is it pool? Error: {0}":["农业合作社地址 ",["normalizedUrl"]," 无效。请确认这是正确的池地址。错误：",["0"]],"The pool URL is not valid. {normalizedUrl}":["此农业合作社地址无效。 ",["normalizedUrl"]],"The pool URL needs to use protocol https. {normalizedUrl}":["农业合作社地址需要使用 https 协议。",["normalizedUrl"]],"The pool URL speciefied is not valid. {normalizedUrl}":["The specified pool URL is not valid. ",["normalizedUrl"]],"The pool URL speciefied is not valid. {poolUrl}":["The pool URL specified is not valid. ",["poolUrl"]],"The pool switching operation was cancelled, please try again by changing pool, or self pooling":"切换农业合作社操作已被取消。请重新尝试切换农业合作社，或自耕种。","The seed used to create the plot. This depends on the pool pk and plot pk.":"密语用于创建农田。农田是与农场池的公钥与农田的公钥相关联的。","The total number of VDF (verifiable delay function) or proof of time iterations on the whole chain up to this sub block.":"截至该子区块为止全链已完成的VDF(延迟验证函数) 或时间证明的迭代次数.","The total number of VDF (verifiable delay function) or proof of time iterations on this block.":"本区块上花费的总VDF(延迟验证函数) 或时间验证的迭代次数.","The total transactions fees in this block. Rewarded to the farmer.":"本区块上奖励给农民的全部交易费用。","These are the instructions for how the farmer wants to get paid. By default this will be an XCH address, but it can be set to any string with a size of less than 1024 characters, so it can represent another blockchain or payment system identifier.":"这是农民耕种奖励的支付方式。默认情况下，这将是一个XCH钱包地址，但是你可以设置任意长度小于1024的字符串，也就是说，你可以指定其它区块链或者支付系统标识。","These plots are invalid, you might want to delete them.":"这些农田无效，你可能想要删除它们。","This access token is verified":"This access token is verified","This access token is whitelisted":"This access token is whitelisted","This allows you to add a directory that has plots in it. If you have not created any plots, go to the plotting screen.":"在这里可以把事先准备好的农田文件添加进来。如果还没有的话，就去开垦几块农田吧。","This difficulty is an artifically lower difficulty than on the real network, and is used when farming, in order to find more proofs and send them to the pool. The more plots you have, the higher difficulty you will have. However, the difficulty does not affect rewards.":"此难度是比实际网络难度人为降低的难度，它将被用于耕作中，用来发现尽量多的验证数据并发送给农业合作社。你的农田越多，你的难度就会越高。这个难度不会影响你的耕作奖励。","This feature is available only from the GUI.":"此功能只在图形界面中提供。","This is the amount of Chia that you can currently use to make transactions. It does not include pending farming rewards, pending incoming transactions, and Chia that you have just spent but is not yet in the blockchain.":"这是你目前可以交易的奇亚币数额。该数额不包含处理中的农场奖励、收入交易，以及未经区块链确认的支付交易。","This is the pending change, which are change coins which you have sent to yourself, but have not been confirmed yet.":"处理中的找零钱, 是你发送支出的找回, 但是没有被链上确认.","This is the sum of the incoming and outgoing pending transactions (not yet included into the blockchain). This does not include farming rewards.":"待处理的收入与支出总和(还未上链). 不包括农场耕种奖励.","This is the time of the latest peak sub block.":"最新的子块时间.","This is the time the block was created by the farmer, which is before it is finalized with a proof of time":"此时间是该区块被农民创建的时间，因此，此时间会早于其被时间认证机制最终确认的时间。","This is the total amount of chia in the blockchain at the current peak sub block that is controlled by your private keys. It includes frozen farming rewards, but not pending incoming and outgoing transactions.":"这里是你的私钥所控制的, 到目前最新子块高度为止的全部奇亚数. 包含被冻结的耕种奖励, 但不包含待处理的收入与支出.","This is the total balance + pending balance: it is what your balance will be after all pending transactions are confirmed.":"目前结余+待处理的结余: 在链上确认完成后就是你的最终总结余.","This is the total number of points this plotNFT has with this pool, since the last payout. The pool will reset the points after making a payout.":"这是此农田NFT自从上次支付奖励后在当前农业合作社中持有的总积点数。在农业合作社支付奖励后，此积点数会归零。","This is the total number of points your farmer has found for this plot NFT. Each k32 plot will get around 10 points per day, so if you have 10TiB, should should expect around 1000 points per day, or 41 points per hour.":"这是你的农民在这个农田NFT对应的农田中找到的总积点数。一个k32的农田每天大约会得到10个积点。也就是说，如果你拥有10TiB的农田，你每天大约会获得1000积点，也就是每小时41积点。","This node is fully caught up and validating the network":"本节点已完成同步，并已开始参与网络验证","This plot NFT is assigned to a different key. You can still create plots for this plot NFT, but you can not make changes.":"此农田NFT绑定了其它密钥。你仍然可以用此农田NFT开垦农田，但是你无法进行任何更改。","This plot NFT is not connected to pool":"此农田NFT尚未连接到农业合作社。","This table shows you the last time your farm attempted to win a block challenge.":"This table shows you the last time your farm attempted to win a block challenge.","This table shows you the last time your farm attempted to win a block challenge. <0>Learn more</0>":"这个表格里面显示的是你的农场尝试过的区块挑战. <0>了解更多</0>","This trade was created at this time":"本交易创建于","This trade was included on blockchain at this block height":"该交易在链上的区块高度","This version of Chia is no longer compatible with the blockchain and can not safely farm.":"此版本的奇亚客户端与区块链已经不再兼容，无法安全地种植。","Time Created":"创建时间","Time this trade was created at this time":"Time this trade was created at this time","Timestamp":"时间戳","To":"目标地址","Token and Asset Issuance Limitations":"Token and Asset Issuance Limitations","Token has empty asset id":"Token has empty asset id","Token has empty name":"Token has empty name","Token has empty tail":"Token has empty tail","Total Balance":"总余额","Total Chia Farmed":"Total Chia Farmed","Total Iterations":"总迭代次数","Total Network Space":"全网占用空间","Total Plot Size:":"总农田大小：","Total Size of Plots":"总农田大小：","Total VDF Iterations":"总VDF迭代数","Total iterations since the start of the blockchain":"自区块开始以来的迭代数","Trade":"Trade","Trade Details":"交易详情","Trade ID":"交易ID","Trade ID:":"交易ID：","Trade Overview":"交易概况","Trades will show up here":"交易会在这里显示","Trading":"交易","Trading History":"交易历史","Transaction has not been sent to node yet":"Transaction has not been sent to node yet","Transaction has sent to a full node and is pending inclusion into the mempool. {0}":["Transaction has sent to a full node and is pending inclusion into the mempool. ",["0"]],"Transaction has successfully been sent to a full node and included in the mempool.":"Transaction has successfully been sent to a full node and included in the mempool.","Transactions":"Transactions","Transactions Filter Hash":"交易过滤器哈希","Try Again":"Try Again","Type":"类别","Unable to create plot NFT":"无法创建农田NFT","Unable to load the list of the keys":"Unable to load the list of the keys","Unclaimed Rewards":"未申领奖励","Unfinished":"未完成","Unique identifier":"唯一识别码","Unknown":"未知","Unlock Keyring":"Unlock Keyring","Unsaved Changes":"未保存的更改","User Pubkey":"用户公钥","VDF Sub Slot Iterations":"VDF子项迭代数","Value is too high":"Value is too high","Value seems high":"值似乎过高","Verify Pool Details":"验证农业合作社详情","View":"查看","View Log":"查看日志","View Offer":"查看出价","View Payout Instructions":"View Payout Instructions","View Pool Login Link":"查看农业合作社登录地址","View pending balances":"查看处理中的余额","View pending balances...":"查看处理中的余额…","Wait for synchronization":"等待同步","Waiting for launch":"Waiting for launch","Waiting for network to launch":"Waiting for network to launch","Waiting for the transaction to be confirmed":"正在等待交易确认","Wallet Status:":"钱包状态：","Wallet does not exists":"钱包不存在","Wallet id is not defined":"Wallet id is not defined","WalletItem.ccWallet":"WalletItem.ccWallet","WalletItem.chiaWallet":"WalletItem.chiaWallet","WalletItem.didWallet":"WalletItem.didWallet","WalletItem.rlWallet":"WalletItem.rlWallet","WalletStatusCard.not_synced":"WalletStatusCard.not_synced","WalletStatusCard.synced":"WalletStatusCard.synced","WalletStatusCard.syncing":"WalletStatusCard.syncing","Wallets":"钱包","Wallets.title":"Wallets.title","Want to Join a Pool? Create a Plot NFT":"想要加入农业合作社？创建一个农田NFT","Want to earn more Chia? Add more plots to your farm.":"想要获得更多奇亚币？ 给你的农场多开垦一些农田吧。","Want to have a delay before the next plot starts?":"想要在开垦下一块农田前等待一会？","Warning: This key is used for a wallet that may have a non-zero balance. By deleting this key you may lose access to this wallet":"警告：此密钥用于可能有余额的钱包。删除此密钥可能会无法访问此钱包。","Warning: This key is used for your farming rewards address. By deleting this key you may lose access to any future farming rewards":"警告：此密钥用于耕作奖励地址。删除此密钥后可能会无法提取将来的耕作奖励。","Warning: This key is used for your pool rewards address. By deleting this key you may lose access to any future pool rewards":"警告：此密钥用于池奖励地址。删除此密钥可能会无法获得将来的池奖励。","Weight":"权重","Weight is the total added difficulty of all sub blocks up to and including this one":"权重是施加给所有子块包括目前这块的额外的困难度.","Welcome to Chia. Please log in with an existing key, or create a new key.":"欢迎来到奇亚。请用已有的密钥登录，或者创建一个新的密钥。","Welcome! The following words are used for your wallet backup. Without them, you will lose access to your wallet, keep them safe! Write down each word along with the order number next to them. (Order is important)":"欢迎使用奇亚! 下面的词语列表是你的钱包密语, 请务必严格完整的把它们记下来(保持顺序). 没有这段密语的话, 你就没办法找回钱包里的币. 千万也不要透露给别人.","When you receive the setup info packet from your admin, enter it below to complete your Rate Limited Wallet setup:":"当你从管理员那收到这个设置信息后, 请把它输入到下面以便完成费用限定型钱包设置:","Window":"窗口","Without fees":"不含交易费","Yes":"是","You are in pending state. Please wait for confirmation":"你当前处于待处理状态，请等待区块链确认。","You are not self pooling":"你目前没有自耕种","You can still create plots for this plot NFT, but you can not make changes until sync is complete.":"你现在可以使用这个农田NFT开垦新农田，但是在同步完成前，你不能进行任何更改。","You do not need to be synced or connected to plot. Temporary files are created during the plotting process which exceed the size of the final plot files. Make sure you have enough space. <0>Learn more</0>":"You do not need to be synced or connected to plot. Temporary files are created during the plotting process which exceed the size of the final plot files. Make sure you have enough space. <0>Learn more</0>","You do not need to be synched or connected to Plot. Temporary files are created during the plotting process which exceed the size of the final plot files. Make sure you have enough space. <0>Learn more</0>":"You do not need to be synched or connected to Plot. Temporary files are created during the plotting process which exceed the size of the final plot files. Make sure you have enough space. <0>Learn more</0>","You do not need to be synched or connected to plot. Temporary files are created during the plotting process which exceed the size of the final plot files. Make sure you have enough space. <0>Learn more</0>":"不需要同步或者连接到网络节点就可以开垦农田。在开垦过程中产生的临时文件尺寸会大于最终的农田大小。请确认你有足够的存储空间。<0>了解更多</0>","You have made changes. Do you want to discard them?":"你已经做了一些更改。你确定要放弃这些修改吗？","You have {0}% of the space on the network, so farming a block will take {expectedTimeToWin} in expectation.":["You have ",["0"],"% of the space on the network, so farming a block will take ",["expectedTimeToWin"]," in expectation."],"You have {0}% of the space on the network, so farming a block will take {expectedTimeToWin} in expectation. Actual results may take 3 to 4 times longer than this estimate.":["您在网络上拥有",["0"],"%的空间，因此收割一个区块所花费的时间在数学期望上是",["expectedTimeToWin"],"。实际耗时可能比这个期望值长3到4倍。"],"You have {0}% of the space on the network, so farming a block will take {expectedTimeToWin} in expectation. Actual results may take 3-4 times longer - actual results may be 3x this estimate.<0>Learn about Expected Value</0>":["You have ",["0"],"% of the space on the network, so farming a block will take ",["expectedTimeToWin"]," in expectation. Actual results may take 3-4 times longer - actual results may be 3x this estimate.<0>Learn about Expected Value</0>"],"You have {0}% of the space on the network, so farming a block will take {expectedTimeToWin} in expectation. Actual results may take 3-4 times longer this estimate. <0>Learn about Expected Value</0>":["You have ",["0"],"% of the space on the network, so farming a block will take ",["expectedTimeToWin"]," in expectation. Actual results may take 3-4 times longer this estimate. <0>Learn about Expected Value</0>"],"You need to claim your rewards first":"你需要先领取你的奖励。","You need to upgrade the Chia application as this version will stop working soon!":"You need to upgrade the Chia application as this version will stop working soon!","You need {currencyCode} to join a pool.":["必须持有 ",["currencyCode"]," 才有加入农业合作社。"],"You will recieve <0/> to {0}":["你将会收到 <0/> 至 ",["0"]],"You will recieve <0/> to {address}":["你将会收到 <0/> 至 ",["address"]],"Your DID requires at least {dids_num_req} attestation file{0} for recovery. Please upload additional files.":["Your DID requires at least ",["dids_num_req"]," attestation file",["0"]," for recovery. Please upload additional files."],"Your Farm Overview":"农场概览","Your Full Node Connection":"你的全节点连接","Your Harvester Network":"你的收割网络","Your Pool Overview":"你的农业合作社概览","Your keyring is locked":"Your keyring is locked","Your keys have not been migrated to a new keyring. You will be unable to create new keys or delete existing keys until migration completes. Are you sure you want to skip migrating your keys?":"Your keys have not been migrated to a new keyring. You will be unable to create new keys or delete existing keys until migration completes. Are you sure you want to skip migrating your keys?","Your keys need to be migrated to a new keyring that is optionally secured by a master passphrase.":"Your keys need to be migrated to a new keyring that is optionally secured by a master passphrase.","Your keys need to be migrated to a new keyring that is secured by a master passphrase.":"Your keys need to be migrated to a new keyring that is secured by a master passphrase.","Your passphrase can be stored in your system's secure credential store. Chia will be able to access your keys without prompting for your passphrase.":"Your passphrase can be stored in your system's secure credential store. Chia will be able to access your keys without prompting for your passphrase.","[Error 13] Permission denied. You are trying to access a file/directory without having the necessary permissions. Most likely one of the plot folders in your config.yaml has an issue.":"[错误 13] 权限被拒绝。您正试图访问的文件或目录没有必需的权限。农田文件夹中的 config.yaml 文件可能存在问题。","[Error 22] File not found. Most likely one of the plot folders in your config.yaml has an issue.":"[错误 22] 文件未找到。农田文件夹中的 config.yaml 文件可能存在问题。","connections:":"连接：","height:":"高度：","not synced":"未同步","p2_singleton_puzzle_hash is not defined":"p2_singleton_puzzle_hash 未定义","status:":"状态：","synced":"已同步","syncing":"同步中","{0} Block Rewards":[["0"]," Block Rewards"],"{0} User Transaction Fees":[["0"]," User Transaction Fees"],"{0} points {1} - {2} hours ago":[["0"]," 积点 ",["1"]," - ",["2"]," 小时前"],"{currencyCode} Block Rewards":[["currencyCode"]," 区块奖励"],"{currencyCode} Total Chia Farmed":[["currencyCode"]," 收获的奇亚币总计"],"{currencyCode} User Transaction Fees":[["currencyCode"]," 用户交易费用"],"{mojo, plural, one {mojo} other {mojos}}":[["mojo","plural",{one:"mojo",other:"mojos"}]]}};
+/*eslint-disable*/module.exports={messages:{"(Optional)":"（可选）","*Want to explore Chia’s blocks further? Check out <0>Chia Explorer</0> built by an open source developer.":"* 想要进一步了解奇亚？来看看一位开源开发者创建的 <0>奇亚区块浏览器</0>。","+":"+","+ Add New Plot NFT":"+ 添加新的农田NFT","+ Add Token":"+ Add Token","+ Add Wallet":"+ 添加钱包","-":"-","0 automatically chooses bucket count":"0 automatically chooses bucket count","128 buckets is recommended":"排序桶数量推荐设置为128","A harvester is a service running on a machine where plot(s) are actually stored. A farmer and harvester talk to a full node to see the state of the chain. View your network of connected harvesters below Learn more":"收割机是一个在存有农田文件的机器上运行的服务。农民和收割机通过和全节点交互来得到链的状态。你可以在下面看到你网络所连接的收割机。了解更多","About Chia Blockchain":"关于奇亚Chia区块链项目","About Chia Wallet":"About Chia Wallet","Accept":"接受","Accept Offer":"Accept Offer","Accepted Offers":"Accepted Offers","Accepted at time:":"接受时间：","Accepted on Date":"Accepted on Date","Action":"操作","Actions":"操作","Add":"添加","Add Backup ID":"添加备份 ID","Add CAT wallets to have more options":"Add CAT wallets to have more options","Add Custom Token":"Add Custom Token","Add Plot Directory":"添加农田文件夹","Add Plot to Queue":"将农田开垦任务加入队列","Add Token":"Add Token","Add Wallet":"Add Wallet","Add a Plot":"开垦农田","Add a Plot NFT":"开垦农田NTF","Add a plot":"开垦农田","Add plot directory":"添加农田文件夹","Add {currencyCode} from the Faucet":["从水龙头获取 ",["currencyCode"]],"Adding {0} token":["Adding ",["0"]," token"],"Address":"地址","Address / Puzzle hash":"地址 / 谜语哈希","After your keys are successfully migrated to the new keyring, you may choose to have your keys removed from the old keyring.":"After your keys are successfully migrated to the new keyring, you may choose to have your keys removed from the old keyring.","All":"All","Amount":"数量","Amount ({cc_unit})":["数量 (",["cc_unit"],")"],"Amount For Initial Coin":"初始币数量","Amount exceeds spendable balance":"Amount exceeds spendable balance","Amount must be an even amount.":"Amount must be an even amount.","Are you sure you want to cancel your offer?":"Are you sure you want to cancel your offer?","Are you sure you want to delete the plot? The plot cannot be recovered.":"你确定要删除这个农田吗？删除后将不可恢复。","Are you sure you want to delete unconfirmed transactions?":"确定要删除未确认的交易吗？","Are you sure you want to disconnect?":"是否要断开连接？","Are you sure you want to quit?":"Are you sure you want to quit?","Are you sure you want to quit? GUI Plotting and farming will stop.":"确定要退出吗? 农田开垦和耕种将会停止。","Are you sure you want to use k={plotSize}":["Are you sure you want to use k=",["plotSize"]],"Are you sure you want to use k={plotSize}?":["你确定要使用 k=",["plotSize"],"？"],"Asset Id":"Asset Id","Asset Type":"Asset Type","Atomic Swap Wallet":"Atomic Swap Wallet","Attempt {attempt}":["Attempt ",["attempt"]],"Authorized Payee Wallet":"Authorized Payee Wallet","Autogenerated name from pool contract address":"从农业合作社合约地址自动生成的名称","Back":"返回","Backup file is used to restore smart wallets.":"备份文件将用于恢复智能钱包。","Balance":"余额","Base Farmer Reward Amount":"基础耕种奖励数量","Before posting an offer in Keybase to the #{KeybaseChannelName} channel, you must first join the {KeybaseTeamName} team. Please note that it might take a few moments to join the channel.":["Before posting an offer in Keybase to the #",["KeybaseChannelName"]," channel, you must first join the ",["KeybaseTeamName"]," team. Please note that it might take a few moments to join the channel."],"Below are the current block challenges. You may or may not have a proof of space for these challenges. These blocks do not currently contain a proof of time.":"下面是当前区块的挑战。 你的农田或许包含了对这些挑战的空间证明。这些区块目前还未包含时间证明。","Best estimate over last 24 hours":"最近24小时的最佳估值","Best estimate over last 5 hours":"Best estimate over last 5 hours","Block":"区块","Block Test":"区块测试","Block VDF Iterations":"区块可验证延迟函数迭代","Block at height {0} in the Chia blockchain":["奇亚区块链上高度为 ",["0"]," 的区块"],"Block hash":"Block hash","Block with hash {headerHash}":["区块哈希",["headerHash"]],"Block with hash {headerHash} does not exist.":["哈希为 ",["headerHash"]," 的区块不存在。"],"Blocks":"区块","Both values are empty. This is ok if you are, for example, sending your rewards to an offline wallet.":"Both values are empty. This is ok if you are, for example, sending your rewards to an offline wallet.","Browse":"浏览","But you are currently farming <0/>":"但你目前正在耕种<0/>","Buy":"购买","CAT Wallet":"CAT Wallet","CC Wallet":"CC Wallet","CCListItems.title":"CCListItems.title","Can be backed up to mnemonic seed":"可备份为助记词","Cancel":"取消","Cancel Offer":"Cancel Offer","Cancel and Spend":"取消并支付","Cancel on blockchain":"Cancel on blockchain","Cannot send chia to coloured address. Please enter a chia address.":"Cannot send chia to coloured address. Please enter a chia address.","Caution, deleting these plots will delete them forever. Check that the storage devices are properly connected.":"注意，删除这些农田的操作是永久的。请检查存储设备连接正常。","Challenge":"挑战","Challenge Hash":"挑战哈希","Change":"变更","Change Pool":"更换农业合作社","Chat on KeyBase":"在Keybase上参与讨论","Chia":"奇亚","Chia Asset Token":"Chia Asset Token","Chia Blockchain Wiki":"奇亚区块链项目维基","Chia Wallet":"奇亚钱包","Choose Number of Plots":"选择要开垦的数量","Choose Plot Size":"选择农田文件大小","Claim Rewards":"申领奖励","Close":"关闭","Closing down node and server":"正在关闭节点与服务","Coin Name":"代币名称","Coins":"Coins","Coins:":"币：","Colour":"颜色","Colour Info":"色彩信息","Colour String":"色彩字符串","Colour:":"色彩：","Coloured Coin":"染色币","Coloured Coin Options":"染色币选项","Confirm":"确认","Confirm Disconnect":"确认断开连接","Confirm Passphrase":"Confirm Passphrase","Confirmation":"确认","Confirmed":"Confirmed","Confirmed at Height":"Confirmed at Height","Confirmed at block:":"区块上已确认：","Confirmed at height {0}":["确认于区块高度 ",["0"]],"Connect":"连接","Connect to other peers":"连接到其它节点","Connect to pool":"连接到农业合作社","Connected":"已连接","Connecting do daemon":"Connecting do daemon","Connecting to daemon":"Connecting to daemon","Connecting to wallet":"正在连接到钱包","Connection Status":"连接状态","Connection Status:":"连接状态：","Connection type":"连接类型","Connections":"连接数","Contribute on GitHub":"在 GitHub 上贡献","Copied":"已复制","Copy":"复制","Copy to Clipboard":"复制到剪贴板","Create":"创建","Create An Attestation Packet":"Create an Attestation Packet","Create Chia Asset Token Wallet":"Create Chia Asset Token Wallet","Create Chia Asset Token Wallet from Existing TAIL":"Create Chia Asset Token Wallet from Existing TAIL","Create DID wallet":"Create DID wallet","Create Distributed Identity Wallet":"创建分布式身份钱包","Create New Chia Asset Token Wallet":"Create New Chia Asset Token Wallet","Create New Wallet":"新建钱包","Create Offer":"创建报价","Create Plot":"Create Plot","Create Rate Limited Admin Wallet":"创建费用限定型管理钱包","Create Rate Limited User Wallet":"创建费用限定型用户钱包","Create Token":"Create Token","Create Trade":"Create Trade","Create Trade Offer":"发起交易出价","Create Transaction":"发起交易","Create a Backup":"创建备份","Create a Plot NFT":"创建农田NFT","Create a new private key":"创建新私钥","Create admin wallet":"创建管理者钱包","Create an Offer":"Create an Offer","Create an offer to exchange XCH or other tokens. View an offer to inspect and accept an offer made by another party.":"Create an offer to exchange XCH or other tokens. View an offer to inspect and accept an offer made by another party.","Create custom CAT Wallet":"Create custom CAT Wallet","Create from Existing":"Create from Existing","Create new coloured coin":"创建新的染色币","Create user wallet":"创建用户钱包","Create wallet for colour":"创建染色钱包","Create wallet for existing colour":"基于已有的染色币来创建一个钱包","CreateWallet.addWallet":"CreateWallet.addWallet","Created At:":"创建于：","Created by us:":"创建人：","Creates and submits a transaction on the blockchain that cancels the offer":"Creates and submits a transaction on the blockchain that cancels the offer","Creating Plot NFT and Joining the Pool":"正在创建农田NFT并加入农业合作社","Creating Plot NFT for Self Pooling":"正在创建自耕种的农田NFT","Creation Date":"Creation Date","Currency code is not defined":"货币代码未定义","Current Difficulty":"当前难度","Current Points Balance":"当前积点余额","Current trade status":"当前交易状态","Custody Wallet":"Custody Wallet","Custom":"Custom","DANGER: permanently delete private key":"注意：永久性删除私钥","DID Wallet":"DID Wallet","DIDListItems.title":"DIDListItems.title","Date":"日期","Delay":"延迟","Delete":"删除","Delete Key":"Delete Key","Delete Plot":"删除农田","Delete Unconfirmed Transactions":"删除未确认的交易","Delete all keys":"删除所有密钥","Delete key":"Delete key","Delete key {fingerprint}":["删除密钥 ",["fingerprint"]],"Deleting all keys will permanently remove the keys from your computer, make sure you have backups. Are you sure you want to continue?":"删除所有密钥将会把你的密钥从你的计算机中永久移除，请确保你已经将其备份。确定继续？","Deleting key":"Deleting key","Deleting the key will permanently remove the key from your computer, make sure you have backups. Are you sure you want to continue?":"删除此密钥将会把密钥从你的计算机中永久移除，请确保你已经将其备份。确定继续？","Details":"Details","Developer":"开发者","Developer Tools":"开发者工具","Difficulty":"难度","Disable bifield plotting when your temporary directory is on fast storage like SSD or you need lower per process RAM requirements. Plotting with bitfield enabled has about 20% less overall writes.":"Disable bifield plotting when your temporary directory is on fast storage like SSD or you need lower per process RAM requirements. Plotting with bitfield enabled has about 20% less overall writes.","Disable bitfield plotting":"开垦时禁用位域","Discard":"放弃","Disconnect":"断开连接","Display Offer Data":"Display Offer Data","Distributed Identity":"分布式身份认证","Distributed Identity Options":"Distributed Identity Options","Do not show this dialog again":"Do not show this dialog again","Do you have existing plots on this machine? <0>Add Plot Directory</0>":"你的机器上有已经完成的农田文件吗？ <0>添加农田文件路径</0>","Does your machine support parallel plotting?":"您的机器是否支持并行开垦？","Drag & drop an offer file below to view its details":"Drag & drop an offer file below to view its details","Drag and drop attestation packet(s)":"Drag and drop attestation packet(s)","Drag and drop offer file":"拖放报价文件","Drag and drop your backup file":"拖放你的备份文件","Drag and drop your recovery backup file":"Drag and drop your recovery backup file","Edit":"编辑","Edit Payout Instructions":"编辑支付指令","Enter a strong passphrase and click Migrate Keys to secure your keys":"Enter a strong passphrase and click Migrate Keys to secure your keys","Enter the 24 word mnemonic that you have saved in order to restore your Chia wallet.":"请输入你保存的24个助记词来恢复奇亚钱包","Error":"错误","Error the entered address appears to be for a different colour.":"Error the entered address appears to be for a different colour.","Error: Cannot send chia to coloured address. Please enter a chia address.":"错误：无法发送奇亚到染色地址。请输入奇亚钱包地址。","Estimated Network Space":"全网已占用空间（估算）","Estimated Time to Win":"预计区块发现时间","Estimated network space":"Estimated network space","Estimated sum of all the plotted disk space of all farmers in the network":"估计网络中所有农民已开垦农田占用磁盘空间量","Exclude final directory":"不耕种最终输出文件夹","Expected Time to Win":"Expected Time to Win","Export Offer File":"Export Offer File","Failed to execute Keybase command: {stderr}":["Failed to execute Keybase command: ",["stderr"]],"Failed to open (invalid plots)":"打开失败（农田无效）","Farm":"农场","Farmer Public Key":"农民公钥","Farmer Puzzle Hash":"农民解谜哈希","Farmer Reward Address":"农民奖励地址","Farmer Reward Address is not properly formatted.":"农民奖励地址格式不正确。","Farmer Reward Address is not properly formatted. Using 'paste' is recommended.":"Farmer Reward Address is not properly formatted. Using 'paste' is recommended.","Farmer Reward Address must not be empty.":"农民奖励地址不能为空。","Farmer is not connected":"农民节点无法连接","Farmer is not running":"农民节点没有运行","Farmer public key:":"农民公钥：","Farmers earn block rewards and transaction fees by committing spare space to the network to help secure transactions. This is where your farm will be once you add a plot. <0>Learn more</0>":"农民们通过贡献自己的空余存储空间，可以帮助提高整个网络的安全性，进而获得区块奖励和交易费用的奖励。当你的第一块农田开垦完成后，这里就是你的农场。<0>了解更多</0>","Farming":"正在耕种","Farming Status":"耕种状态","Fee":"费用","Fee (TXCH)":"Fee (TXCH)","Fee ({0})":["Fee (",["0"],")"],"Fee ({currencyCode})":["交易费 (",["currencyCode"],")"],"Fees Amount":"费用","Fees Puzzle Hash":"Fees Puzzle Hash","File":"文件","Filename":"文件名","Final folder location":"最终输出文件夹位置","Finished":"已完成","Follow on Twitter":"关注我们的Twitter","Frequently Asked Questions":"常见问题解答","Full Node":"全节点","Full Node Overview":"全节点概览","Full Node Status":"Full Node Status","Full Screen":"全屏","Generate New Colour":"生成新的色彩","Get Pool Login Link":"Get Pool Login Link","Go to #{KeybaseChannelName}":["Go to #",["KeybaseChannelName"]],"Grid view":"网格视图","HD or Hierarchical Deterministic keys are a type of public key/private key scheme where one private key can have a nearly infinite number of different public keys (and therefor wallet receive addresses) that will all ultimately come back to and be spendable by a single private key.":"分层确定性钱包中的私钥可以对应有无限多个公钥（也就是你的钱包接受地址），这所有的公钥都受这个单一的私钥控制。","Harvester ID":"Harvester ID","Harvester Plots":"收割机农田","Hashgreen DEX URL":"Hashgreen DEX URL","Header Hash":"头部哈希","Header hash":"头部哈希","Height":"高度","Help":"帮助","Help translate":"改进翻译","Hide Advanced Options":"隐藏高级选项","Hint":"Hint","History":"历史","Host Name":"主机名","IP address":"IP地址","IP address / host":"IP地址 / 主机","If none selected, then it will default to the temporary directory.":"如果未指定，将会使用上述临时目录。","If selected, your offer will be not be shared publicly.":"If selected, your offer will be not be shared publicly.","If you have already shared your offer file, you may need to submit a transaction to cancel the pending offer. Click \"Cancel on blockchain\" to submit a cancellation transaction.":"If you have already shared your offer file, you may need to submit a transaction to cancel the pending offer. Click \"Cancel on blockchain\" to submit a cancellation transaction.","Import":"Import","Import Wallet from Mnemonics":"使用助记词导入钱包","Import from Mnemonics (24 words)":"自助记词导入（24个词）","In Progress":"处理中","In exchange for":"In exchange for","In exchange for:":"In exchange for:","Incoming":"收入","Incorrect value":"数值错误","Index":"索引","Indicated if this offer was created by us":"表明此出价是否由我们创建","Indicated what time this offer was accepted":"表明次出价被接受的时间","Info Packet":"信息包","Initial Amount":"初始金额","Initialize a Rate Limited User Wallet:":"初始化费用限定型用户钱包：","Install Keybase":"Install Keybase","Insufficient funds available to accept offer. Ensure that your\nspendable balance is sufficient to cover the offer amount.":"Insufficient funds available to accept offer. Ensure that your\nspendable balance is sufficient to cover the offer amount.","Insufficient funds available to create offer. Ensure that your\nspendable balance is sufficient to cover the offer amount.":"Insufficient funds available to create offer. Ensure that your\nspendable balance is sufficient to cover the offer amount.","Interval":"间隔","Invalid state":"无效状态","It is a one-time login link that can be used to log in to a pool's website. It contains a signature using the farmer's key from the plot NFT. Not all pools support this feature.":"这是一个用于登录农业合作社网站的一次性链接。此链接包含农田NFT对应的农民密钥的签名。并非所有耕种池都支持此功能。","Join Pool":"加入农业合作社","Join a Pool":"加入农业合作社","Join a pool and get consistent XCH farming rewards. The average returns are the same, but it is much less volatile.":"加入农业合作社来获得持续的XCH耕种收益。平均回报是相同的，但是收益会更加稳定。","Join a pool and get consistent XCH farming rewards. The average returns are the same, but it is much less volatile. Assign plots to a plot NFT. You can easily switch pools without having to re-plot.":"加入农业合作社来获得持续的XCH耕种收益。平均回报是相同的，但是收益会更加稳定。为已开垦的农田分配农田NFT。你可以轻松切换耕种池，而不比重新开垦现有农田。","Join a pool and get more consistent XCH farming rewards. Create a plot NFT and assign your new plots to a group.":"加入农业合作社来获得持续的XCH耕种收益。创建一个农田NFT，然后将新农田分配到组。","Join {KeybaseTeamName}":["Join ",["KeybaseTeamName"]],"K-Size":"K值大小","Keybase command failed {e}. If you haven't installed Keybase, you can download from https://keybase.io":["Keybase command failed ",["e"],". If you haven't installed Keybase, you can download from https://keybase.io"],"Keybase is a secure messaging and file sharing application. To share an offer in the Keybase {KeybaseTeamName} team, you must first have Keybase installed.":["Keybase is a secure messaging and file sharing application. To share an offer in the Keybase ",["KeybaseTeamName"]," team, you must first have Keybase installed."],"Keyring migration failed: {0}":["Keyring migration failed: ",["0"]],"Keys":"密钥","KiB Up/Down":"KiB 上传/下载","Last Attempted Proof":"最近尝试过的证明","Last Height Farmed":"最新收割的区块高度","Latest Block Challenges":"最新的区块挑战","Launcher Id":"启动器ID","Learn More":"了解更多","Leaving Pool":"退出农业合作社","List of connections is empty":"List of connections is empty","List view":"列表视图","Loading Plot NFTs":"正在加载农田NFT","Loading configuration":"Loading configuration","Loading keyring status":"Loading keyring status","Loading list of the keys":"Loading list of the keys","Loading list of wallets":"正在加载钱包列表","Loading...":"正在载入……","Local Harvester Plots":"Local Harvester Plots","Log Out":"Log Out","Logging in":"正在登录","MainWalletList.colouredCoin":"MainWalletList.colouredCoin","MainWalletList.createAdminWallet":"MainWalletList.createAdminWallet","MainWalletList.createDIDWallet":"MainWalletList.createDIDWallet","MainWalletList.createNewColouredCoin":"MainWalletList.createNewColouredCoin","MainWalletList.createUserWallet":"MainWalletList.createUserWallet","MainWalletList.createWalletForExistingColour":"MainWalletList.createWalletForExistingColour","MainWalletList.distributedIdentity":"MainWalletList.distributedIdentity","MainWalletList.rateLimited":"MainWalletList.rateLimited","MainWalletList.recoverDIDWallet":"MainWalletList.recoverDIDWallet","MainWalletList.title":"MainWalletList.title","Manage Farming Rewards":"管理耕种奖励","Manage Offers":"Manage Offers","Manage Recovery DIDs":"管理恢复去中心化标识","Manage Your Farming Rewards Target Addresses":"管理耕种奖励发放地址","Memo":"Memo","Memos":"Memos","MiB Up/Down":"MiB 上传/下载","Migrate Keys":"Migrate Keys","Migration required":"Migration required","Minimum Difficulty":"最小难度","Minutes":"分钟","More memory slightly increases speed":"增加内存可以略微提高速度","Multi Sig Wallet":"Multi Sig Wallet","My DID Wallet":"我的去中心化标识钱包","My Pubkey":"我的公钥","Name":"Name","Network Name":"网络名称","Network has not yet been launched, once it launches full node will start automatically.":"Network has not yet been launched, once it launches full node will start automatically.","Network not launched yet, waiting for genesis challenge!":"Network not launched yet, waiting for genesis challenge!","New Address":"新地址","New Wallet":"新钱包","Next":"下一步","Nickname":"昵称","No":"否","No 24 word seed, since this key is imported.":"没有24个助记词种子，因为此密钥已导入。","No blocks farmed yet":"尚未有区块有收获","No current offers":"No current offers","No previous transactions":"无过往交易","No private keys for one or both addresses. Safe only if you are sending rewards to another wallet.":"本机没有所输入地址对应的私钥。请确认是否要将奖励发送至非本机钱包。","Node Count":"Node Count","Node ID":"节点ID","Node Id":"节点ID","None":"无","None of your plots have passed the plot filter yet.":"你的农田还没有通过过滤器检查.","Not Available":"不可用","Not Synced":"未同步","Not accepted":"Not accepted","Not accepted yet":"尚未被接受","Not confirmed":"Not confirmed","Not confirmed yet":"尚未被确认","Not connected":"未连接","Not found Plots":"未找到农田","Note that this does not change your pooling payout addresses. This only affects old format plots, and the 0.25XCH reward for pooling plots.":"请注意，这将不会改变你的合作社耕作支付地址。这只会影响旧农田文件，以及0.25XCH的合作社耕作奖励。","Number of Plots":"农田数","Number of buckets":"桶数量","Number of threads":"线程数量","OK":"确定","Offer":"出价","Offer Data":"Offer Data","Offer Identifier":"Offer Identifier","Offer Shared":"Offer Shared","Offer has been accepted and is awaiting confirmation.":"Offer has been accepted and is awaiting confirmation.","OfferBin URL":"OfferBin URL","Offered":"Offered","Offers Created":"已创建的出价","Offers you accepted":"Offers you accepted","Offers you created":"Offers you created","On average there is one minute between each transaction block. Unless there is congestion you can expect your transaction to be included in less than a minute.":"每个交易区块平均打包时间为一分钟。通常您的交易将在一分钟内被确认，除非出现拥堵。","Once you accept this offer, you will not be able to cancel the transaction. Are you sure you want to accept this offer?":"Once you accept this offer, you will not be able to cancel the transaction. Are you sure you want to accept this offer?","One or more unknown tokens are being offered. Please verify that the asset IDs of the tokens listed below match the asset IDs of the tokens you expect to receive.":"One or more unknown tokens are being offered. Please verify that the asset IDs of the tokens listed below match the asset IDs of the tokens you expect to receive.","Only one backup file is allowed.":"只允许一个备份文件。","Only one backup file is allowed. To add a new backup file, please delete any previously uploaded files and try again.":"Only one backup file is allowed. To add a new backup file, please delete any previously uploaded files and try again.","Outgoing":"支出","Parent Coin":"Parent Coin","Passphrase":"Passphrase","Passphrase Hint":"Passphrase Hint","Passphrase Hint (Optional)":"Passphrase Hint (Optional)","Paste Offer Data":"Paste Offer Data","Payout Address":"支付地址","Peak Height":"最高高度","Peak Sub-block Height":"Peak Sub-block Height","Peak Time":"最高时间","Pending":"待处理","Pending Balance":"待处理的余额","Pending Change":"待处理的找回","Pending Total Balance":"待处理总余额","Please Confirm":"请确认","Please add a trade pair":"请添加一个交易对","Please enter 0 fee. Positive fees not supported yet for RL.":"请输入费用数字0。正费用目前在RL上还不支持。","Please enter a coin":"Please enter a coin","Please enter a filename":"请输入文件名","Please enter a passphrase":"Please enter a passphrase","Please enter a pubkey":"请输入公钥","Please enter a puzzlehash":"请输入谜题哈希","Please enter a valid CAT name":"Please enter a valid CAT name","Please enter a valid TAIL":"Please enter a valid TAIL","Please enter a valid initial coin amount":"请输入一个有效的初始币数量","Please enter a valid integer of 0 or greater for the number of Backup IDs needed for recovery.":"请输入一个大于或者等于 0 的整数，用于恢复所需的备份ID。","Please enter a valid numeric amount":"请输入一个正确的数目","Please enter a valid numeric amount.":"请输入有效的数字金额。","Please enter a valid numeric fee":"请输入一个正确的费用","Please enter a valid numeric interval length":"请输入一个有效的长度数字","Please enter a valid numeric spendable amount":"请输入一个有效的支出数目","Please enter a valid pubkey":"请输入一个正确的公钥","Please enter a valid token name":"Please enter a valid token name","Please enter an amount for each row":"Please enter an amount for each row","Please enter valid wallet name":"Please enter valid wallet name","Please enter your passphrase":"Please enter your passphrase","Please enter your passphrase to proceed":"Please enter your passphrase to proceed","Please finish syncing before making a transaction":"同步完成后才能发起交易。","Please finish syncing before managing offers":"Please finish syncing before managing offers","Please select amount":"请选择金额","Please select an asset for each row":"Please select an asset for each row","Please select backup file first":"请先选择备份文件","Please select buy or sell":"请选择买或卖","Please select coin colour":"请选择币的色彩","Please select coin type":"Please select coin type","Please specify final directory":"请指定最终输出文件夹","Please specify temporary directory":"请指定临时目录","Please wait for synchronization":"请等待同步","Please wait for wallet synchronization":"请等待钱包同步","Plot":"农田","Plot Count":"农田数","Plot Id":"农田ID","Plot Key":"农田密钥","Plot NFT":"农田NFT","Plot NFT with p2_singleton_puzzle_hash {plotNFTId} does not exists":["p2_singleton_puzzle_hash为 ",["plotNFTId"]," 的农田NTF不存在。"],"Plot Name":"Plot Name","Plot Public Key":"农田公钥","Plot Seed":"Plot Seed","Plot Size":"农田大小","Plot in Parallel":"并发开垦","Plot is duplicate of {0}":["农田与 ",["0"]," 重复"],"PlotNFT is transitioning to (target state)":"PlotNFT is transitioning to (target state)","PlotNFT is transitioning to (target state). This can take a while. Please do not close the application until this is finalized.":"农田NFT正在过渡到（目标状态）。这可能需要一段时间。请在完成之前不要关闭应用程序。","Plots":"农田","Plots Passed Filter":"通过初筛的农田","Plots are allocated space on your hard drive used to farm and earn Chia. <0>Learn more</0>":"农田就是在你的硬盘上分配的一块空间，用来收获奇亚币。<0>了解更多</0>","Plotting":"正在开垦","Plotting in parallel can save time. Otherwise, add plot(s) to the queue.":"并发开垦可以有效节省开垦时间但对系统资源要求很高。系统硬件不够的话建议选择按队列顺序开垦。","Plotting with bitfield enabled has about 30% less overall writes and is now almost always faster. You may see reduced memory requirements with bitfield plotting disabled. If your CPU design is from before 2010 you may have to disable bitfield plotting.":"在开垦农田时启用位域会减少约30%的磁盘写入，而且大多数情况下会更快。但如果您在开垦时禁用位域，占用的内存则可能会减少。如果您的CPU是2010年以前的型号，您可能必须禁用位域才能开垦。","Plotting with bitfield enabled has about 30% less overall writes and is now almost always faster. You may see reduced memory requirements with bitfield plotting disabled. If your CPU design is from before 2010 you may have to disbale bitfield plotting.":"Plotting with bitfield enabled has about 30% less overall writes and is now almost always faster. You may see reduced memory requirements with bitfield plotting disabled. If your CPU design is from before 2010 you may have to disbale bitfield plotting.","Points Found Since Start":"从开始以来获取的积点","Points Found in Last 24 Hours":"最近24小时获取的积点","Points Successful in Last 24 Hours":"最近24小时成功获取的积点","Pool":"农业合作社","Pool Contract Address":"农业合作社合约地址","Pool Key":"池密钥","Pool Login Link":"农业合作社登录链接","Pool Payout Instructions":"农业合作社支付说明","Pool Public Key":"池公钥","Pool Puzzle Hash":"池谜语哈希","Pool Reward Address":"池奖励地址","Pool Reward Address is not properly formatted.":"池奖励地址格式不正确。","Pool Reward Address is not properly formatted. Using 'paste' is recommended.":"Pool Reward Address is not properly formatted. Using 'paste' is recommended.","Pool Reward Address must not be empty.":"池奖励地址不能为空。","Pool Reward Amount":"池奖励金额","Pool does not provide relative_lock_height.":"农业合作社未提供 relative_lock_height。","Pool does not provide target_puzzle_hash.":"农业合作社未提供 target_puzzle_hash。","Pool public key:":"奖励池公钥：","Pool:":"农业合作社：","Pooling":"合作社耕作","Pooling Wallet":"Pooling Wallet","Port":"端口","Preparing Plot NFT":"正在准备农田NFT","Preparing standard wallet":"正在准备标准钱包","Previous":"上一个","Previous Block Hash":"Previous Block Hash","Previous Header Hash":"上个区块的头部哈希","Previous Sub Block Hash":"Previous Sub Block Hash","Private key with public fingerprint {fingerprint}":["公共指纹为 ",["fingerprint"]," 的私钥"],"Private key {0}":["私钥 ",["0"]],"Private key {fingerprint}":["Private key ",["fingerprint"]],"Private key:":"私钥:","Proof of Space Size":"空间证明的大小","Proofs Found":"已找到证明","Protocol Version":"协议版本","Pubkey":"公钥","Public key:":"公钥：","Puzzle Hash":"Puzzle Hash","Puzzlehash":"谜题哈希","Queue Name":"队列名","Queue name":"队列名","Queued":"排队中","Quit":"Quit","RAM max usage":"内存最大使用量","RL Wallet":"RL Wallet","RLListItems.title":"RLListItems.title","Rate Limited":"费用限定","Rate Limited Info":"费用限制信息","Rate Limited Options":"费用限定型选项","Rate Limited User Wallet Setup":"费用限定型用户钱包设置","Receive Address":"接收地址","Recipient address is not a coloured wallet address. Please enter a coloured wallet address":"Recipient address is not a coloured wallet address. Please enter a coloured wallet address","Recover":"恢复","Recover DID Wallet":"恢复去中心化标识钱包","Recover DID wallet":"Recover DID wallet","Recover Distributed Identity Wallet":"恢复分布式身份钱包","Recover Wallet":"恢复钱包","Recoverable Wallet":"Recoverable Wallet","Recovery Wallet":"Recovery Wallet","Refresh Plots":"刷新农田列表","Relative Lock Height":"相对锁高度","Release Notes":"版本发布说明","Remove keys from old keyring upon successful migration":"Remove keys from old keyring upon successful migration","Removing":"删除中","Rename":"重命名","Rename Wallet":"Rename Wallet","Report an Issue...":"报告问题","Requested":"Requested","Reset":"Reset","Restore Metadata for Coloured Coins and other Smart Wallets from Backup":"从备份中恢复染色币与智能钱包的元信息","Restore Smart Wallets From Backup":"Restore Smart Wallets From Backup","Retire":"Retire","SB Height":"SB Height","Safe To Skip":"可安全跳过此步骤","Save":"保存","Save Offer":"Save Offer","Save Offer File":"Save Offer File","Save passphrase":"Save passphrase","Search":"Search","Search block by header hash":"按头部哈希搜索区块","Second temporary folder location":"第二临时文件夹位置","See private key":"查看私钥：","Seed:":"种子：","Select 2nd Temporary Directory":"选择第二临时文件夹","Select Final Directory":"选择最终输出文件夹","Select Key":"选择密钥","Select Offer":"选择出价","Select Offer File":"Select Offer File","Select Temporary Directory":"选择临时文件夹","Select Wallet":"Select Wallet","Select Wallet Type":"选择钱包类别","Select the final destination for the folder where you would like the plot to be stored. We recommend you use a large slow hard drive (like external HDD).":"选择您希望保存农田文件的最终输出文件夹。我们建议您使用大容量、慢速的硬盘（比如外置机械硬盘）。","Select the temporary destination for the folder where you would like the plot to be stored. We recommend you use a fast SSD.":"Select the temporary destination for the folder where you would like the plot to be stored. We recommend you use a fast SSD.","Select the temporary destination for the folder where you would like the plot to be stored. We recommend you use a fast drive.":"Select the temporary destination for the folder where you would like the plot to be stored. We recommend you use a fast drive.","Select your Plot NFT":"选择你的农田NFT","Select your Plot NFT from the dropdown or create a new one.":"从下拉菜单中选择现有农田NFT，或者新建。","Selected":"已选择","Selected recovery file:":"选择恢复文件：","Self Pooling":"自耕种","Self pool. When you win a block you will earn XCH rewards.":"自耕种。一旦发现新区块，你将得到全部XCH区块奖励。","Sell":"卖出","Send":"发送","Send this info packet to your Rate Limited Wallet user who must use it to complete setup of their wallet:":"把这段信息发给你的费用限定型钱包用户, 以便他/她可以完成自己的钱包设置:","Send your pubkey to your Rate Limited Wallet admin:":"将你的公钥发送到你的费用限定型管理者钱包:","Share":"Share","Share Offer":"Share Offer","Share Privately":"Share Privately","Share on Hashgreen DEX":"Share on Hashgreen DEX","Share on Keybase":"Share on Keybase","Share on OfferBin":"Share on OfferBin","Show Advanced Options":"显示高级选项","Show Asset Id":"Show Asset Id","Show Details":"Show Details","Show Token and Asset Issuance Limitations":"Show Token and Asset Issuance Limitations","Show app":"Show app","Side":"边","Sign In":"登录","Skip":"Skip","Skip Keyring Migration":"Skip Keyring Migration","Skips adding [final dir] to harvester for farming":"Skips adding [final dir] to harvester for farming","Skips adding a final directory to harvester for farming":"勾选此项，将不会自动把最终目录添加到收割机进行耕种","Smooth out your XCH farming rewards by joining a pool.":"加入农业合作社来稳定XCH耕作收益。","Something went wrong":"出错了","Speech":"演说","Spendable Amount":"可用金额","Spendable Amount Per Interval":"每次迭代的支出数目","Spendable Balance":"可用余额","Spending Interval (number of blocks): {interval}":["支出迭代(区块数)",["interval"]],"Spending Interval Length (number of blocks)":"支出迭代长度(区块数)","Spending Limit (chia per interval): {0}":["消费限额（单位时间可消费的奇亚币）：",["0"]],"Standard Wallet":"Standard Wallet","Starting Wallet":"Starting Wallet","Starting service {0}":["Starting service ",["0"]],"State":"状态","Status":"状态","Status:":"状态：","StatusCard.connections":"StatusCard.connections","StatusCard.height":"StatusCard.height","StatusCard.status":"StatusCard.status","StatusCard.title":"StatusCard.title","Stripe Size":"Stripe Size","Sub Block Height":"Sub Block Height","Sub Block at height {0} in the Chia blockchain":["Sub Block at height ",["0"]," in the Chia blockchain"],"Submit":"提交","Success":"Success","Summary":"Summary","Synced":"已同步","Syncing":"正在同步","Syncing <0/>/<1/>":"正在同步 <0/> / <1/>","Syncing {progress}/{tip}":["Syncing ",["progress"],"/",["tip"]],"TXCH Block Rewards":"TXCH Block Rewards","TXCH Farming Rewards":"TXCH Farming Rewards","TXCH Fees Collected":"TXCH Fees Collected","TXCH User Fees":"TXCH User Fees","TXCH User Transaction Fees":"TXCH User Transaction Fees","Target Puzzle Hash":"目标迷题哈希","Temporary files are created during the plotting process which exceeds the size of the final plot files. Make sure you have enough space. <0>Learn more</0>":"Temporary files are created during the plotting process which exceeds the size of the final plot files. Make sure you have enough space. <0>Learn more</0>","Temporary folder location":"临时文件夹位置","The application will stop working at block height 193536.":"这个版本会在区块高度达到193536时停止工作。","The full node that your farmer is connected to is below. <0>Learn more</0>":"你的农场连接的全节点如下.<0>了解更多</0>","The minimum required size for mainnet is k=32":"主网要求的最小农田大小为k=32","The node is not synced":"节点没有同步","The node is syncing, which means it is downloading blocks from other nodes, to reach the latest block in the chain":"节点正在同步中, 也就是说它正在从其他节点下载区块, 以便达到最新的区块高度","The number of Backup IDs needed for recovery cannot exceed the number of Backup IDs added.":"The number of Backup IDs needed for recovery cannot exceed the number of Backup IDs added.","The pool URL \"{normalizedUrl}\" is not working. Is it pool? Error: {0}":["农业合作社地址 ",["normalizedUrl"]," 无效。请确认这是正确的池地址。错误：",["0"]],"The pool URL is not valid. {normalizedUrl}":["此农业合作社地址无效。 ",["normalizedUrl"]],"The pool URL needs to use protocol https. {normalizedUrl}":["农业合作社地址需要使用 https 协议。",["normalizedUrl"]],"The pool URL speciefied is not valid. {normalizedUrl}":["The specified pool URL is not valid. ",["normalizedUrl"]],"The pool URL speciefied is not valid. {poolUrl}":["The pool URL specified is not valid. ",["poolUrl"]],"The pool switching operation was cancelled, please try again by changing pool, or self pooling":"切换农业合作社操作已被取消。请重新尝试切换农业合作社，或自耕种。","The seed used to create the plot. This depends on the pool pk and plot pk.":"密语用于创建农田。农田是与农场池的公钥与农田的公钥相关联的。","The total number of VDF (verifiable delay function) or proof of time iterations on the whole chain up to this sub block.":"截至该子区块为止全链已完成的VDF(延迟验证函数) 或时间证明的迭代次数.","The total number of VDF (verifiable delay function) or proof of time iterations on this block.":"本区块上花费的总VDF(延迟验证函数) 或时间验证的迭代次数.","The total transactions fees in this block. Rewarded to the farmer.":"本区块上奖励给农民的全部交易费用。","These are the instructions for how the farmer wants to get paid. By default this will be an XCH address, but it can be set to any string with a size of less than 1024 characters, so it can represent another blockchain or payment system identifier.":"这是农民耕种奖励的支付方式。默认情况下，这将是一个XCH钱包地址，但是你可以设置任意长度小于1024的字符串，也就是说，你可以指定其它区块链或者支付系统标识。","These plots are invalid, you might want to delete them.":"这些农田无效，你可能想要删除它们。","This access token is verified":"This access token is verified","This access token is whitelisted":"This access token is whitelisted","This allows you to add a directory that has plots in it. If you have not created any plots, go to the plotting screen.":"在这里可以把事先准备好的农田文件添加进来。如果还没有的话，就去开垦几块农田吧。","This difficulty is an artifically lower difficulty than on the real network, and is used when farming, in order to find more proofs and send them to the pool. The more plots you have, the higher difficulty you will have. However, the difficulty does not affect rewards.":"此难度是比实际网络难度人为降低的难度，它将被用于耕作中，用来发现尽量多的验证数据并发送给农业合作社。你的农田越多，你的难度就会越高。这个难度不会影响你的耕作奖励。","This feature is available only from the GUI.":"此功能只在图形界面中提供。","This is the amount of Chia that you can currently use to make transactions. It does not include pending farming rewards, pending incoming transactions, and Chia that you have just spent but is not yet in the blockchain.":"这是你目前可以交易的奇亚币数额。该数额不包含处理中的农场奖励、收入交易，以及未经区块链确认的支付交易。","This is the pending change, which are change coins which you have sent to yourself, but have not been confirmed yet.":"处理中的找零钱, 是你发送支出的找回, 但是没有被链上确认.","This is the sum of the incoming and outgoing pending transactions (not yet included into the blockchain). This does not include farming rewards.":"待处理的收入与支出总和(还未上链). 不包括农场耕种奖励.","This is the time of the latest peak sub block.":"最新的子块时间.","This is the time the block was created by the farmer, which is before it is finalized with a proof of time":"此时间是该区块被农民创建的时间，因此，此时间会早于其被时间认证机制最终确认的时间。","This is the total amount of chia in the blockchain at the current peak sub block that is controlled by your private keys. It includes frozen farming rewards, but not pending incoming and outgoing transactions.":"这里是你的私钥所控制的, 到目前最新子块高度为止的全部奇亚数. 包含被冻结的耕种奖励, 但不包含待处理的收入与支出.","This is the total balance + pending balance: it is what your balance will be after all pending transactions are confirmed.":"目前结余+待处理的结余: 在链上确认完成后就是你的最终总结余.","This is the total number of points this plotNFT has with this pool, since the last payout. The pool will reset the points after making a payout.":"这是此农田NFT自从上次支付奖励后在当前农业合作社中持有的总积点数。在农业合作社支付奖励后，此积点数会归零。","This is the total number of points your farmer has found for this plot NFT. Each k32 plot will get around 10 points per day, so if you have 10TiB, should should expect around 1000 points per day, or 41 points per hour.":"这是你的农民在这个农田NFT对应的农田中找到的总积点数。一个k32的农田每天大约会得到10个积点。也就是说，如果你拥有10TiB的农田，你每天大约会获得1000积点，也就是每小时41积点。","This node is fully caught up and validating the network":"本节点已完成同步，并已开始参与网络验证","This number reflects the number of nodes that the accepted SpendBundle has been sent to":"This number reflects the number of nodes that the accepted SpendBundle has been sent to","This offer has completed successfully":"This offer has completed successfully","This offer is no longer valid":"This offer is no longer valid","This plot NFT is assigned to a different key. You can still create plots for this plot NFT, but you can not make changes.":"此农田NFT绑定了其它密钥。你仍然可以用此农田NFT开垦农田，但是你无法进行任何更改。","This plot NFT is not connected to pool":"此农田NFT尚未连接到农业合作社。","This table shows you the last time your farm attempted to win a block challenge.":"This table shows you the last time your farm attempted to win a block challenge.","This table shows you the last time your farm attempted to win a block challenge. <0>Learn more</0>":"这个表格里面显示的是你的农场尝试过的区块挑战. <0>了解更多</0>","This trade was created at this time":"本交易创建于","This trade was included on blockchain at this block height":"该交易在链上的区块高度","This version of Chia is no longer compatible with the blockchain and can not safely farm.":"此版本的奇亚客户端与区块链已经不再兼容，无法安全地种植。","Time Created":"创建时间","Time this trade was created at this time":"Time this trade was created at this time","Timestamp":"时间戳","To":"目标地址","Token and Asset Issuance Limitations":"Token and Asset Issuance Limitations","Token has empty name":"Token has empty name","Token has empty tail":"Token has empty tail","Total Balance":"总余额","Total Chia Farmed":"Total Chia Farmed","Total Iterations":"总迭代次数","Total Network Space":"全网占用空间","Total Plot Size:":"总农田大小：","Total Size of Plots":"总农田大小：","Total VDF Iterations":"总VDF迭代数","Total iterations since the start of the blockchain":"自区块开始以来的迭代数","Trade":"Trade","Trade Details":"交易详情","Trade ID":"交易ID","Trade ID:":"交易ID：","Trade Overview":"交易概况","Trades will show up here":"交易会在这里显示","Trading":"交易","Trading History":"交易历史","Transaction has not been sent to node yet":"Transaction has not been sent to node yet","Transaction has sent to a full node and is pending inclusion into the mempool. {0}":["Transaction has sent to a full node and is pending inclusion into the mempool. ",["0"]],"Transaction has successfully been sent to a full node and included in the mempool.":"Transaction has successfully been sent to a full node and included in the mempool.","Transactions":"Transactions","Transactions Filter Hash":"交易过滤器哈希","Try Again":"Try Again","Type":"类别","Unable to create plot NFT":"无法创建农田NFT","Unable to load the list of the keys":"Unable to load the list of the keys","Unable to open Keybase. Install Keybase from https://keybase.io":"Unable to open Keybase. Install Keybase from https://keybase.io","Unable to open browser. Install Keybase from https://keybase.io":"Unable to open browser. Install Keybase from https://keybase.io","Unclaimed Rewards":"未申领奖励","Unfinished":"未完成","Unique identifier":"唯一识别码","Unknown":"未知","Unknown CAT":"Unknown CAT","Unlock Keyring":"Unlock Keyring","Unsaved Changes":"未保存的更改","User Pubkey":"用户公钥","VDF Sub Slot Iterations":"VDF子项迭代数","Value is too high":"Value is too high","Value seems high":"值似乎过高","Verify Pool Details":"验证农业合作社详情","View":"查看","View Log":"查看日志","View Offer":"查看出价","View Payout Instructions":"View Payout Instructions","View Pool Login Link":"查看农业合作社登录地址","View an Offer":"View an Offer","View on Hashgreen DEX":"View on Hashgreen DEX","View on OfferBin":"View on OfferBin","View pending balances":"查看处理中的余额","View pending balances...":"查看处理中的余额…","Viewing offer":"Viewing offer","Viewing offer created at {0}":["Viewing offer created at ",["0"]],"Viewing offer: {offerFilePath}":["Viewing offer: ",["offerFilePath"]],"Wait for synchronization":"等待同步","Waiting for launch":"Waiting for launch","Waiting for network to launch":"Waiting for network to launch","Waiting for the transaction to be confirmed":"正在等待交易确认","Wallet Status:":"钱包状态：","Wallet does not exists":"钱包不存在","Wallet id is not defined":"Wallet id is not defined","WalletItem.ccWallet":"WalletItem.ccWallet","WalletItem.chiaWallet":"WalletItem.chiaWallet","WalletItem.didWallet":"WalletItem.didWallet","WalletItem.rlWallet":"WalletItem.rlWallet","WalletStatusCard.not_synced":"WalletStatusCard.not_synced","WalletStatusCard.synced":"WalletStatusCard.synced","WalletStatusCard.syncing":"WalletStatusCard.syncing","Wallets":"钱包","Wallets.title":"Wallets.title","Want to Join a Pool? Create a Plot NFT":"想要加入农业合作社？创建一个农田NFT","Want to earn more Chia? Add more plots to your farm.":"想要获得更多奇亚币？ 给你的农场多开垦一些农田吧。","Want to have a delay before the next plot starts?":"想要在开垦下一块农田前等待一会？","Warning":"Warning","Warning: This key is used for a wallet that may have a non-zero balance. By deleting this key you may lose access to this wallet":"警告：此密钥用于可能有余额的钱包。删除此密钥可能会无法访问此钱包。","Warning: This key is used for your farming rewards address. By deleting this key you may lose access to any future farming rewards":"警告：此密钥用于耕作奖励地址。删除此密钥后可能会无法提取将来的耕作奖励。","Warning: This key is used for your pool rewards address. By deleting this key you may lose access to any future pool rewards":"警告：此密钥用于池奖励地址。删除此密钥可能会无法获得将来的池奖励。","Warning: Verify that the offered CAT asset IDs match the asset IDs of the tokens you expect to receive.":"Warning: Verify that the offered CAT asset IDs match the asset IDs of the tokens you expect to receive.","Weight":"权重","Weight is the total added difficulty of all sub blocks up to and including this one":"权重是施加给所有子块包括目前这块的额外的困难度.","Welcome to Chia. Please log in with an existing key, or create a new key.":"欢迎来到奇亚。请用已有的密钥登录，或者创建一个新的密钥。","Welcome! The following words are used for your wallet backup. Without them, you will lose access to your wallet, keep them safe! Write down each word along with the order number next to them. (Order is important)":"欢迎使用奇亚! 下面的词语列表是你的钱包密语, 请务必严格完整的把它们记下来(保持顺序). 没有这段密语的话, 你就没办法找回钱包里的币. 千万也不要透露给别人.","When you receive the setup info packet from your admin, enter it below to complete your Rate Limited Wallet setup:":"当你从管理员那收到这个设置信息后, 请把它输入到下面以便完成费用限定型钱包设置:","Window":"窗口","Without fees":"不含交易费","Yes":"是","You are in pending state. Please wait for confirmation":"你当前处于待处理状态，请等待区块链确认。","You are not self pooling":"你目前没有自耕种","You can still create plots for this plot NFT, but you can not make changes until sync is complete.":"你现在可以使用这个农田NFT开垦新农田，但是在同步完成前，你不能进行任何更改。","You created this offer":"You created this offer","You do not need to be synced or connected to plot. Temporary files are created during the plotting process which exceed the size of the final plot files. Make sure you have enough space. <0>Learn more</0>":"You do not need to be synced or connected to plot. Temporary files are created during the plotting process which exceed the size of the final plot files. Make sure you have enough space. <0>Learn more</0>","You do not need to be synched or connected to Plot. Temporary files are created during the plotting process which exceed the size of the final plot files. Make sure you have enough space. <0>Learn more</0>":"You do not need to be synched or connected to Plot. Temporary files are created during the plotting process which exceed the size of the final plot files. Make sure you have enough space. <0>Learn more</0>","You do not need to be synched or connected to plot. Temporary files are created during the plotting process which exceed the size of the final plot files. Make sure you have enough space. <0>Learn more</0>":"不需要同步或者连接到网络节点就可以开垦农田。在开垦过程中产生的临时文件尺寸会大于最终的农田大小。请确认你有足够的存储空间。<0>了解更多</0>","You have made changes. Do you want to discard them?":"你已经做了一些更改。你确定要放弃这些修改吗？","You have {0}% of the space on the network, so farming a block will take {expectedTimeToWin} in expectation.":["You have ",["0"],"% of the space on the network, so farming a block will take ",["expectedTimeToWin"]," in expectation."],"You have {0}% of the space on the network, so farming a block will take {expectedTimeToWin} in expectation. Actual results may take 3 to 4 times longer than this estimate.":["您在网络上拥有",["0"],"%的空间，因此收割一个区块所花费的时间在数学期望上是",["expectedTimeToWin"],"。实际耗时可能比这个期望值长3到4倍。"],"You have {0}% of the space on the network, so farming a block will take {expectedTimeToWin} in expectation. Actual results may take 3-4 times longer - actual results may be 3x this estimate.<0>Learn about Expected Value</0>":["You have ",["0"],"% of the space on the network, so farming a block will take ",["expectedTimeToWin"]," in expectation. Actual results may take 3-4 times longer - actual results may be 3x this estimate.<0>Learn about Expected Value</0>"],"You have {0}% of the space on the network, so farming a block will take {expectedTimeToWin} in expectation. Actual results may take 3-4 times longer this estimate. <0>Learn about Expected Value</0>":["You have ",["0"],"% of the space on the network, so farming a block will take ",["expectedTimeToWin"]," in expectation. Actual results may take 3-4 times longer this estimate. <0>Learn about Expected Value</0>"],"You need to claim your rewards first":"你需要先领取你的奖励。","You need to upgrade the Chia application as this version will stop working soon!":"You need to upgrade the Chia application as this version will stop working soon!","You need {currencyCode} to join a pool.":["必须持有 ",["currencyCode"]," 才有加入农业合作社。"],"You will offer":"You will offer","You will receive":"You will receive","You will recieve <0/> to {0}":["你将会收到 <0/> 至 ",["0"]],"You will recieve <0/> to {address}":["你将会收到 <0/> 至 ",["address"]],"Your DID requires at least {dids_num_req} attestation file{0} for recovery. Please upload additional files.":["Your DID requires at least ",["dids_num_req"]," attestation file",["0"]," for recovery. Please upload additional files."],"Your Farm Overview":"农场概览","Your Full Node Connection":"你的全节点连接","Your Harvester Network":"你的收割网络","Your Offers":"Your Offers","Your Pool Overview":"你的农业合作社概览","Your keyring is locked":"Your keyring is locked","Your keys have not been migrated to a new keyring. You will be unable to create new keys or delete existing keys until migration completes. Are you sure you want to skip migrating your keys?":"Your keys have not been migrated to a new keyring. You will be unable to create new keys or delete existing keys until migration completes. Are you sure you want to skip migrating your keys?","Your keys need to be migrated to a new keyring that is optionally secured by a master passphrase.":"Your keys need to be migrated to a new keyring that is optionally secured by a master passphrase.","Your keys need to be migrated to a new keyring that is secured by a master passphrase.":"Your keys need to be migrated to a new keyring that is secured by a master passphrase.","Your offer has been successfully posted to Keybase.":"Your offer has been successfully posted to Keybase.","Your offer:":"Your offer:","Your passphrase can be stored in your system's secure credential store. Chia will be able to access your keys without prompting for your passphrase.":"Your passphrase can be stored in your system's secure credential store. Chia will be able to access your keys without prompting for your passphrase.","[Error 13] Permission denied. You are trying to access a file/directory without having the necessary permissions. Most likely one of the plot folders in your config.yaml has an issue.":"[错误 13] 权限被拒绝。您正试图访问的文件或目录没有必需的权限。农田文件夹中的 config.yaml 文件可能存在问题。","[Error 22] File not found. Most likely one of the plot folders in your config.yaml has an issue.":"[错误 22] 文件未找到。农田文件夹中的 config.yaml 文件可能存在问题。","connections:":"连接：","height:":"高度：","not synced":"未同步","p2_singleton_puzzle_hash is not defined":"p2_singleton_puzzle_hash 未定义","status:":"状态：","synced":"已同步","syncing":"同步中","{0}":[["0"]],"{0} Block Rewards":[["0"]," Block Rewards"],"{0} User Transaction Fees":[["0"]," User Transaction Fees"],"{0} points {1} - {2} hours ago":[["0"]," 积点 ",["1"]," - ",["2"]," 小时前"],"{currencyCode} Block Rewards":[["currencyCode"]," 区块奖励"],"{currencyCode} Total Chia Farmed":[["currencyCode"]," 收获的奇亚币总计"],"{currencyCode} User Transaction Fees":[["currencyCode"]," 用户交易费用"],"{mojo, plural, one {mojo} other {mojos}}":[["mojo","plural",{one:"mojo",other:"mojos"}]],"{mojos, plural, one {mojo} other {mojos}}":[["mojos","plural",{one:"mojo",other:"mojos"}]]}};
