@@ -1,1 +1,1 @@
-/*eslint-disable*/module.exports={messages:{"(Optional)":"(選用)","*Want to explore Chia’s blocks further? Check out <0>Chia Explorer</0> built by an open source developer.":"＊想更進一步探索 Chia 區塊鍵？看看開源開發者打造的 <0>Chia Explorer</0>","+":"+","+ Add New Plot NFT":"+ 新增新的農場","+ Add Token":"+ Add Token","+ Add Wallet":"+ 新增錢包","-":"-","0 automatically chooses bucket count":"0 automatically chooses bucket count","128 buckets is recommended":"建議使用 128 個桶","A harvester is a service running on a machine where plot(s) are actually stored. A farmer and harvester talk to a full node to see the state of the chain. View your network of connected harvesters below Learn more":"收割機是在實際存放耕地的機器上運作的服務。塊農和收割機會與一個完整節點溝通以取得區塊鍵狀態。以下是你的收割機與塊農網路 學習更多","About Chia Blockchain":"關於 Chia 區塊鏈","About Chia Wallet":"About Chia Wallet","Accept":"接受","Accepted at time:":"接受於：","Action":"動作","Actions":"動作","Add":"新增","Add Backup ID":"增加備份 ID","Add Custom Token":"Add Custom Token","Add Plot Directory":"新增耕地資料夾","Add Plot to Queue":"新增耕地至隊列","Add Token":"Add Token","Add Wallet":"Add Wallet","Add a Plot":"新增耕地","Add a Plot NFT":"新增一塊農場","Add a plot":"新增耕地","Add plot directory":"新增耕地資料夾","Add {currencyCode} from the Faucet":["從水龍頭新增",["currencyCode"],"."],"Adding {0} token":["Adding ",["0"]," token"],"Address":"位址","Address / Puzzle hash":"位址 / 拼圖雜湊值","After your keys are successfully migrated to the new keyring, you may choose to have your keys removed from the old keyring.":"After your keys are successfully migrated to the new keyring, you may choose to have your keys removed from the old keyring.","All":"All","Amount":"數量","Amount ({cc_unit})":["數量 (",["cc_unit"],")"],"Amount For Initial Coin":"初始代幣數量","Amount must be an even amount.":"Amount must be an even amount.","Are you sure you want to delete the plot? The plot cannot be recovered.":"你確要刪除這個耕地嗎？此耕地將無法尋回。","Are you sure you want to delete unconfirmed transactions?":"你確認要刪除未確認的交易?","Are you sure you want to disconnect?":"您確定要斷線嗎？","Are you sure you want to quit?":"Are you sure you want to quit?","Are you sure you want to quit? GUI Plotting and farming will stop.":"你確定要退出？耕地規劃和耕種將會停止。","Are you sure you want to use k={plotSize}":["Are you sure you want to use k=",["plotSize"]],"Are you sure you want to use k={plotSize}?":["你確定 k 的大小為 ",["plotSize"],"?"],"Asset Id":"Asset Id","Atomic Swap Wallet":"Atomic Swap Wallet","Attempt {attempt}":["Attempt ",["attempt"]],"Authorized Payee Wallet":"Authorized Payee Wallet","Autogenerated name from pool contract address":"自動從農會合約地址建立名稱","Back":"返回","Backup file is used to restore smart wallets.":"備份檔案可用於復原智慧錢包。","Balance":"餘額","Base Farmer Reward Amount":"基礎塊農獎勵數量","Below are the current block challenges. You may or may not have a proof of space for these challenges. These blocks do not currently contain a proof of time.":"以下是目前的區塊挑戰。你可能會或可能不會有這些挑戰的空間證明。這些區塊目前不包含時間證明。","Best estimate over last 24 hours":"已過 24 小時最佳預估值","Best estimate over last 5 hours":"Best estimate over last 5 hours","Block":"區塊","Block Test":"區塊測試","Block VDF Iterations":"區塊 VDF 疊代次數","Block at height {0} in the Chia blockchain":["Chia 區塊鏈區塊高度 ",["0"]],"Block hash":"Block hash","Block with hash {headerHash}":["帶雜湊值 ",["headerHash"]," 的區塊"],"Block with hash {headerHash} does not exist.":["帶雜湊值 ",["headerHash"]," 的區塊不存在。"],"Blocks":"區塊","Both values are empty. This is ok if you are, for example, sending your rewards to an offline wallet.":"Both values are empty. This is ok if you are, for example, sending your rewards to an offline wallet.","Browse":"瀏覽","But you are currently farming <0/>":"但你目前耕種 <0/>","Buy":"購買","CAT Wallet":"CAT Wallet","CC Wallet":"CC Wallet","CCListItems.title":"CCListItems.title","Can be backed up to mnemonic seed":"可以備份到助記符號種子","Cancel":"取消","Cancel and Spend":"取消和花費","Cannot send chia to coloured address. Please enter a chia address.":"Cannot send chia to coloured address. Please enter a chia address.","Caution, deleting these plots will delete them forever. Check that the storage devices are properly connected.":"警告，刪除這些農田區將會是永久刪除的。檢查儲存裝置是否正確連接。","Challenge":"挑戰","Challenge Hash":"挑戰雜湊值","Change":"變更","Change Pool":"變更農會","Chat on KeyBase":"在 KeyBase 上閒聊","Chia":"Chia","Chia Asset Token":"Chia Asset Token","Chia Blockchain Wiki":"Chia 區塊鏈維基","Chia Wallet":"Chia 錢包","Choose Number of Plots":"選擇耕地數量","Choose Plot Size":"選擇耕地大小","Claim Rewards":"領取獎勵","Close":"關閉","Closing down node and server":"關閉節點和伺服器","Coin Name":"幣種名稱","Coins:":"幣種：","Colour":"顏色","Colour Info":"顏色資訊","Colour String":"顏色字串","Colour:":"顏色：","Coloured Coin":"有色代幣","Coloured Coin Options":"有色代幣選項","Confirm":"確認","Confirm Disconnect":"確認斷線","Confirm Passphrase":"Confirm Passphrase","Confirmation":"確認","Confirmed":"Confirmed","Confirmed at Height":"Confirmed at Height","Confirmed at block:":"已確認區塊：","Confirmed at height {0}":["已確認於高度 ",["0"]],"Connect":"連線","Connect to other peers":"連線到其他同儕","Connect to pool":"連線至農會","Connected":"已連線","Connecting do daemon":"Connecting do daemon","Connecting to daemon":"Connecting to daemon","Connecting to wallet":"連線至錢包中","Connection Status":"連線狀態","Connection Status:":"連線狀態：","Connection type":"連線類型","Connections":"連線","Contribute on GitHub":"在 GitHub 上貢獻","Copied":"已複製","Copy":"複製","Copy to Clipboard":"複製到剪貼板","Create":"建立","Create An Attestation Packet":"建立一個證明封包","Create Chia Asset Token Wallet":"Create Chia Asset Token Wallet","Create Chia Asset Token Wallet from Existing TAIL":"Create Chia Asset Token Wallet from Existing TAIL","Create DID wallet":"Create DID wallet","Create Distributed Identity Wallet":"建立分散式識別的錢包","Create New Chia Asset Token Wallet":"Create New Chia Asset Token Wallet","Create New Wallet":"建立新錢包","Create Offer":"建立提案","Create Plot":"Create Plot","Create Rate Limited Admin Wallet":"建立速率限制管理員錢包","Create Rate Limited User Wallet":"建立速率限制使用者錢包","Create Token":"Create Token","Create Trade":"Create Trade","Create Trade Offer":"建立貿易提案","Create Transaction":"建立交易","Create a Backup":"建立備份","Create a Plot NFT":"建立一塊農場","Create a new private key":"建立一個新的私鑰","Create admin wallet":"建立管理員錢包","Create custom CAT Wallet":"Create custom CAT Wallet","Create from Existing":"Create from Existing","Create new coloured coin":"建立新的有色代幣","Create user wallet":"建立使用者錢包","Create wallet for colour":"建立有色代幣錢包","Create wallet for existing colour":"用已有的有色代幣去建立錢包","CreateWallet.addWallet":"CreateWallet.addWallet","Created At:":"建立於：","Created by us:":"由我們創造：","Creating Plot NFT and Joining the Pool":"建立一塊農場並加入農會","Creating Plot NFT for Self Pooling":"正在為自己的農會建立農場","Currency code is not defined":"未定義貨幣代碼","Current Difficulty":"目前難度","Current Points Balance":"目前積分餘額","Current trade status":"目前交易狀態","Custody Wallet":"Custody Wallet","Custom":"Custom","DANGER: permanently delete private key":"危險：將永久刪除私鑰","DID Wallet":"DID Wallet","DIDListItems.title":"DIDListItems.title","Date":"日期","Delay":"延遲","Delete":"刪除","Delete Key":"Delete Key","Delete Plot":"刪除耕地","Delete Unconfirmed Transactions":"刪除未確認的交易","Delete all keys":"刪除全部金鑰","Delete key":"Delete key","Delete key {fingerprint}":["刪除帳號指紋：",["fingerprint"]],"Deleting all keys will permanently remove the keys from your computer, make sure you have backups. Are you sure you want to continue?":"刪除全部金鑰將會永久從你的電腦中移除，請確保你有備份。你確定要繼續？","Deleting key":"Deleting key","Deleting the key will permanently remove the key from your computer, make sure you have backups. Are you sure you want to continue?":"刪除此金鑰將會永久從你的電腦中移除，請確保你有備份。你確定要繼續？","Developer":"開發者","Developer Tools":"開發者工具","Difficulty":"難度","Disable bifield plotting when your temporary directory is on fast storage like SSD or you need lower per process RAM requirements. Plotting with bitfield enabled has about 20% less overall writes.":"Disable bifield plotting when your temporary directory is on fast storage like SSD or you need lower per process RAM requirements. Plotting with bitfield enabled has about 20% less overall writes.","Disable bitfield plotting":"禁用 bitfield 耕地","Discard":"放棄","Disconnect":"斷線","Distributed Identity":"分散式識別","Distributed Identity Options":"Distributed Identity Options","Do you have existing plots on this machine? <0>Add Plot Directory</0>":"你有已存在的耕地在本機上？ <0>增加耕地資料夾</0>","Does your machine support parallel plotting?":"你的機器支援平行耕地嗎？","Drag and drop attestation packet(s)":"拖放證明封包","Drag and drop offer file":"拖放提案檔案","Drag and drop your backup file":"拖放你的備份檔案","Drag and drop your recovery backup file":"拖放您的備份檔案","Edit":"編輯","Edit Payout Instructions":"編輯支付說明","Enter a strong passphrase and click Migrate Keys to secure your keys":"Enter a strong passphrase and click Migrate Keys to secure your keys","Enter the 24 word mnemonic that you have saved in order to restore your Chia wallet.":"輸入你已存的 24 字助記符號以復原您的 Chia 錢包。","Error":"錯誤","Error the entered address appears to be for a different colour.":"Error the entered address appears to be for a different colour.","Error: Cannot send chia to coloured address. Please enter a chia address.":"錯誤：不能發送 Chia 到有色地址。請輸入 Chia 地址。","Estimated Network Space":"估計的網路空間","Estimated Time to Win":"預計贏取時間","Estimated network space":"Estimated network space","Estimated sum of all the plotted disk space of all farmers in the network":"網路中所有農夫的所有耕地磁盤空間估計總和","Exclude final directory":"排除最終資料夾","Expected Time to Win":"Expected Time to Win","Failed to open (invalid plots)":"開啟失敗（無效耕地）","Farm":"耕種","Farmer Public Key":"農夫公鑰","Farmer Puzzle Hash":"塊農拼圖雜湊值","Farmer Reward Address":"塊農獎勵地址","Farmer Reward Address is not properly formatted.":"農夫獎勵地址並不符合格式。","Farmer Reward Address is not properly formatted. Using 'paste' is recommended.":"Farmer Reward Address is not properly formatted. Using 'paste' is recommended.","Farmer Reward Address must not be empty.":"農夫獎勵地址不能為空。","Farmer is not connected":"塊農目前未連線","Farmer is not running":"塊農未運作","Farmer public key:":"農夫公鑰:","Farmers earn block rewards and transaction fees by committing spare space to the network to help secure transactions. This is where your farm will be once you add a plot. <0>Learn more</0>":"農夫可通過提交閒置空間到網路上來幫助保護交易，從而獲得區塊獎勵和交易費。這便是新增耕地後耕種的地方。<0>了解詳情</0>","Farming":"耕種中","Farming Status":"耕種狀態","Fee":"費用","Fee (TXCH)":"Fee (TXCH)","Fee ({0})":["Fee (",["0"],")"],"Fee ({currencyCode})":["費用 (",["currencyCode"],")"],"Fees Amount":"費用數量","Fees Puzzle Hash":"Fees Puzzle Hash","File":"檔案","Filename":"檔案名稱","Final folder location":"最終資料夾位置","Finished":"已完成","Follow on Twitter":"Twitter 上追隨","Frequently Asked Questions":"常見問題","Full Node":"完整節點","Full Node Overview":"完整節點概觀","Full Node Status":"Full Node Status","Full Screen":"全畫面","Generate New Colour":"生成新顏色","Get Pool Login Link":"Get Pool Login Link","Grid view":"網格檢視","HD or Hierarchical Deterministic keys are a type of public key/private key scheme where one private key can have a nearly infinite number of different public keys (and therefor wallet receive addresses) that will all ultimately come back to and be spendable by a single private key.":"層及式確定性金鑰是一種公鑰/私鑰格式，一個私鑰可以具有幾乎無限數量的不同公鑰（以此類推，同錢包接收地址），最終所有這些公鑰都將回朔於一個私鑰使用／花用。","Harvester ID":"Harvester ID","Harvester Plots":"耕地收割機","Header Hash":"表頭雜湊值","Header hash":"表頭雜湊值","Height":"高度","Help":"幫助","Help translate":"協助翻譯","Hide Advanced Options":"隱藏進階選項","Hint":"Hint","History":"歷史","Host Name":"主機名稱","IP address":"IP位址","IP address / host":"IP 位址 / 主機","If none selected, then it will default to the temporary directory.":"如無選擇，將會預設至臨時資料夾。","Import Wallet from Mnemonics":"以助記符號匯入錢包","Import from Mnemonics (24 words)":"從助記詞(24 字) 匯入","In Progress":"進行中","Incoming":"轉入","Incorrect value":"數值不正確","Index":"索引","Indicated if this offer was created by us":"指出如果此提案由我方創建","Indicated what time this offer was accepted":"指出此提案於何時被接受","Info Packet":"資訊封包","Initial Amount":"初始數量","Initialize a Rate Limited User Wallet:":"初始化速率限制使用者錢包：","Interval":"間隔","Invalid state":"無效狀態","It is a one-time login link that can be used to log in to a pool's website. It contains a signature using the farmer's key from the plot NFT. Not all pools support this feature.":"這是一個一次性登錄連結, 可用於登記在農會的網站. 它來自農場, 其包含農民公鑰的簽名. 並非所有農會都支援此功能.","Join Pool":"加入礦池","Join a Pool":"加入一個農場","Join a pool and get consistent XCH farming rewards. The average returns are the same, but it is much less volatile.":"加入農會並獲得持續的 XCH 耕種獎勵。雖然平均報酬一樣但波動較小相對穩定。","Join a pool and get consistent XCH farming rewards. The average returns are the same, but it is much less volatile. Assign plots to a plot NFT. You can easily switch pools without having to re-plot.":"加入農會並獲得持續的XCH 耕種獎勵。雖然平均報酬一樣但波動較小相對穩定。分配耕地到農場。你可以更輕鬆的轉換農會而不用重新進行耕地。","Join a pool and get more consistent XCH farming rewards. Create a plot NFT and assign your new plots to a group.":"加入農會以獲得更多持續性的 XCH 耕種獎勵。建立農場並分配您的新耕地到群組。","K-Size":"K-大小","Keyring migration failed: {0}":["Keyring migration failed: ",["0"]],"Keys":"金鑰","KiB Up/Down":"KiB 上傳/下傳","Last Attempted Proof":"最後試過證明","Last Height Farmed":"最後耕種高度","Latest Block Challenges":"最新區塊挑戰","Launcher Id":"啟動器 ID","Learn More":"了解更多","Leaving Pool":"離開農會","List of connections is empty":"List of connections is empty","List view":"清單檢視","Loading Plot NFTs":"正在讀取農場","Loading configuration":"Loading configuration","Loading keyring status":"Loading keyring status","Loading list of the keys":"Loading list of the keys","Loading list of wallets":"正在讀取錢包列表","Loading...":"載入中...","Local Harvester Plots":"Local Harvester Plots","Log Out":"Log Out","Logging in":"登入中","MainWalletList.colouredCoin":"MainWalletList.colouredCoin","MainWalletList.createAdminWallet":"MainWalletList.createAdminWallet","MainWalletList.createDIDWallet":"MainWalletList.createDIDWallet","MainWalletList.createNewColouredCoin":"MainWalletList.createNewColouredCoin","MainWalletList.createUserWallet":"MainWalletList.createUserWallet","MainWalletList.createWalletForExistingColour":"MainWalletList.createWalletForExistingColour","MainWalletList.distributedIdentity":"MainWalletList.distributedIdentity","MainWalletList.rateLimited":"MainWalletList.rateLimited","MainWalletList.recoverDIDWallet":"MainWalletList.recoverDIDWallet","MainWalletList.title":"MainWalletList.title","Manage Farming Rewards":"管理耕種獎勵","Manage Recovery DIDs":"管理還原的 DID","Manage Your Farming Rewards Target Addresses":"管理你的耕種獎勵目標地址","Memo":"Memo","Memos":"Memos","MiB Up/Down":"MiB 上傳/下傳","Migrate Keys":"Migrate Keys","Migration required":"Migration required","Minimum Difficulty":"最小難度","Minutes":"分鐘","More memory slightly increases speed":"更多記憶體可略微提高速度","Multi Sig Wallet":"Multi Sig Wallet","My DID Wallet":"我的 DID 錢包","My Pubkey":"我的公鑰","Name":"Name","Network Name":"網路名稱","Network has not yet been launched, once it launches full node will start automatically.":"Network has not yet been launched, once it launches full node will start automatically.","Network not launched yet, waiting for genesis challenge!":"Network not launched yet, waiting for genesis challenge!","New Address":"新位址","New Wallet":"新錢包","Next":"下一步","Nickname":"暱稱","No":"否","No 24 word seed, since this key is imported.":"由這個鑰匙是匯入的，沒有 24 個字種子。","No blocks farmed yet":"還沒有耕種收穫到的區塊","No previous transactions":"沒有上一個交易","No private keys for one or both addresses. Safe only if you are sending rewards to another wallet.":"一邊或雙邊地址沒有私鑰。只有你發送獎勵到另一個錢包時才算安全。","Node ID":"節點 ID","Node Id":"節點 ID","None":"無","None of your plots have passed the plot filter yet.":"你還沒有任何的耕地通過篩選。","Not Available":"尚無","Not Synced":"未同步","Not accepted yet":"還未接受","Not confirmed yet":"還未確認","Not connected":"未連線","Not found Plots":"沒找到的耕地","Note that this does not change your pooling payout addresses. This only affects old format plots, and the 0.25XCH reward for pooling plots.":"請注意，這不會變更您的農會支付位址。這只會影響舊的耕地格式和 0.25XCH 的農會耕地獎勵。","Number of Plots":"耕地數量","Number of buckets":"桶數量","Number of threads":"執行緒數：","OK":"好","Offer":"提案","Offers Created":"已建立提案","On average there is one minute between each transaction block. Unless there is congestion you can expect your transaction to be included in less than a minute.":"平均每個交易區塊之間有 1 分鐘的時間。除非出現擁塞，否則你可以在不到 1 分鐘的時間內完成交易。","Only one backup file is allowed.":"只允許一個備份檔案。","Only one backup file is allowed. To add a new backup file, please delete any previously uploaded files and try again.":"Only one backup file is allowed. To add a new backup file, please delete any previously uploaded files and try again.","Outgoing":"匯出","Passphrase":"Passphrase","Passphrase Hint":"Passphrase Hint","Passphrase Hint (Optional)":"Passphrase Hint (Optional)","Payout Address":"付款地址","Peak Height":"峰值高度","Peak Sub-block Height":"Peak Sub-block Height","Peak Time":"峰值時間","Pending":"待確認中","Pending Balance":"未結餘額","Pending Change":"未結更動","Pending Total Balance":"未結總餘額","Please Confirm":"請確認","Please add a trade pair":"請增加一對貿易方","Please enter 0 fee. Positive fees not supported yet for RL.":"請輸入 0 費用。 RL 尚未支援正向費用。","Please enter a coin":"請輸入幣種","Please enter a filename":"請輸入檔案名稱","Please enter a passphrase":"Please enter a passphrase","Please enter a pubkey":"請輸入公鑰","Please enter a puzzlehash":"請輸入一個 PuzzleHash","Please enter a valid CAT name":"Please enter a valid CAT name","Please enter a valid TAIL":"Please enter a valid TAIL","Please enter a valid asset id":"Please enter a valid asset id","Please enter a valid initial coin amount":"請輸入有效的初始金額","Please enter a valid integer of 0 or greater for the number of Backup IDs needed for recovery.":"Please enter an integer value of 0 or greater for the number of Backup IDs needed for recovery.","Please enter a valid numeric amount":"請輸入有效的數字值","Please enter a valid numeric amount.":"請輸入有效的數值。","Please enter a valid numeric fee":"請輸入有效的數字費用值","Please enter a valid numeric interval length":"請輸入有效的數字間隔長度值","Please enter a valid numeric spendable amount":"請輸入有效的可花用數字量值","Please enter a valid pubkey":"請輸入有效的公鑰","Please enter a valid token name":"Please enter a valid token name","Please enter valid wallet name":"Please enter valid wallet name","Please enter your passphrase":"Please enter your passphrase","Please enter your passphrase to proceed":"Please enter your passphrase to proceed","Please finish syncing before making a transaction":"請在交易前先完成同步","Please select amount":"請選擇數量","Please select backup file first":"請先選擇一個備份檔","Please select buy or sell":"請選擇購入或售出","Please select coin colour":"請選擇代幣顏色","Please select coin type":"Please select coin type","Please specify final directory":"請指定最終資料夾","Please specify temporary directory":"請指定暫存目錄","Please wait for synchronization":"請等候同步","Please wait for wallet synchronization":"請等候錢包同步","Plot":"耕地","Plot Count":"耕地數量","Plot Id":"耕地 ID","Plot Key":"耕地金鑰","Plot NFT":"農場","Plot NFT with p2_singleton_puzzle_hash {plotNFTId} does not exists":["耕地 NFT 的 p2_singleton_puzzle_hash ",["plotNFTId"]," 不存在"],"Plot Name":"Plot Name","Plot Public Key":"耕地公鑰","Plot Seed":"Plot Seed","Plot Size":"耕地大小","Plot in Parallel":"平行耕地","Plot is duplicate of {0}":["耕地與 ",["0"]," 重複"],"PlotNFT is transitioning to (target state)":"PlotNFT is transitioning to (target state)","PlotNFT is transitioning to (target state). This can take a while. Please do not close the application until this is finalized.":"農場正在切換到 (target state). 這可能會需要一段時間. 在完成之前, 請不要關閉應用程式.","Plots":"耕地","Plots Passed Filter":"已通過篩選耕地","Plots are allocated space on your hard drive used to farm and earn Chia. <0>Learn more</0>":"耕地是在你硬碟上劃出的空間，用於耕種來賺取 Chia 。 <0>了解更多</0>","Plotting":"規劃耕地中","Plotting in parallel can save time. Otherwise, add plot(s) to the queue.":"平行規劃耕地可節省時間。若不，新增耕地至隊列。","Plotting with bitfield enabled has about 30% less overall writes and is now almost always faster. You may see reduced memory requirements with bitfield plotting disabled. If your CPU design is from before 2010 you may have to disable bitfield plotting.":"以 bitfield 規劃耕地可減少約 30％ 的總寫入量且如今也幾乎較快。 關閉 Bitfield 規劃耕地有可能降低記憶體需求。如果你的 CPU 是 2010 年以前的設計很有可能需要關閉 bitfield 規劃耕地。","Plotting with bitfield enabled has about 30% less overall writes and is now almost always faster. You may see reduced memory requirements with bitfield plotting disabled. If your CPU design is from before 2010 you may have to disbale bitfield plotting.":"Plotting with bitfield enabled has about 30% less overall writes and is now almost always faster. You may see reduced memory requirements with bitfield plotting disabled. If your CPU design is from before 2010 you may have to disbale bitfield plotting.","Points Found Since Start":"自開始以來所發現的積分","Points Found in Last 24 Hours":"最近 24 小時內所發現的積分","Points Successful in Last 24 Hours":"過去24小時有效的積分","Pool":"農會","Pool Contract Address":"農會合約地址","Pool Key":"池金鑰","Pool Login Link":"農會登入連結","Pool Payout Instructions":"查看農會支付說明","Pool Public Key":"礦池公鑰","Pool Puzzle Hash":"池拼圖雜湊值","Pool Reward Address":"礦池獎勵地址","Pool Reward Address is not properly formatted.":"農會獎勵地址並不符合格式。","Pool Reward Address is not properly formatted. Using 'paste' is recommended.":"Pool Reward Address is not properly formatted. Using 'paste' is recommended.","Pool Reward Address must not be empty.":"農會獎勵地址不能為空。","Pool Reward Amount":"礦池勵數量","Pool does not provide relative_lock_height.":"農會未提供 相對_鎖定_高度。","Pool does not provide target_puzzle_hash.":"農會未提供 目標_拼圖_雜湊值。","Pool public key:":"礦池公鑰 (PPK):","Pool:":"農會：","Pooling":"正在為農會耕種中","Pooling Wallet":"Pooling Wallet","Port":"通訊埠","Preparing Plot NFT":"正在準備農場","Preparing standard wallet":"正在準備標準錢包","Previous":"上一個","Previous Block Hash":"Previous Block Hash","Previous Header Hash":"前一個表頭雜湊值","Previous Sub Block Hash":"Previous Sub Block Hash","Private key with public fingerprint {fingerprint}":["帶有公有指紋 ",["fingerprint"]," 的私鑰"],"Private key {0}":["私鑰 ",["0"]],"Private key {fingerprint}":["Private key ",["fingerprint"]],"Private key:":"私鑰：","Proof of Space Size":"空間證明的大小","Proofs Found":"發現證明數","Protocol Version":"協議版本","Pubkey":"公鑰","Public key:":"公鑰：","Puzzlehash":"拼圖哈希值","Queue Name":"隊列名稱","Queue name":"隊列名稱","Queued":"已隊列","Quit":"Quit","RAM max usage":"最大記憶體使用量","RL Wallet":"RL Wallet","RLListItems.title":"RLListItems.title","Rate Limited":"已限速","Rate Limited Info":"速率限制資訊","Rate Limited Options":"速率限制選項","Rate Limited User Wallet Setup":"速率限制使用者錢包設定","Receive Address":"接收位址","Recipient address is not a coloured wallet address. Please enter a coloured wallet address":"Recipient address is not a coloured wallet address. Please enter a coloured wallet address","Recover":"恢復","Recover DID Wallet":"恢復 DID 錢包","Recover DID wallet":"Recover DID wallet","Recover Distributed Identity Wallet":"還原分散式識別的錢包","Recover Wallet":"恢復錢包","Recoverable Wallet":"Recoverable Wallet","Recovery Wallet":"Recovery Wallet","Refresh Plots":"刷新耕地","Relative Lock Height":"相對鎖定高度","Release Notes":"發行說明","Remove keys from old keyring upon successful migration":"Remove keys from old keyring upon successful migration","Removing":"移除中","Rename":"更改名稱","Rename Wallet":"Rename Wallet","Report an Issue...":"回報問題...","Restore Metadata for Coloured Coins and other Smart Wallets from Backup":"從備份還原有色代幣和其他智慧錢包的元數據","Restore Smart Wallets From Backup":"Restore Smart Wallets From Backup","Retire":"Retire","SB Height":"SB Height","Safe To Skip":"可安全跳過","Save":"儲存","Save passphrase":"Save passphrase","Search":"Search","Search block by header hash":"以表頭雜湊值搜尋區塊","Second temporary folder location":"第二暫存資料夾位置","See private key":"看私鑰","Seed:":"種子：","Select 2nd Temporary Directory":"選擇第二暫存資料夾","Select Final Directory":"選擇最終資料夾","Select Key":"選擇金鑰","Select Offer":"選擇提案","Select Temporary Directory":"選擇暫存資料夾","Select Wallet":"Select Wallet","Select Wallet Type":"選擇錢包類型","Select the final destination for the folder where you would like the plot to be stored. We recommend you use a large slow hard drive (like external HDD).":"選擇你想要儲存耕地的最終資料夾。我們建議你使用較大的慢速硬碟（像是外接硬碟）。","Select the temporary destination for the folder where you would like the plot to be stored. We recommend you use a fast SSD.":"Select the temporary destination for the folder where you would like the plot to be stored. We recommend you use a fast SSD.","Select the temporary destination for the folder where you would like the plot to be stored. We recommend you use a fast drive.":"選擇你想要儲存耕地的暫存資料夾。我們建議你使用快速硬碟。","Select your Plot NFT":"選擇您的農場","Select your Plot NFT from the dropdown or create a new one.":"在下拉式選單中選擇你的農場或是創建一個新的。","Selected":"已選擇","Selected recovery file:":"選擇的還原檔案：","Self Pooling":"正在為自己的農會耕種中","Self pool. When you win a block you will earn XCH rewards.":"自己的農會。當你贏取一個區塊時你將獲得 XCH 獎勵。","Sell":"賣出","Send":"發送","Send this info packet to your Rate Limited Wallet user who must use it to complete setup of their wallet:":"發送這個資訊封包給你必須需要完成設定他們錢包的速率限制錢包使用者：","Send your pubkey to your Rate Limited Wallet admin:":"發送你的公鑰到你的速率限制錢包管理員：","Show Advanced Options":"顯示進階選項","Show Asset Id":"Show Asset Id","Show Token and Asset Issuance Limitations":"Show Token and Asset Issuance Limitations","Show app":"Show app","Side":"方向","Sign In":"登入","Skip":"Skip","Skip Keyring Migration":"Skip Keyring Migration","Skips adding [final dir] to harvester for farming":"Skips adding [final dir] to harvester for farming","Skips adding a final directory to harvester for farming":"跳過新增最終資料夾到收割機進行耕種","Smooth out your XCH farming rewards by joining a pool.":"加入農會來獲得穩定的XCH獎勵。","Something went wrong":"某個地方出錯了","Speech":"語音","Spendable Amount":"可花用數量","Spendable Amount Per Interval":"每間隔可花用數量","Spendable Balance":"可花用餘額","Spending Interval (number of blocks): {interval}":["花用間隔 (區塊數量)：",["interval"]],"Spending Interval Length (number of blocks)":"花費間隔長度 (區塊數量)","Spending Limit (chia per interval): {0}":["花用限制（每間隔可用 Chia）：",["0"]],"Standard Wallet":"Standard Wallet","Starting Wallet":"Starting Wallet","Starting service {0}":["Starting service ",["0"]],"Starting {0}":["Starting ",["0"]],"State":"狀態","Status":"狀態","Status:":"狀態：","StatusCard.connections":"StatusCard.connections","StatusCard.height":"StatusCard.height","StatusCard.status":"StatusCard.status","StatusCard.title":"StatusCard.title","Stripe Size":"Stripe Size","Sub Block Height":"Sub Block Height","Sub Block at height {0} in the Chia blockchain":["Sub Block at height ",["0"]," in the Chia blockchain"],"Submit":"確認","Success":"Success","Synced":"已同步","Syncing":"同步中","Syncing <0/>/<1/>":"同步中 <0/> / <1/>","Syncing {progress}/{tip}":["Syncing ",["progress"],"/",["tip"]],"TXCH Block Rewards":"TXCH Block Rewards","TXCH Farming Rewards":"TXCH Farming Rewards","TXCH Fees Collected":"TXCH Fees Collected","TXCH User Fees":"TXCH User Fees","TXCH User Transaction Fees":"TXCH User Transaction Fees","Target Puzzle Hash":"目標拼圖雜湊值","Temporary files are created during the plotting process which exceeds the size of the final plot files. Make sure you have enough space. <0>Learn more</0>":"Temporary files are created during the plotting process which exceeds the size of the final plot files. Make sure you have enough space. <0>Learn more</0>","Temporary folder location":"暫存資料夾位置","The application will stop working at block height 193536.":"這個應用程式會在區塊高度 193536 時停止運作。","The full node that your farmer is connected to is below. <0>Learn more</0>":"你的塊農連接上的完整節點列在下方。<0>了解更多</0>","The minimum required size for mainnet is k=32":"主網最小要求的大小是 k=32","The node is not synced":"此節點尚未同步","The node is syncing, which means it is downloading blocks from other nodes, to reach the latest block in the chain":"節點同步中，這代表正從其他節點中下載區塊，直到到達區塊鏈的最新區塊","The number of Backup IDs needed for recovery cannot exceed the number of Backup IDs added.":"The number of Backup IDs needed for recovery cannot exceed the number of Backup IDs added.","The pool URL \"{normalizedUrl}\" is not working. Is it pool? Error: {0}":["農會網址：「 ",["normalizedUrl"]," 」錯誤。這是農會嗎？錯誤碼： ",["0"]],"The pool URL is not valid. {normalizedUrl}":["農會網址無效. ",["normalizedUrl"]],"The pool URL needs to use protocol https. {normalizedUrl}":["農會網址需要使用https協議. ",["normalizedUrl"]],"The pool URL speciefied is not valid. {normalizedUrl}":["The specified pool URL is not valid. ",["normalizedUrl"]],"The pool URL speciefied is not valid. {poolUrl}":["The pool URL specified is not valid. ",["poolUrl"]],"The pool switching operation was cancelled, please try again by changing pool, or self pooling":"切換農會操作已被取消，請重試切換農會或自己的農會。","The seed used to create the plot. This depends on the pool pk and plot pk.":"用來建立耕地的種子。這取決於池主鍵和耕地主鍵。","The total number of VDF (verifiable delay function) or proof of time iterations on the whole chain up to this sub block.":"總 VDF（可驗證的延遲函數）數量或整個鏈上直到此子區塊的時間疊代證明。","The total number of VDF (verifiable delay function) or proof of time iterations on this block.":"VDF（可驗證的延遲函數）數量或此區塊上的時間疊代證明。","The total transactions fees in this block. Rewarded to the farmer.":"在這個區塊的總交易費用。是獎勵給塊農的。","These are the instructions for how the farmer wants to get paid. By default this will be an XCH address, but it can be set to any string with a size of less than 1024 characters, so it can represent another blockchain or payment system identifier.":"這些是關於農夫如何獲得報酬的說明. 正常的狀況下, 這將是一個 XCH 地址, 但它可以設置為小於 1024 個字符的任何字符串，因此它可以代表另一個區塊鏈或支付系統識別碼.","These plots are invalid, you might want to delete them.":"這些耕地是無效的，你可能想要刪除他們。","This access token is verified":"This access token is verified","This access token is whitelisted":"This access token is whitelisted","This allows you to add a directory that has plots in it. If you have not created any plots, go to the plotting screen.":"這允許你增加已存有耕地的資料夾。如果你還沒建立任何的耕地，去耕地規劃畫面吧。","This difficulty is an artifically lower difficulty than on the real network, and is used when farming, in order to find more proofs and send them to the pool. The more plots you have, the higher difficulty you will have. However, the difficulty does not affect rewards.":"為了在耕種時找到更多證明以傳送至網路，此處難度與實際網路相比已人工降低過。你擁有的耕地越多，你的難度就越高。 但是，難度不影響獎勵。","This feature is available only from the GUI.":"這個功能只能在 GUI 中提供。","This is the amount of Chia that you can currently use to make transactions. It does not include pending farming rewards, pending incoming transactions, and Chia that you have just spent but is not yet in the blockchain.":"這是你目前可用來交易的 Chia 數量。他不包括未結的耕種獎勵，未結的匯入交易和你剛花用，但還沒在區塊鏈內的 Chia 。","This is the pending change, which are change coins which you have sent to yourself, but have not been confirmed yet.":"這是待處理的改變，是你發送給自己的但還沒確認的改變。","This is the sum of the incoming and outgoing pending transactions (not yet included into the blockchain). This does not include farming rewards.":"這是匯入和匯出的未結交易的總和（還沒包含在區塊鏈中）。這不包括耕種獎勵。","This is the time of the latest peak sub block.":"這是最後的峰值子區塊的時間。","This is the time the block was created by the farmer, which is before it is finalized with a proof of time":"這是塊農建立區塊的時間，這是在用時間證明最終確定之前的時間","This is the total amount of chia in the blockchain at the current peak sub block that is controlled by your private keys. It includes frozen farming rewards, but not pending incoming and outgoing transactions.":"這是由你的私鑰控制的目前峰值子區塊中區塊鏈中 Chia 的總量。他包括凍結耕種獎勵，但不包括未結的匯入和匯出交易。","This is the total balance + pending balance: it is what your balance will be after all pending transactions are confirmed.":"這是總餘額 + 未結餘額：這是在全部未結交易結清後你的餘額。","This is the total number of points this plotNFT has with this pool, since the last payout. The pool will reset the points after making a payout.":"自上次支付以來該農場在該池中的總積分。農會在支付過後將重置積分。","This is the total number of points your farmer has found for this plot NFT. Each k32 plot will get around 10 points per day, so if you have 10TiB, should should expect around 1000 points per day, or 41 points per hour.":"這是您的農夫為此農場找到的總積分。 每個 K32 圖每天將獲得大約 10 點積分，所以如果您有 10TiB，則每天大約 1000 點積分，或每小時 41 點積分。","This node is fully caught up and validating the network":"此節點已完全跟上和驗證網路","This plot NFT is assigned to a different key. You can still create plots for this plot NFT, but you can not make changes.":"此農場分配給其他的金鑰。你仍然可以為此農場建立耕地，但不能進行更改。","This plot NFT is not connected to pool":"這個農場目前沒有連線到農會.","This table shows you the last time your farm attempted to win a block challenge.":"This table shows you the last time your farm attempted to win a block challenge.","This table shows you the last time your farm attempted to win a block challenge. <0>Learn more</0>":"這個表格顯示上次你的農場試著獲勝的區塊挑戰。<0>了解更多</0>","This trade was created at this time":"此貿易是在這個時候建立的","This trade was included on blockchain at this block height":"此貿易已包含在區塊鏈中的這個區塊高度","This version of Chia is no longer compatible with the blockchain and can not safely farm.":"此版本的 Chia 不再與區塊鏈相容，因此無法安全地進行耕種。","Time Created":"建立時間","Time this trade was created at this time":"Time this trade was created at this time","Timestamp":"時間戳記","To":"到","Token and Asset Issuance Limitations":"Token and Asset Issuance Limitations","Token has empty asset id":"Token has empty asset id","Token has empty name":"Token has empty name","Token has empty tail":"Token has empty tail","Total Balance":"總餘額","Total Chia Farmed":"Total Chia Farmed","Total Iterations":"總疊代次數","Total Network Space":"總網路空間","Total Plot Size:":"總耕地大小：","Total Size of Plots":"總耕地大小","Total VDF Iterations":"總 VDF 疊代次數","Total iterations since the start of the blockchain":"從區塊鏈開始後的總疊代次數","Trade":"Trade","Trade Details":"交易明細","Trade ID":"貿易 ID","Trade ID:":"貿易 ID:","Trade Overview":"貿易概觀","Trades will show up here":"貿易將會在此顯示","Trading":"貿易中","Trading History":"貿易歷史","Transaction has not been sent to node yet":"Transaction has not been sent to node yet","Transaction has sent to a full node and is pending inclusion into the mempool. {0}":["Transaction has sent to a full node and is pending inclusion into the mempool. ",["0"]],"Transaction has successfully been sent to a full node and included in the mempool.":"Transaction has successfully been sent to a full node and included in the mempool.","Transactions":"Transactions","Transactions Filter Hash":"交易過濾雜湊值","Try Again":"Try Again","Type":"類型","Unable to create plot NFT":"無法建立農場","Unable to load the list of the keys":"Unable to load the list of the keys","Unclaimed Rewards":"尚未領取的獎勵","Unfinished":"未完成","Unique identifier":"唯一識別碼","Unknown":"未知","Unlock Keyring":"Unlock Keyring","Unsaved Changes":"尚未儲存變更","User Pubkey":"使用者公鑰","VDF Sub Slot Iterations":"VDF 子欄位疊代次數","Value is too high":"Value is too high","Value seems high":"數值有點高","Verify Pool Details":"驗證農會細節","View":"檢視","View Log":"檢視日誌","View Offer":"檢視提案","View Payout Instructions":"View Payout Instructions","View Pool Login Link":"顯示農會登入連結","View pending balances":"檢視未結餘額","View pending balances...":"檢視未結餘額...","Wait for synchronization":"等待同步中","Waiting for launch":"Waiting for launch","Waiting for network to launch":"Waiting for network to launch","Waiting for the transaction to be confirmed":"正在等待交易結算","Wallet Status:":"錢包狀態：","Wallet does not exists":"錢包不存在","Wallet id is not defined":"Wallet id is not defined","WalletItem.ccWallet":"WalletItem.ccWallet","WalletItem.chiaWallet":"WalletItem.chiaWallet","WalletItem.didWallet":"WalletItem.didWallet","WalletItem.rlWallet":"WalletItem.rlWallet","WalletStatusCard.not_synced":"WalletStatusCard.not_synced","WalletStatusCard.synced":"WalletStatusCard.synced","WalletStatusCard.syncing":"WalletStatusCard.syncing","Wallets":"錢包","Wallets.title":"Wallets.title","Want to Join a Pool? Create a Plot NFT":"想加入農會？建立一個農場","Want to earn more Chia? Add more plots to your farm.":"想要獲得更多 Chia？增加更多耕地到你的農場。","Want to have a delay before the next plot starts?":"想要在下次規劃耕地開始前有一段延遲？","Warning: This key is used for a wallet that may have a non-zero balance. By deleting this key you may lose access to this wallet":"警告：這個金鑰的錢包還有餘額，如果刪除金鑰可能會導致你遺失這個錢包","Warning: This key is used for your farming rewards address. By deleting this key you may lose access to any future farming rewards":"警告：這個金鑰用於你的耕種獎勵地址。刪除金鑰後可能會無法領取之後的耕種獎勵。","Warning: This key is used for your pool rewards address. By deleting this key you may lose access to any future pool rewards":"警告：這個金鑰用於你的農會獎勵地址。刪除金鑰後可能會無法領取之後的農會獎勵。","Weight":"權重","Weight is the total added difficulty of all sub blocks up to and including this one":"權重是所有包括這個的子區塊的總增加難度","Welcome to Chia. Please log in with an existing key, or create a new key.":"歡迎來到 Chia。請用已有的金鑰登入，或建立新的金鑰。","Welcome! The following words are used for your wallet backup. Without them, you will lose access to your wallet, keep them safe! Write down each word along with the order number next to them. (Order is important)":"歡迎！以下的字是用來為你的錢包備份。沒有這些你將會無法進入你的錢包。將這些安全保存好！寫下每個字，包括他們旁邊的順序數字。（順序很重要）","When you receive the setup info packet from your admin, enter it below to complete your Rate Limited Wallet setup:":"當你從你的管理員收到設定資訊封包時，在下面輸入以完成你的速率限制錢包設定：","Window":"視窗","Without fees":"未計雜費","Yes":"是","You are in pending state. Please wait for confirmation":"您處於待確認的狀態。請等待確認","You are not self pooling":"你目前不在自己的農會。","You can still create plots for this plot NFT, but you can not make changes until sync is complete.":"您仍然可以為此農場創建耕地，但在同步完成之前您不能進行更改。","You do not need to be synced or connected to plot. Temporary files are created during the plotting process which exceed the size of the final plot files. Make sure you have enough space. <0>Learn more</0>":"You do not need to be synced or connected to plot. Temporary files are created during the plotting process which exceed the size of the final plot files. Make sure you have enough space. <0>Learn more</0>","You do not need to be synched or connected to Plot. Temporary files are created during the plotting process which exceed the size of the final plot files. Make sure you have enough space. <0>Learn more</0>":"You do not need to be synched or connected to Plot. Temporary files are created during the plotting process which exceed the size of the final plot files. Make sure you have enough space. <0>Learn more</0>","You do not need to be synched or connected to plot. Temporary files are created during the plotting process which exceed the size of the final plot files. Make sure you have enough space. <0>Learn more</0>":"你不需要同步或連線後才能規劃耕地。規劃耕地時所產生的暫存檔將大於最終的耕地大小。確保你有足夠的空間。<0>了解更多</0>","You have made changes. Do you want to discard them?":"您有做過變更。請問您想要放棄它們嗎？","You have {0}% of the space on the network, so farming a block will take {expectedTimeToWin} in expectation.":["You have ",["0"],"% of the space on the network, so farming a block will take ",["expectedTimeToWin"]," in expectation."],"You have {0}% of the space on the network, so farming a block will take {expectedTimeToWin} in expectation. Actual results may take 3 to 4 times longer than this estimate.":["你擁有 ",["0"]," ％的網路空間，耕種一個區塊預計要 ",["expectedTimeToWin"]," 。實際結果可能是預計的 3 至 4 倍。"],"You have {0}% of the space on the network, so farming a block will take {expectedTimeToWin} in expectation. Actual results may take 3-4 times longer - actual results may be 3x this estimate.<0>Learn about Expected Value</0>":["You have ",["0"],"% of the space on the network, so farming a block will take ",["expectedTimeToWin"]," in expectation. Actual results may take 3-4 times longer - actual results may be 3x this estimate.<0>Learn about Expected Value</0>"],"You have {0}% of the space on the network, so farming a block will take {expectedTimeToWin} in expectation. Actual results may take 3-4 times longer this estimate. <0>Learn about Expected Value</0>":["You have ",["0"],"% of the space on the network, so farming a block will take ",["expectedTimeToWin"]," in expectation. Actual results may take 3-4 times longer this estimate. <0>Learn about Expected Value</0>"],"You need to claim your rewards first":"您需要先領取你的獎勵","You need to upgrade the Chia application as this version will stop working soon!":"You need to upgrade the Chia application as this version will stop working soon!","You need {currencyCode} to join a pool.":["你需要 ",["currencyCode"]," 來加入農會。"],"You will recieve <0/> to {0}":["你將會收到 <0/> 到 ",["0"]],"You will recieve <0/> to {address}":["你將會收到 <0/> 到 ",["address"]],"Your DID requires at least {dids_num_req} attestation file{0} for recovery. Please upload additional files.":["Your DID requires at least ",["dids_num_req"]," attestation file",["0"]," for recovery. Please upload additional files."],"Your Farm Overview":"你的農場概觀","Your Full Node Connection":"你的完整節點連線","Your Harvester Network":"你的收割機網路","Your Pool Overview":"您的農場概觀","Your keyring is locked":"Your keyring is locked","Your keys have not been migrated to a new keyring. You will be unable to create new keys or delete existing keys until migration completes. Are you sure you want to skip migrating your keys?":"Your keys have not been migrated to a new keyring. You will be unable to create new keys or delete existing keys until migration completes. Are you sure you want to skip migrating your keys?","Your keys need to be migrated to a new keyring that is optionally secured by a master passphrase.":"Your keys need to be migrated to a new keyring that is optionally secured by a master passphrase.","Your keys need to be migrated to a new keyring that is secured by a master passphrase.":"Your keys need to be migrated to a new keyring that is secured by a master passphrase.","Your passphrase can be stored in your system's secure credential store. Chia will be able to access your keys without prompting for your passphrase.":"Your passphrase can be stored in your system's secure credential store. Chia will be able to access your keys without prompting for your passphrase.","[Error 13] Permission denied. You are trying to access a file/directory without having the necessary permissions. Most likely one of the plot folders in your config.yaml has an issue.":"［錯誤 13］權限被拒。你想存取一個檔案／目錄但沒有所需的權限。很有可能你 config.yaml 裡其中一個耕地資料夾有問題。","[Error 22] File not found. Most likely one of the plot folders in your config.yaml has an issue.":"［錯誤 22］找不到檔案。很有可能你 config.yaml 裡其中一個耕地資料夾有問題。","connections:":"連線：","height:":"高度：","not synced":"未同步","p2_singleton_puzzle_hash is not defined":"\"p2_singleton_puzzle_hash\"參數未定義","status:":"狀態：","synced":"已同步","syncing":"同步中","{0} Block Rewards":[["0"]," Block Rewards"],"{0} User Transaction Fees":[["0"]," User Transaction Fees"],"{0} points {1} - {2} hours ago":[["0"]," 積分 ",["1"]," - ",["2"]," 小時前"],"{currencyCode} Block Rewards":[["currencyCode"]," 區塊獎勵"],"{currencyCode} Total Chia Farmed":[["currencyCode"]," 已收耕 Chia"],"{currencyCode} User Transaction Fees":[["currencyCode"]," 使用者交易費用"],"{mojo, plural, one {mojo} other {mojos}}":[["mojo","plural",{other:"mojos"}]]}};
+/*eslint-disable*/module.exports={messages:{"(Optional)":"(選用)","*Want to explore Chia’s blocks further? Check out <0>Chia Explorer</0> built by an open source developer.":"＊想更進一步探索 Chia 區塊鍵？看看開源開發者打造的 <0>Chia Explorer</0>","+":"+","+ Add New Plot NFT":"+ 新增新的農場","+ Add Token":"+ Add Token","+ Add Wallet":"+ 新增錢包","-":"-","0 automatically chooses bucket count":"0 automatically chooses bucket count","128 buckets is recommended":"建議使用 128 個桶","A harvester is a service running on a machine where plot(s) are actually stored. A farmer and harvester talk to a full node to see the state of the chain. View your network of connected harvesters below Learn more":"收割機是在實際存放耕地的機器上運作的服務。塊農和收割機會與一個完整節點溝通以取得區塊鍵狀態。以下是你的收割機與塊農網路 學習更多","About Chia Blockchain":"關於 Chia 區塊鏈","About Chia Wallet":"About Chia Wallet","Accept":"接受","Accept Offer":"Accept Offer","Accepted Offers":"Accepted Offers","Accepted at time:":"接受於：","Accepted on Date":"Accepted on Date","Action":"動作","Actions":"動作","Add":"新增","Add Backup ID":"增加備份 ID","Add CAT wallets to have more options":"Add CAT wallets to have more options","Add Custom Token":"Add Custom Token","Add Plot Directory":"新增耕地資料夾","Add Plot to Queue":"新增耕地至隊列","Add Token":"Add Token","Add Wallet":"Add Wallet","Add a Plot":"新增耕地","Add a Plot NFT":"新增一塊農場","Add a plot":"新增耕地","Add plot directory":"新增耕地資料夾","Add {currencyCode} from the Faucet":["從水龍頭新增",["currencyCode"],"."],"Adding {0} token":["Adding ",["0"]," token"],"Address":"位址","Address / Puzzle hash":"位址 / 拼圖雜湊值","After your keys are successfully migrated to the new keyring, you may choose to have your keys removed from the old keyring.":"After your keys are successfully migrated to the new keyring, you may choose to have your keys removed from the old keyring.","All":"All","Amount":"數量","Amount ({cc_unit})":["數量 (",["cc_unit"],")"],"Amount For Initial Coin":"初始代幣數量","Amount exceeds spendable balance":"Amount exceeds spendable balance","Amount must be an even amount.":"Amount must be an even amount.","Are you sure you want to cancel your offer?":"Are you sure you want to cancel your offer?","Are you sure you want to delete the plot? The plot cannot be recovered.":"你確要刪除這個耕地嗎？此耕地將無法尋回。","Are you sure you want to delete unconfirmed transactions?":"你確認要刪除未確認的交易?","Are you sure you want to disconnect?":"您確定要斷線嗎？","Are you sure you want to quit?":"Are you sure you want to quit?","Are you sure you want to quit? GUI Plotting and farming will stop.":"你確定要退出？耕地規劃和耕種將會停止。","Are you sure you want to use k={plotSize}":["Are you sure you want to use k=",["plotSize"]],"Are you sure you want to use k={plotSize}?":["你確定 k 的大小為 ",["plotSize"],"?"],"Asset Id":"Asset Id","Asset Type":"Asset Type","Atomic Swap Wallet":"Atomic Swap Wallet","Attempt {attempt}":["Attempt ",["attempt"]],"Authorized Payee Wallet":"Authorized Payee Wallet","Autogenerated name from pool contract address":"自動從農會合約地址建立名稱","Back":"返回","Backup file is used to restore smart wallets.":"備份檔案可用於復原智慧錢包。","Balance":"餘額","Base Farmer Reward Amount":"基礎塊農獎勵數量","Before posting an offer in Keybase to the #{KeybaseChannelName} channel, you must first join the {KeybaseTeamName} team. Please note that it might take a few moments to join the channel.":["Before posting an offer in Keybase to the #",["KeybaseChannelName"]," channel, you must first join the ",["KeybaseTeamName"]," team. Please note that it might take a few moments to join the channel."],"Below are the current block challenges. You may or may not have a proof of space for these challenges. These blocks do not currently contain a proof of time.":"以下是目前的區塊挑戰。你可能會或可能不會有這些挑戰的空間證明。這些區塊目前不包含時間證明。","Best estimate over last 24 hours":"已過 24 小時最佳預估值","Best estimate over last 5 hours":"Best estimate over last 5 hours","Block":"區塊","Block Test":"區塊測試","Block VDF Iterations":"區塊 VDF 疊代次數","Block at height {0} in the Chia blockchain":["Chia 區塊鏈區塊高度 ",["0"]],"Block hash":"Block hash","Block with hash {headerHash}":["帶雜湊值 ",["headerHash"]," 的區塊"],"Block with hash {headerHash} does not exist.":["帶雜湊值 ",["headerHash"]," 的區塊不存在。"],"Blocks":"區塊","Both values are empty. This is ok if you are, for example, sending your rewards to an offline wallet.":"Both values are empty. This is ok if you are, for example, sending your rewards to an offline wallet.","Browse":"瀏覽","But you are currently farming <0/>":"但你目前耕種 <0/>","Buy":"購買","CAT Wallet":"CAT Wallet","CC Wallet":"CC Wallet","CCListItems.title":"CCListItems.title","Can be backed up to mnemonic seed":"可以備份到助記符號種子","Cancel":"取消","Cancel Offer":"Cancel Offer","Cancel and Spend":"取消和花費","Cancel on blockchain":"Cancel on blockchain","Cannot send chia to coloured address. Please enter a chia address.":"Cannot send chia to coloured address. Please enter a chia address.","Caution, deleting these plots will delete them forever. Check that the storage devices are properly connected.":"警告，刪除這些農田區將會是永久刪除的。檢查儲存裝置是否正確連接。","Challenge":"挑戰","Challenge Hash":"挑戰雜湊值","Change":"變更","Change Pool":"變更農會","Chat on KeyBase":"在 KeyBase 上閒聊","Chia":"Chia","Chia Asset Token":"Chia Asset Token","Chia Blockchain Wiki":"Chia 區塊鏈維基","Chia Wallet":"Chia 錢包","Choose Number of Plots":"選擇耕地數量","Choose Plot Size":"選擇耕地大小","Claim Rewards":"領取獎勵","Close":"關閉","Closing down node and server":"關閉節點和伺服器","Coin Name":"幣種名稱","Coins":"Coins","Coins:":"幣種：","Colour":"顏色","Colour Info":"顏色資訊","Colour String":"顏色字串","Colour:":"顏色：","Coloured Coin":"有色代幣","Coloured Coin Options":"有色代幣選項","Confirm":"確認","Confirm Disconnect":"確認斷線","Confirm Passphrase":"Confirm Passphrase","Confirmation":"確認","Confirmed":"Confirmed","Confirmed at Height":"Confirmed at Height","Confirmed at block:":"已確認區塊：","Confirmed at height {0}":["已確認於高度 ",["0"]],"Connect":"連線","Connect to other peers":"連線到其他同儕","Connect to pool":"連線至農會","Connected":"已連線","Connecting do daemon":"Connecting do daemon","Connecting to daemon":"Connecting to daemon","Connecting to wallet":"連線至錢包中","Connection Status":"連線狀態","Connection Status:":"連線狀態：","Connection type":"連線類型","Connections":"連線","Contribute on GitHub":"在 GitHub 上貢獻","Copied":"已複製","Copy":"複製","Copy to Clipboard":"複製到剪貼板","Create":"建立","Create An Attestation Packet":"建立一個證明封包","Create Chia Asset Token Wallet":"Create Chia Asset Token Wallet","Create Chia Asset Token Wallet from Existing TAIL":"Create Chia Asset Token Wallet from Existing TAIL","Create DID wallet":"Create DID wallet","Create Distributed Identity Wallet":"建立分散式識別的錢包","Create New Chia Asset Token Wallet":"Create New Chia Asset Token Wallet","Create New Wallet":"建立新錢包","Create Offer":"建立提案","Create Plot":"Create Plot","Create Rate Limited Admin Wallet":"建立速率限制管理員錢包","Create Rate Limited User Wallet":"建立速率限制使用者錢包","Create Token":"Create Token","Create Trade":"Create Trade","Create Trade Offer":"建立貿易提案","Create Transaction":"建立交易","Create a Backup":"建立備份","Create a Plot NFT":"建立一塊農場","Create a new private key":"建立一個新的私鑰","Create admin wallet":"建立管理員錢包","Create an Offer":"Create an Offer","Create an offer to exchange XCH or other tokens. View an offer to inspect and accept an offer made by another party.":"Create an offer to exchange XCH or other tokens. View an offer to inspect and accept an offer made by another party.","Create custom CAT Wallet":"Create custom CAT Wallet","Create from Existing":"Create from Existing","Create new coloured coin":"建立新的有色代幣","Create user wallet":"建立使用者錢包","Create wallet for colour":"建立有色代幣錢包","Create wallet for existing colour":"用已有的有色代幣去建立錢包","CreateWallet.addWallet":"CreateWallet.addWallet","Created At:":"建立於：","Created by us:":"由我們創造：","Creates and submits a transaction on the blockchain that cancels the offer":"Creates and submits a transaction on the blockchain that cancels the offer","Creating Plot NFT and Joining the Pool":"建立一塊農場並加入農會","Creating Plot NFT for Self Pooling":"正在為自己的農會建立農場","Creation Date":"Creation Date","Currency code is not defined":"未定義貨幣代碼","Current Difficulty":"目前難度","Current Points Balance":"目前積分餘額","Current trade status":"目前交易狀態","Custody Wallet":"Custody Wallet","Custom":"Custom","DANGER: permanently delete private key":"危險：將永久刪除私鑰","DID Wallet":"DID Wallet","DIDListItems.title":"DIDListItems.title","Date":"日期","Delay":"延遲","Delete":"刪除","Delete Key":"Delete Key","Delete Plot":"刪除耕地","Delete Unconfirmed Transactions":"刪除未確認的交易","Delete all keys":"刪除全部金鑰","Delete key":"Delete key","Delete key {fingerprint}":["刪除帳號指紋：",["fingerprint"]],"Deleting all keys will permanently remove the keys from your computer, make sure you have backups. Are you sure you want to continue?":"刪除全部金鑰將會永久從你的電腦中移除，請確保你有備份。你確定要繼續？","Deleting key":"Deleting key","Deleting the key will permanently remove the key from your computer, make sure you have backups. Are you sure you want to continue?":"刪除此金鑰將會永久從你的電腦中移除，請確保你有備份。你確定要繼續？","Details":"Details","Developer":"開發者","Developer Tools":"開發者工具","Difficulty":"難度","Disable bifield plotting when your temporary directory is on fast storage like SSD or you need lower per process RAM requirements. Plotting with bitfield enabled has about 20% less overall writes.":"Disable bifield plotting when your temporary directory is on fast storage like SSD or you need lower per process RAM requirements. Plotting with bitfield enabled has about 20% less overall writes.","Disable bitfield plotting":"禁用 bitfield 耕地","Discard":"放棄","Disconnect":"斷線","Display Offer Data":"Display Offer Data","Distributed Identity":"分散式識別","Distributed Identity Options":"Distributed Identity Options","Do not show this dialog again":"Do not show this dialog again","Do you have existing plots on this machine? <0>Add Plot Directory</0>":"你有已存在的耕地在本機上？ <0>增加耕地資料夾</0>","Does your machine support parallel plotting?":"你的機器支援平行耕地嗎？","Drag & drop an offer file below to view its details":"Drag & drop an offer file below to view its details","Drag and drop attestation packet(s)":"拖放證明封包","Drag and drop offer file":"拖放提案檔案","Drag and drop your backup file":"拖放你的備份檔案","Drag and drop your recovery backup file":"拖放您的備份檔案","Edit":"編輯","Edit Payout Instructions":"編輯支付說明","Enter a strong passphrase and click Migrate Keys to secure your keys":"Enter a strong passphrase and click Migrate Keys to secure your keys","Enter the 24 word mnemonic that you have saved in order to restore your Chia wallet.":"輸入你已存的 24 字助記符號以復原您的 Chia 錢包。","Error":"錯誤","Error the entered address appears to be for a different colour.":"Error the entered address appears to be for a different colour.","Error: Cannot send chia to coloured address. Please enter a chia address.":"錯誤：不能發送 Chia 到有色地址。請輸入 Chia 地址。","Estimated Network Space":"估計的網路空間","Estimated Time to Win":"預計贏取時間","Estimated network space":"Estimated network space","Estimated sum of all the plotted disk space of all farmers in the network":"網路中所有農夫的所有耕地磁盤空間估計總和","Exclude final directory":"排除最終資料夾","Expected Time to Win":"Expected Time to Win","Export Offer File":"Export Offer File","Failed to execute Keybase command: {stderr}":["Failed to execute Keybase command: ",["stderr"]],"Failed to open (invalid plots)":"開啟失敗（無效耕地）","Farm":"耕種","Farmer Public Key":"農夫公鑰","Farmer Puzzle Hash":"塊農拼圖雜湊值","Farmer Reward Address":"塊農獎勵地址","Farmer Reward Address is not properly formatted.":"農夫獎勵地址並不符合格式。","Farmer Reward Address is not properly formatted. Using 'paste' is recommended.":"Farmer Reward Address is not properly formatted. Using 'paste' is recommended.","Farmer Reward Address must not be empty.":"農夫獎勵地址不能為空。","Farmer is not connected":"塊農目前未連線","Farmer is not running":"塊農未運作","Farmer public key:":"農夫公鑰:","Farmers earn block rewards and transaction fees by committing spare space to the network to help secure transactions. This is where your farm will be once you add a plot. <0>Learn more</0>":"農夫可通過提交閒置空間到網路上來幫助保護交易，從而獲得區塊獎勵和交易費。這便是新增耕地後耕種的地方。<0>了解詳情</0>","Farming":"耕種中","Farming Status":"耕種狀態","Fee":"費用","Fee (TXCH)":"Fee (TXCH)","Fee ({0})":["Fee (",["0"],")"],"Fee ({currencyCode})":["費用 (",["currencyCode"],")"],"Fees Amount":"費用數量","Fees Puzzle Hash":"Fees Puzzle Hash","File":"檔案","Filename":"檔案名稱","Final folder location":"最終資料夾位置","Finished":"已完成","Follow on Twitter":"Twitter 上追隨","Frequently Asked Questions":"常見問題","Full Node":"完整節點","Full Node Overview":"完整節點概觀","Full Node Status":"Full Node Status","Full Screen":"全畫面","Generate New Colour":"生成新顏色","Get Pool Login Link":"Get Pool Login Link","Go to #{KeybaseChannelName}":["Go to #",["KeybaseChannelName"]],"Grid view":"網格檢視","HD or Hierarchical Deterministic keys are a type of public key/private key scheme where one private key can have a nearly infinite number of different public keys (and therefor wallet receive addresses) that will all ultimately come back to and be spendable by a single private key.":"層及式確定性金鑰是一種公鑰/私鑰格式，一個私鑰可以具有幾乎無限數量的不同公鑰（以此類推，同錢包接收地址），最終所有這些公鑰都將回朔於一個私鑰使用／花用。","Harvester ID":"Harvester ID","Harvester Plots":"耕地收割機","Hashgreen DEX URL":"Hashgreen DEX URL","Header Hash":"表頭雜湊值","Header hash":"表頭雜湊值","Height":"高度","Help":"幫助","Help translate":"協助翻譯","Hide Advanced Options":"隱藏進階選項","Hint":"Hint","History":"歷史","Host Name":"主機名稱","IP address":"IP位址","IP address / host":"IP 位址 / 主機","If none selected, then it will default to the temporary directory.":"如無選擇，將會預設至臨時資料夾。","If selected, your offer will be not be shared publicly.":"If selected, your offer will be not be shared publicly.","If you have already shared your offer file, you may need to submit a transaction to cancel the pending offer. Click \"Cancel on blockchain\" to submit a cancellation transaction.":"If you have already shared your offer file, you may need to submit a transaction to cancel the pending offer. Click \"Cancel on blockchain\" to submit a cancellation transaction.","Import":"Import","Import Wallet from Mnemonics":"以助記符號匯入錢包","Import from Mnemonics (24 words)":"從助記詞(24 字) 匯入","In Progress":"進行中","In exchange for":"In exchange for","In exchange for:":"In exchange for:","Incoming":"轉入","Incorrect value":"數值不正確","Index":"索引","Indicated if this offer was created by us":"指出如果此提案由我方創建","Indicated what time this offer was accepted":"指出此提案於何時被接受","Info Packet":"資訊封包","Initial Amount":"初始數量","Initialize a Rate Limited User Wallet:":"初始化速率限制使用者錢包：","Install Keybase":"Install Keybase","Insufficient funds available to accept offer. Ensure that your\nspendable balance is sufficient to cover the offer amount.":"Insufficient funds available to accept offer. Ensure that your\nspendable balance is sufficient to cover the offer amount.","Insufficient funds available to create offer. Ensure that your\nspendable balance is sufficient to cover the offer amount.":"Insufficient funds available to create offer. Ensure that your\nspendable balance is sufficient to cover the offer amount.","Interval":"間隔","Invalid state":"無效狀態","It is a one-time login link that can be used to log in to a pool's website. It contains a signature using the farmer's key from the plot NFT. Not all pools support this feature.":"這是一個一次性登錄連結, 可用於登記在農會的網站. 它來自農場, 其包含農民公鑰的簽名. 並非所有農會都支援此功能.","Join Pool":"加入礦池","Join a Pool":"加入一個農場","Join a pool and get consistent XCH farming rewards. The average returns are the same, but it is much less volatile.":"加入農會並獲得持續的 XCH 耕種獎勵。雖然平均報酬一樣但波動較小相對穩定。","Join a pool and get consistent XCH farming rewards. The average returns are the same, but it is much less volatile. Assign plots to a plot NFT. You can easily switch pools without having to re-plot.":"加入農會並獲得持續的XCH 耕種獎勵。雖然平均報酬一樣但波動較小相對穩定。分配耕地到農場。你可以更輕鬆的轉換農會而不用重新進行耕地。","Join a pool and get more consistent XCH farming rewards. Create a plot NFT and assign your new plots to a group.":"加入農會以獲得更多持續性的 XCH 耕種獎勵。建立農場並分配您的新耕地到群組。","Join {KeybaseTeamName}":["Join ",["KeybaseTeamName"]],"K-Size":"K-大小","Keybase command failed {e}. If you haven't installed Keybase, you can download from https://keybase.io":["Keybase command failed ",["e"],". If you haven't installed Keybase, you can download from https://keybase.io"],"Keybase is a secure messaging and file sharing application. To share an offer in the Keybase {KeybaseTeamName} team, you must first have Keybase installed.":["Keybase is a secure messaging and file sharing application. To share an offer in the Keybase ",["KeybaseTeamName"]," team, you must first have Keybase installed."],"Keyring migration failed: {0}":["Keyring migration failed: ",["0"]],"Keys":"金鑰","KiB Up/Down":"KiB 上傳/下傳","Last Attempted Proof":"最後試過證明","Last Height Farmed":"最後耕種高度","Latest Block Challenges":"最新區塊挑戰","Launcher Id":"啟動器 ID","Learn More":"了解更多","Leaving Pool":"離開農會","List of connections is empty":"List of connections is empty","List view":"清單檢視","Loading Plot NFTs":"正在讀取農場","Loading configuration":"Loading configuration","Loading keyring status":"Loading keyring status","Loading list of the keys":"Loading list of the keys","Loading list of wallets":"正在讀取錢包列表","Loading...":"載入中...","Local Harvester Plots":"Local Harvester Plots","Log Out":"Log Out","Logging in":"登入中","MainWalletList.colouredCoin":"MainWalletList.colouredCoin","MainWalletList.createAdminWallet":"MainWalletList.createAdminWallet","MainWalletList.createDIDWallet":"MainWalletList.createDIDWallet","MainWalletList.createNewColouredCoin":"MainWalletList.createNewColouredCoin","MainWalletList.createUserWallet":"MainWalletList.createUserWallet","MainWalletList.createWalletForExistingColour":"MainWalletList.createWalletForExistingColour","MainWalletList.distributedIdentity":"MainWalletList.distributedIdentity","MainWalletList.rateLimited":"MainWalletList.rateLimited","MainWalletList.recoverDIDWallet":"MainWalletList.recoverDIDWallet","MainWalletList.title":"MainWalletList.title","Manage Farming Rewards":"管理耕種獎勵","Manage Offers":"Manage Offers","Manage Recovery DIDs":"管理還原的 DID","Manage Your Farming Rewards Target Addresses":"管理你的耕種獎勵目標地址","Memo":"Memo","Memos":"Memos","MiB Up/Down":"MiB 上傳/下傳","Migrate Keys":"Migrate Keys","Migration required":"Migration required","Minimum Difficulty":"最小難度","Minutes":"分鐘","More memory slightly increases speed":"更多記憶體可略微提高速度","Multi Sig Wallet":"Multi Sig Wallet","My DID Wallet":"我的 DID 錢包","My Pubkey":"我的公鑰","Name":"Name","Network Name":"網路名稱","Network has not yet been launched, once it launches full node will start automatically.":"Network has not yet been launched, once it launches full node will start automatically.","Network not launched yet, waiting for genesis challenge!":"Network not launched yet, waiting for genesis challenge!","New Address":"新位址","New Wallet":"新錢包","Next":"下一步","Nickname":"暱稱","No":"否","No 24 word seed, since this key is imported.":"由這個鑰匙是匯入的，沒有 24 個字種子。","No blocks farmed yet":"還沒有耕種收穫到的區塊","No current offers":"No current offers","No previous transactions":"沒有上一個交易","No private keys for one or both addresses. Safe only if you are sending rewards to another wallet.":"一邊或雙邊地址沒有私鑰。只有你發送獎勵到另一個錢包時才算安全。","Node Count":"Node Count","Node ID":"節點 ID","Node Id":"節點 ID","None":"無","None of your plots have passed the plot filter yet.":"你還沒有任何的耕地通過篩選。","Not Available":"尚無","Not Synced":"未同步","Not accepted":"Not accepted","Not accepted yet":"還未接受","Not confirmed":"Not confirmed","Not confirmed yet":"還未確認","Not connected":"未連線","Not found Plots":"沒找到的耕地","Note that this does not change your pooling payout addresses. This only affects old format plots, and the 0.25XCH reward for pooling plots.":"請注意，這不會變更您的農會支付位址。這只會影響舊的耕地格式和 0.25XCH 的農會耕地獎勵。","Number of Plots":"耕地數量","Number of buckets":"桶數量","Number of threads":"執行緒數：","OK":"好","Offer":"提案","Offer Data":"Offer Data","Offer Identifier":"Offer Identifier","Offer Shared":"Offer Shared","Offer has been accepted and is awaiting confirmation.":"Offer has been accepted and is awaiting confirmation.","OfferBin URL":"OfferBin URL","Offered":"Offered","Offers Created":"已建立提案","Offers you accepted":"Offers you accepted","Offers you created":"Offers you created","On average there is one minute between each transaction block. Unless there is congestion you can expect your transaction to be included in less than a minute.":"平均每個交易區塊之間有 1 分鐘的時間。除非出現擁塞，否則你可以在不到 1 分鐘的時間內完成交易。","Once you accept this offer, you will not be able to cancel the transaction. Are you sure you want to accept this offer?":"Once you accept this offer, you will not be able to cancel the transaction. Are you sure you want to accept this offer?","One or more unknown tokens are being offered. Please verify that the asset IDs of the tokens listed below match the asset IDs of the tokens you expect to receive.":"One or more unknown tokens are being offered. Please verify that the asset IDs of the tokens listed below match the asset IDs of the tokens you expect to receive.","Only one backup file is allowed.":"只允許一個備份檔案。","Only one backup file is allowed. To add a new backup file, please delete any previously uploaded files and try again.":"Only one backup file is allowed. To add a new backup file, please delete any previously uploaded files and try again.","Outgoing":"匯出","Parent Coin":"Parent Coin","Passphrase":"Passphrase","Passphrase Hint":"Passphrase Hint","Passphrase Hint (Optional)":"Passphrase Hint (Optional)","Paste Offer Data":"Paste Offer Data","Payout Address":"付款地址","Peak Height":"峰值高度","Peak Sub-block Height":"Peak Sub-block Height","Peak Time":"峰值時間","Pending":"待確認中","Pending Balance":"未結餘額","Pending Change":"未結更動","Pending Total Balance":"未結總餘額","Please Confirm":"請確認","Please add a trade pair":"請增加一對貿易方","Please enter 0 fee. Positive fees not supported yet for RL.":"請輸入 0 費用。 RL 尚未支援正向費用。","Please enter a coin":"請輸入幣種","Please enter a filename":"請輸入檔案名稱","Please enter a passphrase":"Please enter a passphrase","Please enter a pubkey":"請輸入公鑰","Please enter a puzzlehash":"請輸入一個 PuzzleHash","Please enter a valid CAT name":"Please enter a valid CAT name","Please enter a valid TAIL":"Please enter a valid TAIL","Please enter a valid initial coin amount":"請輸入有效的初始金額","Please enter a valid integer of 0 or greater for the number of Backup IDs needed for recovery.":"Please enter an integer value of 0 or greater for the number of Backup IDs needed for recovery.","Please enter a valid numeric amount":"請輸入有效的數字值","Please enter a valid numeric amount.":"請輸入有效的數值。","Please enter a valid numeric fee":"請輸入有效的數字費用值","Please enter a valid numeric interval length":"請輸入有效的數字間隔長度值","Please enter a valid numeric spendable amount":"請輸入有效的可花用數字量值","Please enter a valid pubkey":"請輸入有效的公鑰","Please enter a valid token name":"Please enter a valid token name","Please enter an amount for each row":"Please enter an amount for each row","Please enter valid wallet name":"Please enter valid wallet name","Please enter your passphrase":"Please enter your passphrase","Please enter your passphrase to proceed":"Please enter your passphrase to proceed","Please finish syncing before making a transaction":"請在交易前先完成同步","Please finish syncing before managing offers":"Please finish syncing before managing offers","Please select amount":"請選擇數量","Please select an asset for each row":"Please select an asset for each row","Please select backup file first":"請先選擇一個備份檔","Please select buy or sell":"請選擇購入或售出","Please select coin colour":"請選擇代幣顏色","Please select coin type":"Please select coin type","Please specify final directory":"請指定最終資料夾","Please specify temporary directory":"請指定暫存目錄","Please wait for synchronization":"請等候同步","Please wait for wallet synchronization":"請等候錢包同步","Plot":"耕地","Plot Count":"耕地數量","Plot Id":"耕地 ID","Plot Key":"耕地金鑰","Plot NFT":"農場","Plot NFT with p2_singleton_puzzle_hash {plotNFTId} does not exists":["耕地 NFT 的 p2_singleton_puzzle_hash ",["plotNFTId"]," 不存在"],"Plot Name":"Plot Name","Plot Public Key":"耕地公鑰","Plot Seed":"Plot Seed","Plot Size":"耕地大小","Plot in Parallel":"平行耕地","Plot is duplicate of {0}":["耕地與 ",["0"]," 重複"],"PlotNFT is transitioning to (target state)":"PlotNFT is transitioning to (target state)","PlotNFT is transitioning to (target state). This can take a while. Please do not close the application until this is finalized.":"農場正在切換到 (target state). 這可能會需要一段時間. 在完成之前, 請不要關閉應用程式.","Plots":"耕地","Plots Passed Filter":"已通過篩選耕地","Plots are allocated space on your hard drive used to farm and earn Chia. <0>Learn more</0>":"耕地是在你硬碟上劃出的空間，用於耕種來賺取 Chia 。 <0>了解更多</0>","Plotting":"規劃耕地中","Plotting in parallel can save time. Otherwise, add plot(s) to the queue.":"平行規劃耕地可節省時間。若不，新增耕地至隊列。","Plotting with bitfield enabled has about 30% less overall writes and is now almost always faster. You may see reduced memory requirements with bitfield plotting disabled. If your CPU design is from before 2010 you may have to disable bitfield plotting.":"以 bitfield 規劃耕地可減少約 30％ 的總寫入量且如今也幾乎較快。 關閉 Bitfield 規劃耕地有可能降低記憶體需求。如果你的 CPU 是 2010 年以前的設計很有可能需要關閉 bitfield 規劃耕地。","Plotting with bitfield enabled has about 30% less overall writes and is now almost always faster. You may see reduced memory requirements with bitfield plotting disabled. If your CPU design is from before 2010 you may have to disbale bitfield plotting.":"Plotting with bitfield enabled has about 30% less overall writes and is now almost always faster. You may see reduced memory requirements with bitfield plotting disabled. If your CPU design is from before 2010 you may have to disbale bitfield plotting.","Points Found Since Start":"自開始以來所發現的積分","Points Found in Last 24 Hours":"最近 24 小時內所發現的積分","Points Successful in Last 24 Hours":"過去24小時有效的積分","Pool":"農會","Pool Contract Address":"農會合約地址","Pool Key":"池金鑰","Pool Login Link":"農會登入連結","Pool Payout Instructions":"查看農會支付說明","Pool Public Key":"礦池公鑰","Pool Puzzle Hash":"池拼圖雜湊值","Pool Reward Address":"礦池獎勵地址","Pool Reward Address is not properly formatted.":"農會獎勵地址並不符合格式。","Pool Reward Address is not properly formatted. Using 'paste' is recommended.":"Pool Reward Address is not properly formatted. Using 'paste' is recommended.","Pool Reward Address must not be empty.":"農會獎勵地址不能為空。","Pool Reward Amount":"礦池勵數量","Pool does not provide relative_lock_height.":"農會未提供 相對_鎖定_高度。","Pool does not provide target_puzzle_hash.":"農會未提供 目標_拼圖_雜湊值。","Pool public key:":"礦池公鑰 (PPK):","Pool:":"農會：","Pooling":"正在為農會耕種中","Pooling Wallet":"Pooling Wallet","Port":"通訊埠","Preparing Plot NFT":"正在準備農場","Preparing standard wallet":"正在準備標準錢包","Previous":"上一個","Previous Block Hash":"Previous Block Hash","Previous Header Hash":"前一個表頭雜湊值","Previous Sub Block Hash":"Previous Sub Block Hash","Private key with public fingerprint {fingerprint}":["帶有公有指紋 ",["fingerprint"]," 的私鑰"],"Private key {0}":["私鑰 ",["0"]],"Private key {fingerprint}":["Private key ",["fingerprint"]],"Private key:":"私鑰：","Proof of Space Size":"空間證明的大小","Proofs Found":"發現證明數","Protocol Version":"協議版本","Pubkey":"公鑰","Public key:":"公鑰：","Puzzle Hash":"Puzzle Hash","Puzzlehash":"拼圖哈希值","Queue Name":"隊列名稱","Queue name":"隊列名稱","Queued":"已隊列","Quit":"Quit","RAM max usage":"最大記憶體使用量","RL Wallet":"RL Wallet","RLListItems.title":"RLListItems.title","Rate Limited":"已限速","Rate Limited Info":"速率限制資訊","Rate Limited Options":"速率限制選項","Rate Limited User Wallet Setup":"速率限制使用者錢包設定","Receive Address":"接收位址","Recipient address is not a coloured wallet address. Please enter a coloured wallet address":"Recipient address is not a coloured wallet address. Please enter a coloured wallet address","Recover":"恢復","Recover DID Wallet":"恢復 DID 錢包","Recover DID wallet":"Recover DID wallet","Recover Distributed Identity Wallet":"還原分散式識別的錢包","Recover Wallet":"恢復錢包","Recoverable Wallet":"Recoverable Wallet","Recovery Wallet":"Recovery Wallet","Refresh Plots":"刷新耕地","Relative Lock Height":"相對鎖定高度","Release Notes":"發行說明","Remove keys from old keyring upon successful migration":"Remove keys from old keyring upon successful migration","Removing":"移除中","Rename":"更改名稱","Rename Wallet":"Rename Wallet","Report an Issue...":"回報問題...","Requested":"Requested","Reset":"Reset","Restore Metadata for Coloured Coins and other Smart Wallets from Backup":"從備份還原有色代幣和其他智慧錢包的元數據","Restore Smart Wallets From Backup":"Restore Smart Wallets From Backup","Retire":"Retire","SB Height":"SB Height","Safe To Skip":"可安全跳過","Save":"儲存","Save Offer":"Save Offer","Save Offer File":"Save Offer File","Save passphrase":"Save passphrase","Search":"Search","Search block by header hash":"以表頭雜湊值搜尋區塊","Second temporary folder location":"第二暫存資料夾位置","See private key":"看私鑰","Seed:":"種子：","Select 2nd Temporary Directory":"選擇第二暫存資料夾","Select Final Directory":"選擇最終資料夾","Select Key":"選擇金鑰","Select Offer":"選擇提案","Select Offer File":"Select Offer File","Select Temporary Directory":"選擇暫存資料夾","Select Wallet":"Select Wallet","Select Wallet Type":"選擇錢包類型","Select the final destination for the folder where you would like the plot to be stored. We recommend you use a large slow hard drive (like external HDD).":"選擇你想要儲存耕地的最終資料夾。我們建議你使用較大的慢速硬碟（像是外接硬碟）。","Select the temporary destination for the folder where you would like the plot to be stored. We recommend you use a fast SSD.":"Select the temporary destination for the folder where you would like the plot to be stored. We recommend you use a fast SSD.","Select the temporary destination for the folder where you would like the plot to be stored. We recommend you use a fast drive.":"選擇你想要儲存耕地的暫存資料夾。我們建議你使用快速硬碟。","Select your Plot NFT":"選擇您的農場","Select your Plot NFT from the dropdown or create a new one.":"在下拉式選單中選擇你的農場或是創建一個新的。","Selected":"已選擇","Selected recovery file:":"選擇的還原檔案：","Self Pooling":"正在為自己的農會耕種中","Self pool. When you win a block you will earn XCH rewards.":"自己的農會。當你贏取一個區塊時你將獲得 XCH 獎勵。","Sell":"賣出","Send":"發送","Send this info packet to your Rate Limited Wallet user who must use it to complete setup of their wallet:":"發送這個資訊封包給你必須需要完成設定他們錢包的速率限制錢包使用者：","Send your pubkey to your Rate Limited Wallet admin:":"發送你的公鑰到你的速率限制錢包管理員：","Share":"Share","Share Offer":"Share Offer","Share Privately":"Share Privately","Share on Hashgreen DEX":"Share on Hashgreen DEX","Share on Keybase":"Share on Keybase","Share on OfferBin":"Share on OfferBin","Show Advanced Options":"顯示進階選項","Show Asset Id":"Show Asset Id","Show Details":"Show Details","Show Token and Asset Issuance Limitations":"Show Token and Asset Issuance Limitations","Show app":"Show app","Side":"方向","Sign In":"登入","Skip":"Skip","Skip Keyring Migration":"Skip Keyring Migration","Skips adding [final dir] to harvester for farming":"Skips adding [final dir] to harvester for farming","Skips adding a final directory to harvester for farming":"跳過新增最終資料夾到收割機進行耕種","Smooth out your XCH farming rewards by joining a pool.":"加入農會來獲得穩定的XCH獎勵。","Something went wrong":"某個地方出錯了","Speech":"語音","Spendable Amount":"可花用數量","Spendable Amount Per Interval":"每間隔可花用數量","Spendable Balance":"可花用餘額","Spending Interval (number of blocks): {interval}":["花用間隔 (區塊數量)：",["interval"]],"Spending Interval Length (number of blocks)":"花費間隔長度 (區塊數量)","Spending Limit (chia per interval): {0}":["花用限制（每間隔可用 Chia）：",["0"]],"Standard Wallet":"Standard Wallet","Starting Wallet":"Starting Wallet","Starting service {0}":["Starting service ",["0"]],"State":"狀態","Status":"狀態","Status:":"狀態：","StatusCard.connections":"StatusCard.connections","StatusCard.height":"StatusCard.height","StatusCard.status":"StatusCard.status","StatusCard.title":"StatusCard.title","Stripe Size":"Stripe Size","Sub Block Height":"Sub Block Height","Sub Block at height {0} in the Chia blockchain":["Sub Block at height ",["0"]," in the Chia blockchain"],"Submit":"確認","Success":"Success","Summary":"Summary","Synced":"已同步","Syncing":"同步中","Syncing <0/>/<1/>":"同步中 <0/> / <1/>","Syncing {progress}/{tip}":["Syncing ",["progress"],"/",["tip"]],"TXCH Block Rewards":"TXCH Block Rewards","TXCH Farming Rewards":"TXCH Farming Rewards","TXCH Fees Collected":"TXCH Fees Collected","TXCH User Fees":"TXCH User Fees","TXCH User Transaction Fees":"TXCH User Transaction Fees","Target Puzzle Hash":"目標拼圖雜湊值","Temporary files are created during the plotting process which exceeds the size of the final plot files. Make sure you have enough space. <0>Learn more</0>":"Temporary files are created during the plotting process which exceeds the size of the final plot files. Make sure you have enough space. <0>Learn more</0>","Temporary folder location":"暫存資料夾位置","The application will stop working at block height 193536.":"這個應用程式會在區塊高度 193536 時停止運作。","The full node that your farmer is connected to is below. <0>Learn more</0>":"你的塊農連接上的完整節點列在下方。<0>了解更多</0>","The minimum required size for mainnet is k=32":"主網最小要求的大小是 k=32","The node is not synced":"此節點尚未同步","The node is syncing, which means it is downloading blocks from other nodes, to reach the latest block in the chain":"節點同步中，這代表正從其他節點中下載區塊，直到到達區塊鏈的最新區塊","The number of Backup IDs needed for recovery cannot exceed the number of Backup IDs added.":"The number of Backup IDs needed for recovery cannot exceed the number of Backup IDs added.","The pool URL \"{normalizedUrl}\" is not working. Is it pool? Error: {0}":["農會網址：「 ",["normalizedUrl"]," 」錯誤。這是農會嗎？錯誤碼： ",["0"]],"The pool URL is not valid. {normalizedUrl}":["農會網址無效. ",["normalizedUrl"]],"The pool URL needs to use protocol https. {normalizedUrl}":["農會網址需要使用https協議. ",["normalizedUrl"]],"The pool URL speciefied is not valid. {normalizedUrl}":["The specified pool URL is not valid. ",["normalizedUrl"]],"The pool URL speciefied is not valid. {poolUrl}":["The pool URL specified is not valid. ",["poolUrl"]],"The pool switching operation was cancelled, please try again by changing pool, or self pooling":"切換農會操作已被取消，請重試切換農會或自己的農會。","The seed used to create the plot. This depends on the pool pk and plot pk.":"用來建立耕地的種子。這取決於池主鍵和耕地主鍵。","The total number of VDF (verifiable delay function) or proof of time iterations on the whole chain up to this sub block.":"總 VDF（可驗證的延遲函數）數量或整個鏈上直到此子區塊的時間疊代證明。","The total number of VDF (verifiable delay function) or proof of time iterations on this block.":"VDF（可驗證的延遲函數）數量或此區塊上的時間疊代證明。","The total transactions fees in this block. Rewarded to the farmer.":"在這個區塊的總交易費用。是獎勵給塊農的。","These are the instructions for how the farmer wants to get paid. By default this will be an XCH address, but it can be set to any string with a size of less than 1024 characters, so it can represent another blockchain or payment system identifier.":"這些是關於農夫如何獲得報酬的說明. 正常的狀況下, 這將是一個 XCH 地址, 但它可以設置為小於 1024 個字符的任何字符串，因此它可以代表另一個區塊鏈或支付系統識別碼.","These plots are invalid, you might want to delete them.":"這些耕地是無效的，你可能想要刪除他們。","This access token is verified":"This access token is verified","This access token is whitelisted":"This access token is whitelisted","This allows you to add a directory that has plots in it. If you have not created any plots, go to the plotting screen.":"這允許你增加已存有耕地的資料夾。如果你還沒建立任何的耕地，去耕地規劃畫面吧。","This difficulty is an artifically lower difficulty than on the real network, and is used when farming, in order to find more proofs and send them to the pool. The more plots you have, the higher difficulty you will have. However, the difficulty does not affect rewards.":"為了在耕種時找到更多證明以傳送至網路，此處難度與實際網路相比已人工降低過。你擁有的耕地越多，你的難度就越高。 但是，難度不影響獎勵。","This feature is available only from the GUI.":"這個功能只能在 GUI 中提供。","This is the amount of Chia that you can currently use to make transactions. It does not include pending farming rewards, pending incoming transactions, and Chia that you have just spent but is not yet in the blockchain.":"這是你目前可用來交易的 Chia 數量。他不包括未結的耕種獎勵，未結的匯入交易和你剛花用，但還沒在區塊鏈內的 Chia 。","This is the pending change, which are change coins which you have sent to yourself, but have not been confirmed yet.":"這是待處理的改變，是你發送給自己的但還沒確認的改變。","This is the sum of the incoming and outgoing pending transactions (not yet included into the blockchain). This does not include farming rewards.":"這是匯入和匯出的未結交易的總和（還沒包含在區塊鏈中）。這不包括耕種獎勵。","This is the time of the latest peak sub block.":"這是最後的峰值子區塊的時間。","This is the time the block was created by the farmer, which is before it is finalized with a proof of time":"這是塊農建立區塊的時間，這是在用時間證明最終確定之前的時間","This is the total amount of chia in the blockchain at the current peak sub block that is controlled by your private keys. It includes frozen farming rewards, but not pending incoming and outgoing transactions.":"這是由你的私鑰控制的目前峰值子區塊中區塊鏈中 Chia 的總量。他包括凍結耕種獎勵，但不包括未結的匯入和匯出交易。","This is the total balance + pending balance: it is what your balance will be after all pending transactions are confirmed.":"這是總餘額 + 未結餘額：這是在全部未結交易結清後你的餘額。","This is the total number of points this plotNFT has with this pool, since the last payout. The pool will reset the points after making a payout.":"自上次支付以來該農場在該池中的總積分。農會在支付過後將重置積分。","This is the total number of points your farmer has found for this plot NFT. Each k32 plot will get around 10 points per day, so if you have 10TiB, should should expect around 1000 points per day, or 41 points per hour.":"這是您的農夫為此農場找到的總積分。 每個 K32 圖每天將獲得大約 10 點積分，所以如果您有 10TiB，則每天大約 1000 點積分，或每小時 41 點積分。","This node is fully caught up and validating the network":"此節點已完全跟上和驗證網路","This number reflects the number of nodes that the accepted SpendBundle has been sent to":"This number reflects the number of nodes that the accepted SpendBundle has been sent to","This offer has completed successfully":"This offer has completed successfully","This offer is no longer valid":"This offer is no longer valid","This plot NFT is assigned to a different key. You can still create plots for this plot NFT, but you can not make changes.":"此農場分配給其他的金鑰。你仍然可以為此農場建立耕地，但不能進行更改。","This plot NFT is not connected to pool":"這個農場目前沒有連線到農會.","This table shows you the last time your farm attempted to win a block challenge.":"This table shows you the last time your farm attempted to win a block challenge.","This table shows you the last time your farm attempted to win a block challenge. <0>Learn more</0>":"這個表格顯示上次你的農場試著獲勝的區塊挑戰。<0>了解更多</0>","This trade was created at this time":"此貿易是在這個時候建立的","This trade was included on blockchain at this block height":"此貿易已包含在區塊鏈中的這個區塊高度","This version of Chia is no longer compatible with the blockchain and can not safely farm.":"此版本的 Chia 不再與區塊鏈相容，因此無法安全地進行耕種。","Time Created":"建立時間","Time this trade was created at this time":"Time this trade was created at this time","Timestamp":"時間戳記","To":"到","Token and Asset Issuance Limitations":"Token and Asset Issuance Limitations","Token has empty name":"Token has empty name","Token has empty tail":"Token has empty tail","Total Balance":"總餘額","Total Chia Farmed":"Total Chia Farmed","Total Iterations":"總疊代次數","Total Network Space":"總網路空間","Total Plot Size:":"總耕地大小：","Total Size of Plots":"總耕地大小","Total VDF Iterations":"總 VDF 疊代次數","Total iterations since the start of the blockchain":"從區塊鏈開始後的總疊代次數","Trade":"Trade","Trade Details":"交易明細","Trade ID":"貿易 ID","Trade ID:":"貿易 ID:","Trade Overview":"貿易概觀","Trades will show up here":"貿易將會在此顯示","Trading":"貿易中","Trading History":"貿易歷史","Transaction has not been sent to node yet":"Transaction has not been sent to node yet","Transaction has sent to a full node and is pending inclusion into the mempool. {0}":["Transaction has sent to a full node and is pending inclusion into the mempool. ",["0"]],"Transaction has successfully been sent to a full node and included in the mempool.":"Transaction has successfully been sent to a full node and included in the mempool.","Transactions":"Transactions","Transactions Filter Hash":"交易過濾雜湊值","Try Again":"Try Again","Type":"類型","Unable to create plot NFT":"無法建立農場","Unable to load the list of the keys":"Unable to load the list of the keys","Unable to open Keybase. Install Keybase from https://keybase.io":"Unable to open Keybase. Install Keybase from https://keybase.io","Unable to open browser. Install Keybase from https://keybase.io":"Unable to open browser. Install Keybase from https://keybase.io","Unclaimed Rewards":"尚未領取的獎勵","Unfinished":"未完成","Unique identifier":"唯一識別碼","Unknown":"未知","Unknown CAT":"Unknown CAT","Unlock Keyring":"Unlock Keyring","Unsaved Changes":"尚未儲存變更","User Pubkey":"使用者公鑰","VDF Sub Slot Iterations":"VDF 子欄位疊代次數","Value is too high":"Value is too high","Value seems high":"數值有點高","Verify Pool Details":"驗證農會細節","View":"檢視","View Log":"檢視日誌","View Offer":"檢視提案","View Payout Instructions":"View Payout Instructions","View Pool Login Link":"顯示農會登入連結","View an Offer":"View an Offer","View on Hashgreen DEX":"View on Hashgreen DEX","View on OfferBin":"View on OfferBin","View pending balances":"檢視未結餘額","View pending balances...":"檢視未結餘額...","Viewing offer":"Viewing offer","Viewing offer created at {0}":["Viewing offer created at ",["0"]],"Viewing offer: {offerFilePath}":["Viewing offer: ",["offerFilePath"]],"Wait for synchronization":"等待同步中","Waiting for launch":"Waiting for launch","Waiting for network to launch":"Waiting for network to launch","Waiting for the transaction to be confirmed":"正在等待交易結算","Wallet Status:":"錢包狀態：","Wallet does not exists":"錢包不存在","Wallet id is not defined":"Wallet id is not defined","WalletItem.ccWallet":"WalletItem.ccWallet","WalletItem.chiaWallet":"WalletItem.chiaWallet","WalletItem.didWallet":"WalletItem.didWallet","WalletItem.rlWallet":"WalletItem.rlWallet","WalletStatusCard.not_synced":"WalletStatusCard.not_synced","WalletStatusCard.synced":"WalletStatusCard.synced","WalletStatusCard.syncing":"WalletStatusCard.syncing","Wallets":"錢包","Wallets.title":"Wallets.title","Want to Join a Pool? Create a Plot NFT":"想加入農會？建立一個農場","Want to earn more Chia? Add more plots to your farm.":"想要獲得更多 Chia？增加更多耕地到你的農場。","Want to have a delay before the next plot starts?":"想要在下次規劃耕地開始前有一段延遲？","Warning":"Warning","Warning: This key is used for a wallet that may have a non-zero balance. By deleting this key you may lose access to this wallet":"警告：這個金鑰的錢包還有餘額，如果刪除金鑰可能會導致你遺失這個錢包","Warning: This key is used for your farming rewards address. By deleting this key you may lose access to any future farming rewards":"警告：這個金鑰用於你的耕種獎勵地址。刪除金鑰後可能會無法領取之後的耕種獎勵。","Warning: This key is used for your pool rewards address. By deleting this key you may lose access to any future pool rewards":"警告：這個金鑰用於你的農會獎勵地址。刪除金鑰後可能會無法領取之後的農會獎勵。","Warning: Verify that the offered CAT asset IDs match the asset IDs of the tokens you expect to receive.":"Warning: Verify that the offered CAT asset IDs match the asset IDs of the tokens you expect to receive.","Weight":"權重","Weight is the total added difficulty of all sub blocks up to and including this one":"權重是所有包括這個的子區塊的總增加難度","Welcome to Chia. Please log in with an existing key, or create a new key.":"歡迎來到 Chia。請用已有的金鑰登入，或建立新的金鑰。","Welcome! The following words are used for your wallet backup. Without them, you will lose access to your wallet, keep them safe! Write down each word along with the order number next to them. (Order is important)":"歡迎！以下的字是用來為你的錢包備份。沒有這些你將會無法進入你的錢包。將這些安全保存好！寫下每個字，包括他們旁邊的順序數字。（順序很重要）","When you receive the setup info packet from your admin, enter it below to complete your Rate Limited Wallet setup:":"當你從你的管理員收到設定資訊封包時，在下面輸入以完成你的速率限制錢包設定：","Window":"視窗","Without fees":"未計雜費","Yes":"是","You are in pending state. Please wait for confirmation":"您處於待確認的狀態。請等待確認","You are not self pooling":"你目前不在自己的農會。","You can still create plots for this plot NFT, but you can not make changes until sync is complete.":"您仍然可以為此農場創建耕地，但在同步完成之前您不能進行更改。","You created this offer":"You created this offer","You do not need to be synced or connected to plot. Temporary files are created during the plotting process which exceed the size of the final plot files. Make sure you have enough space. <0>Learn more</0>":"You do not need to be synced or connected to plot. Temporary files are created during the plotting process which exceed the size of the final plot files. Make sure you have enough space. <0>Learn more</0>","You do not need to be synched or connected to Plot. Temporary files are created during the plotting process which exceed the size of the final plot files. Make sure you have enough space. <0>Learn more</0>":"You do not need to be synched or connected to Plot. Temporary files are created during the plotting process which exceed the size of the final plot files. Make sure you have enough space. <0>Learn more</0>","You do not need to be synched or connected to plot. Temporary files are created during the plotting process which exceed the size of the final plot files. Make sure you have enough space. <0>Learn more</0>":"你不需要同步或連線後才能規劃耕地。規劃耕地時所產生的暫存檔將大於最終的耕地大小。確保你有足夠的空間。<0>了解更多</0>","You have made changes. Do you want to discard them?":"您有做過變更。請問您想要放棄它們嗎？","You have {0}% of the space on the network, so farming a block will take {expectedTimeToWin} in expectation.":["You have ",["0"],"% of the space on the network, so farming a block will take ",["expectedTimeToWin"]," in expectation."],"You have {0}% of the space on the network, so farming a block will take {expectedTimeToWin} in expectation. Actual results may take 3 to 4 times longer than this estimate.":["你擁有 ",["0"]," ％的網路空間，耕種一個區塊預計要 ",["expectedTimeToWin"]," 。實際結果可能是預計的 3 至 4 倍。"],"You have {0}% of the space on the network, so farming a block will take {expectedTimeToWin} in expectation. Actual results may take 3-4 times longer - actual results may be 3x this estimate.<0>Learn about Expected Value</0>":["You have ",["0"],"% of the space on the network, so farming a block will take ",["expectedTimeToWin"]," in expectation. Actual results may take 3-4 times longer - actual results may be 3x this estimate.<0>Learn about Expected Value</0>"],"You have {0}% of the space on the network, so farming a block will take {expectedTimeToWin} in expectation. Actual results may take 3-4 times longer this estimate. <0>Learn about Expected Value</0>":["You have ",["0"],"% of the space on the network, so farming a block will take ",["expectedTimeToWin"]," in expectation. Actual results may take 3-4 times longer this estimate. <0>Learn about Expected Value</0>"],"You need to claim your rewards first":"您需要先領取你的獎勵","You need to upgrade the Chia application as this version will stop working soon!":"You need to upgrade the Chia application as this version will stop working soon!","You need {currencyCode} to join a pool.":["你需要 ",["currencyCode"]," 來加入農會。"],"You will offer":"You will offer","You will receive":"You will receive","You will recieve <0/> to {0}":["你將會收到 <0/> 到 ",["0"]],"You will recieve <0/> to {address}":["你將會收到 <0/> 到 ",["address"]],"Your DID requires at least {dids_num_req} attestation file{0} for recovery. Please upload additional files.":["Your DID requires at least ",["dids_num_req"]," attestation file",["0"]," for recovery. Please upload additional files."],"Your Farm Overview":"你的農場概觀","Your Full Node Connection":"你的完整節點連線","Your Harvester Network":"你的收割機網路","Your Offers":"Your Offers","Your Pool Overview":"您的農場概觀","Your keyring is locked":"Your keyring is locked","Your keys have not been migrated to a new keyring. You will be unable to create new keys or delete existing keys until migration completes. Are you sure you want to skip migrating your keys?":"Your keys have not been migrated to a new keyring. You will be unable to create new keys or delete existing keys until migration completes. Are you sure you want to skip migrating your keys?","Your keys need to be migrated to a new keyring that is optionally secured by a master passphrase.":"Your keys need to be migrated to a new keyring that is optionally secured by a master passphrase.","Your keys need to be migrated to a new keyring that is secured by a master passphrase.":"Your keys need to be migrated to a new keyring that is secured by a master passphrase.","Your offer has been successfully posted to Keybase.":"Your offer has been successfully posted to Keybase.","Your offer:":"Your offer:","Your passphrase can be stored in your system's secure credential store. Chia will be able to access your keys without prompting for your passphrase.":"Your passphrase can be stored in your system's secure credential store. Chia will be able to access your keys without prompting for your passphrase.","[Error 13] Permission denied. You are trying to access a file/directory without having the necessary permissions. Most likely one of the plot folders in your config.yaml has an issue.":"［錯誤 13］權限被拒。你想存取一個檔案／目錄但沒有所需的權限。很有可能你 config.yaml 裡其中一個耕地資料夾有問題。","[Error 22] File not found. Most likely one of the plot folders in your config.yaml has an issue.":"［錯誤 22］找不到檔案。很有可能你 config.yaml 裡其中一個耕地資料夾有問題。","connections:":"連線：","height:":"高度：","not synced":"未同步","p2_singleton_puzzle_hash is not defined":"\"p2_singleton_puzzle_hash\"參數未定義","status:":"狀態：","synced":"已同步","syncing":"同步中","{0}":[["0"]],"{0} Block Rewards":[["0"]," Block Rewards"],"{0} User Transaction Fees":[["0"]," User Transaction Fees"],"{0} points {1} - {2} hours ago":[["0"]," 積分 ",["1"]," - ",["2"]," 小時前"],"{currencyCode} Block Rewards":[["currencyCode"]," 區塊獎勵"],"{currencyCode} Total Chia Farmed":[["currencyCode"]," 已收耕 Chia"],"{currencyCode} User Transaction Fees":[["currencyCode"]," 使用者交易費用"],"{mojo, plural, one {mojo} other {mojos}}":[["mojo","plural",{other:"mojos"}]],"{mojos, plural, one {mojo} other {mojos}}":[["mojos","plural",{one:"mojo",other:"mojos"}]]}};
