@@ -1,6 +1,6 @@
 import { truncate } from 'lodash';
 
-export default function getCatUnit(name?: string, length: number = 10): string {
+export default function getCatUnit(name?: string, length = 10): string {
   if (!name) {
     return '';
   }
